Grab a gun and shoot me
