+++ src/components/Icon.tsx
@@ -7,6 +7,7 @@ export enum Icons {
   BOOKMARK_OUTLINE = "bookmark_outline",
   CLOCK = "clock",
   EYE = "eye",
+  COMPASS = "compass",
   EYE_SLASH = "eyeSlash",
   ARROW_LEFT = "arrowLeft",
   ARROW_RIGHT = "arrowRight",
@@ -79,6 +80,7 @@ export interface IconProps {
 }
 
 const iconList: Record<Icons, string> = {
+  compass: `<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 512 512"><!--!Font Awesome Free 6.5.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path fill="currentColor" d="M464 256A208 208 0 1 0 48 256a208 208 0 1 0 416 0zM0 256a256 256 0 1 1 512 0A256 256 0 1 1 0 256zm306.7 69.1L162.4 380.6c-19.4 7.5-38.5-11.6-31-31l55.5-144.3c3.3-8.5 9.9-15.1 18.4-18.4l144.3-55.5c19.4-7.5 38.5 11.6 31 31L325.1 306.7c-3.2 8.5-9.9 15.1-18.4 18.4zM288 256a32 32 0 1 0 -64 0 32 32 0 1 0 64 0z"/></svg>`,
   Hammer: `<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 512 512"><!--!Font Awesome Free 6.5.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path fill="currentColor" d="M318.6 9.4c-12.5-12.5-32.8-12.5-45.3 0l-120 120c-12.5 12.5-12.5 32.8 0 45.3l16 16c12.5 12.5 32.8 12.5 45.3 0l4-4L325.4 293.4l-4 4c-12.5 12.5-12.5 32.8 0 45.3l16 16c12.5 12.5 32.8 12.5 45.3 0l120-120c12.5-12.5 12.5-32.8 0-45.3l-16-16c-12.5-12.5-32.8-12.5-45.3 0l-4 4L330.6 74.6l4-4c12.5-12.5 12.5-32.8 0-45.3l-16-16zm-152 288c-12.5-12.5-32.8-12.5-45.3 0l-112 112c-12.5 12.5-12.5 32.8 0 45.3l48 48c12.5 12.5 32.8 12.5 45.3 0l112-112c12.5-12.5 12.5-32.8 0-45.3l-1.4-1.4L272 285.3 226.7 240 168 298.7l-1.4-1.4z"></path></svg>`,
   cloud: `<svg xmlns="http://www.w3.org/2000/svg"  width="1em" height="1em" viewBox="0 0 640 512"><!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path fill="currentColor" d="M0 336c0 79.5 64.5 144 144 144l368 0c70.7 0 128-57.3 128-128c0-61.9-44-113.6-102.4-125.4c4.1-10.7 6.4-22.4 6.4-34.6c0-53-43-96-96-96c-19.7 0-38.1 6-53.3 16.2C367 64.2 315.3 32 256 32C167.6 32 96 103.6 96 192c0 2.7 .1 5.4 .2 8.1C40.2 219.8 0 273.2 0 336z"/></svg>`,
   kermit: `<svg xmlns="http://www.w3.org/2000/svg" width="1.65em" height="1.65em" viewBox="0 0 512 512" style="shape-rendering:geometricPrecision;text-rendering:geometricPrecision;image-rendering:optimizeQuality;fill-rule:evenodd;clip-rule:evenodd"><path style="opacity:.007" fill="#dbe0c1" d="M298.5 6.5q2.947.672 1 3-1.329-1.238-1-3"/><path style="opacity:.004" fill="#bfc77f" d="M246.5 9.5q2 1 0 0"/><path style="opacity:1" fill="#e6e6dc" d="M213.5 6.5a4.93 4.93 0 0 1-.5 3q-16.531-1.207-32.5 5 .186.91 1 1a30.6 30.6 0 0 0-8 7q-1.222-2.703-4.5-3-4.8 1.06-9.5 2.5-3 1-4 4a4.93 4.93 0 0 1-3 .5q-4.31 11.055-4 23a5.73 5.73 0 0 1-3-1q2.051-43.799 45.5-48 12.198-.39 22.5 6"/><path style="opacity:.004" fill="#b1bc4a" d="M265.5 14.5q.595-1.538 2-.5a2.43 2.43 0 0 1-2 .5"/><path style="opacity:1" fill="#eff1e7" d="M213.5 6.5q10.868 5.65 17 16a19.5 19.5 0 0 0 1 7.5q-.504 1.545-2 1.5a21.1 21.1 0 0 0-1-9 3.65 3.65 0 0 0-1.5 1q-6.1-7.844-15.5-5v4q-7.23-2.368-15-2.5-6.562-5.102-15-4.5-.814-.09-1-1 15.969-6.207 32.5-5a4.93 4.93 0 0 0 .5-3"/><path style="opacity:1" fill="#0c0c0a" d="M181.5 15.5q8.438-.602 15 4.5 7.77.132 15 2.5 2.501 3.735-1 6.5l-14 1a20.65 20.65 0 0 1-5 7 67.4 67.4 0 0 1-15 2 63 63 0 0 0-7 8q-6.504 1.25-6-5.5 3.494-5.939 8.5-11a168 168 0 0 0 1.5-8 30.6 30.6 0 0 1 8-7"/><path style="opacity:.018" fill="#e3e5da" d="M300.5 14.5q1.226 3.495 2 7.5l-1 1.5q-2.311-4.479-1-9"/><path style="opacity:.004" fill="#afbd43" d="M266.5 18.5q1.538.595.5 2a2.43 2.43 0 0 1-.5-2"/><path style="opacity:1" fill="#f1f1e8" d="M173.5 22.5a168 168 0 0 1-1.5 8q-5.006 5.061-8.5 11a2.43 2.43 0 0 0-.5-2q-2.81-.884-5 2-.76-2.244 1.5-3.5l-5-1q4.245-3.966-1-2.5-3.406 8.787-1 18 .128 2.242-2 3a20.8 20.8 0 0 1-2-6q-.31-11.945 4-23a4.93 4.93 0 0 0 3-.5q1-3 4-4 4.7-1.44 9.5-2.5 3.278.297 4.5 3"/><path style="opacity:.004" fill="#abba3c" d="M266.5 23.5q2 1 0 0"/><path style="opacity:.978" fill="#b6c153" d="M255.5 23.5q-.538 2.807-3 4.5-6.055 2.747-11.5-.5a3.94 3.94 0 0 0-2 1l-4-4-1 10q-1.006 1.129-2.5 1 2.158-2.233 2-5.5a42.3 42.3 0 0 0-2-7.5 18.4 18.4 0 0 0 6-.5q3.184-4.69 7-1.5 2.014-2.103 2.5 1l7-1z"/><path style="opacity:.004" fill="#e8e9e2" d="M301.5 25.5q2.377 1.461 3 4.5-3.41-.475-3-4.5"/><path style="opacity:1" fill="#e2e4b7" d="M230.5 22.5h1a42.3 42.3 0 0 1 2 7.5q.158 3.267-2 5.5-1.487-.258-2 1a73.7 73.7 0 0 1-14 4q4.818-5.163 12-6.5 1.321-1.062 2-2.5 1.496.045 2-1.5a19.5 19.5 0 0 1-1-7.5"/><path style="opacity:.986" fill="#a4b23d" d="M285.5 40.5a17.8 17.8 0 0 1-6 4q-.926.165-1.5 1a28.7 28.7 0 0 0 .5 13q3.255 7.065 6.5 14 1.063 1.32 2.5 2a16.3 16.3 0 0 0 2 5 2.43 2.43 0 0 1-2-.5 63 63 0 0 0-5-6.5 8.4 8.4 0 0 0-.5-4 101 101 0 0 0-5.5-11q-.531-5.385-4-9.5 3.954-2.985 8-6-.747-1.737-2.5-2.5a16.3 16.3 0 0 1-4.5 2q-1.374-.064-2-1 2.472-2.934 2-7 2.457-.989 5-2 1.857 3.205 5.5 4a27.6 27.6 0 0 0 1.5 5"/><path style="opacity:.821" fill="#e3e3d7" d="M339.5 111.5q-3.373.87-5 4l-2 1q-7.572 1.99-13 4a25.9 25.9 0 0 0-1-10q-2.433 2.908-5.5 1-2.76-8.145-5.5-16a192 192 0 0 0 5.5 5 40.7 40.7 0 0 0 3.5 7q4.146 1.892 7 5.5 4.36-.127 8-1 14.937-11.324 14-30.5l-.5-8q-.913-2.35-2.5-4-.874-5.17-4-9-4.875-9.102-15-12.5l1-1q-19.406-9.175-34 6 1.423.925 2.5-.5.255 4.995 3 10 1.035 1.857 1.5 0a34.5 34.5 0 0 1 8 6 25.9 25.9 0 0 1 1 10q-1.465-4.18-1-9h-6a12.9 12.9 0 0 0-.5-5q-1.5 2-3 0-.49 7.827 2.5 15 1.27 7.953-2.5 0a34.3 34.3 0 0 0-1.5-8 31 31 0 0 1-2 2 41.1 41.1 0 0 0-6-16q.673-8.86 0-17 37.136-10.044 56.5 23 9.994 24.831-3.5 48"/><path style="opacity:1" fill="#f9f9ee" d="M229.5 31.5q-.679 1.438-2 2.5-7.182 1.337-12 6.5-4.68 2.04-10 2.5 1.255 1.517 1 3.5-5.665 2.22-12 2.5-4.513 3.11-10 4-1.255 1.517-1 3.5-10.23 3.477-20.5 4a23.4 23.4 0 0 1-5.5-2 4.45 4.45 0 0 1 2-1.5 795 795 0 0 1-7-4.5q-2.406-9.213 1-18 5.245-1.466 1 2.5l5 1q-2.26 1.256-1.5 3.5 2.19-2.884 5-2 .696.891.5 2-.504 6.75 6 5.5a63 63 0 0 1 7-8 67.4 67.4 0 0 0 15-2 20.65 20.65 0 0 0 5-7l14-1q3.501-2.765 1-6.5v-4q9.4-2.844 15.5 5 .621-.685 1.5-1a21.1 21.1 0 0 1 1 9"/><path style="opacity:1" fill="#c0ca69" d="M229.5 36.5q-6.981 4.34-15 7-2.062 4.57-7 4a31.5 31.5 0 0 1 8-7 73.7 73.7 0 0 0 14-4"/><path style="opacity:1" fill="#a3b710" d="M260.5 42.5h-4q.21 2.436-2 3.5 3.946 2.536-1 3.5a97 97 0 0 1 9.5 7q.405-4.086 1.5-8a19 19 0 0 0-7-2.5 44 44 0 0 0 11-2.5q.594 7.236 5 13 1.237 1.329 3 1a101 101 0 0 1 5.5 11 8.4 8.4 0 0 1 .5 4 27.6 27.6 0 0 0-4.5-2q-1.545.504-1.5 2-3.346 1.686-7 1.5-3.089-1.779-.5-3.5 2.838 4.335 2.5-1-6.508-.183-1-1v-3h3q-.329 1.762 1 3 3.474-3.433-.5-5 .224-6.555-5.5-4a6.8 6.8 0 0 0 3 1.5 10.1 10.1 0 0 1-6 1.5h-5q.393 2.217-1 4-2.15-2.58-2.5-6-1.062-1.32-2.5-2-3.34-1.35-3-5a12 12 0 0 0-6 1 6.54 6.54 0 0 1-3-3q-2.318-1.428-5-1-.32 2.466 1 4.5a19 19 0 0 0 8 1.5v5h-7q-.255-1.983 1-3.5-3.453-3.001-7.5-1.5-2.1-3.51-1.5-7-1.329 1.238-1 3-1.937-.356-3 1-1.487.258-2-1 .472-2.589.5-5l-1 1q-1.362-2.909-4.5-3-.814-.09-1-1a26 26 0 0 0 5-3q9.213-3.915 20-7 3.029 1.47 5.5 4l3-3q1 2 2 0l3 2q.502 3.005 2.5 5"/><path style="opacity:1" fill="#97ab0a" d="M271.5 40.5q.626.936 2 1a16.3 16.3 0 0 0 4.5-2q1.753.763 2.5 2.5-4.046 3.015-8 6 3.469 4.114 4 9.5-1.763.329-3-1-4.406-5.764-5-13a44 44 0 0 1-11 2.5q3.765.564 7 2.5a47 47 0 0 0-1.5 8 97 97 0 0 0-9.5-7q4.946-.964 1-3.5 2.21-1.064 2-3.5h10q-.329-1.762 1-3a6.85 6.85 0 0 0 4 1"/><path style="opacity:1" fill="#abbb2c" d="M273.5 33.5q.472 4.066-2 7a6.85 6.85 0 0 1-4-1q-1.329 1.238-1 3h-6q-1.998-1.995-2.5-5l-3-2q-1 2-2 0l-3 3q-2.471-2.53-5.5-4-10.787 3.085-20 7a26 26 0 0 1-5 3q.186.91 1 1-2.128 1.427-5 1 .03 2.464-2 4-4.975 3.332-11 3.5-5.25 2.25-7.5 7.5a6.3 6.3 0 0 0-3-1 34 34 0 0 1-6.5 3q-4.044-1.68-8 0-4.81 3.001-8.5 7a30.8 30.8 0 0 0-6.5-3v-2q7.428.525 13-4a47 47 0 0 0 11-1.5 56.8 56.8 0 0 1 11-5 179 179 0 0 0 10-7.5q4.938.57 7-4 8.019-2.66 15-7 .513-1.258 2-1 1.494.129 2.5-1l1-10 4 4a3.94 3.94 0 0 1 2-1q5.445 3.247 11.5.5 2.462-1.694 3-4.5 1.782-.821 3.5-1 4.101 1.293.5 2.5 2.031 1.94 4 .5 3.59.892 5.5 4 1.61-1.17 3.5.5l-1 1q1.321 1.063 2 2.5"/><path style="opacity:1" fill="#f1f0e8" d="M338.5 60.5a16.5 16.5 0 0 1-5-2.5q-8.577-.315-17-1.5-2.166 3.162-2 7-5.804-5.544-13.5-8-6.414 1.148-3.5 7-.465 1.857-1.5 0-2.745-5.005-3-10-1.077 1.425-2.5.5 14.594-15.175 34-6l-1 1q10.125 3.398 15 12.5"/><path style="opacity:1" fill="#edeed0" d="M215.5 40.5a31.5 31.5 0 0 0-8 7q-7.383 4.176-16 5v3q-7.011 3.195-15 4.5-.834.575-1 1.5a88.6 88.6 0 0 1-25 2.5q-3.928-6.077-4.5-13.5-.575-.835-1.5-1 0-1 1-1 1.337.915 3 1a20.8 20.8 0 0 0 2 6q2.128-.758 2-3a795 795 0 0 0 7 4.5 4.45 4.45 0 0 0-2 1.5 23.4 23.4 0 0 0 5.5 2q10.27-.523 20.5-4-.255-1.983 1-3.5 5.487-.89 10-4 6.335-.28 12-2.5.255-1.983-1-3.5 5.32-.46 10-2.5"/><path style="opacity:1" fill="#080807" d="M314.5 63.5q16.763.507 14.5 17-3.195 5.508-2.5 12 .793 5.538-3 9.5-2.5 1-5 0a99 99 0 0 1-1.5-13.5 20.9 20.9 0 0 0-6.5-3.5q-3.35-2.457-4-6.5a25.9 25.9 0 0 0-1-10 34.5 34.5 0 0 0-8-6q-2.914-5.852 3.5-7 7.696 2.456 13.5 8"/><path style="opacity:.493" fill="#aeb943" d="M144.5 49.5q.925.165 1.5 1 .572 7.423 4.5 13.5a88.6 88.6 0 0 0 25-2.5q-5.572 4.525-13 4h-16a55 55 0 0 0-1-12q-10.845 2.362-20 8.5a53 53 0 0 1-7 0q-5.781 2.411-3 8a16.14 16.14 0 0 1-7 5.5l-2 4a25.9 25.9 0 0 1-10 1 59.5 59.5 0 0 0 5.5-7q3.194 2.045 1.5-1a99.6 99.6 0 0 0 13-14 7.25 7.25 0 0 0 2.5-2l1 1a4.93 4.93 0 0 0 .5-3q4.381.531 7-3 3.156 2.645 6.5 2 4.633-3.49 10.5-4"/><path style="opacity:1" fill="#b9c557" d="M207.5 47.5a179 179 0 0 1-10 7.5 56.8 56.8 0 0 0-11 5 47 47 0 0 1-11 1.5q.166-.925 1-1.5 7.989-1.305 15-4.5v-3q8.617-.824 16-5"/><path style="opacity:1" fill="#fbfcf5" d="M338.5 60.5q3.126 3.83 4 9a259 259 0 0 0 1 17q-2.327 4.288-3 9a25 25 0 0 0-7 .5q-2.503 2.503-6 2.5.442-3.352-1-6-.695-6.492 2.5-12 2.263-16.493-14.5-17-.166-3.838 2-7 8.423 1.185 17 1.5a16.5 16.5 0 0 0 5 2.5"/><path style="opacity:1" fill="#92a901" d="M254.5 58.5q-1.37 4.842-1 10h-5q-.332 3.246-3 5-3.161-1.453-7-1 1.115 5.175-4 5a5.73 5.73 0 0 0 1-3q2.322-.481 2-3h4q.399-5.37 4-1 1.64-2.052-1-3.5a18.2 18.2 0 0 0-8 .5v-5q-1.763-.329-3 1 3.548 5.409-.5 10-.892-2.098.5-4-1.237-1.329-3-1v-10q-1.763.329-3-1 1.96-2.047 0-4 1.063-1.356 3-1-.329-1.762 1-3-.6 3.49 1.5 7 4.047-1.501 7.5 1.5-1.255 1.517-1 3.5h7v-5q-4.156.12-8-1.5-1.32-2.034-1-4.5 2.682-.428 5 1a6.54 6.54 0 0 0 3 3 12 12 0 0 1 6-1q-.34 3.65 3 5"/><path style="opacity:1" fill="#98af04" d="M225.5 52.5q.513 1.258 2 1 1.96 1.953 0 4 1.237 1.329 3 1v10q1.763-.329 3 1-1.392 1.902-.5 4 4.048-4.591.5-10 1.237-1.329 3-1v5a18.2 18.2 0 0 1 8-.5q2.64 1.448 1 3.5-3.601-4.37-4 1h-4q.322 2.519-2 3-3.844 1.05-7.5 3-2.607-3.755-6.5-8.5-2.5-1-5 0 5.626 1.142.5 3.5a70 70 0 0 1-3.5-5q-1.98-2.163-5-2 .003 3.495-2.5 6-.616-4.523 2.5-8-2.881-1.712-6-.5a4.45 4.45 0 0 1 2 1.5q-1.914 1.718-4 0-1.393 1.783-1 4h-3q.455 4.631-4 5v3q-4.167.717-6-3 3.897-1.77 5-6 3.07 2.105 5-1 .09.814 1 1a123 123 0 0 0 2-8 27.3 27.3 0 0 1 9-2q-.255 1.983 1 3.5 3.396.354 6.5 1.5-2.79-4.452-4.5-9 1.726-1.847 3.5 0 1.782-1.595 1.5-4a6.85 6.85 0 0 0 4-1q5.493.507 2 3.5 1.422.725 3 .5"/><path style="opacity:1" fill="#d6d6bf" d="M285.5 40.5h1q.673 8.14 0 17a41.1 41.1 0 0 1 6 16q-.309 3.074-2.5 1a30 30 0 0 1-2-9 31 31 0 0 1-4.5-3q-4.885-8.436-4-18a17.8 17.8 0 0 0 6-4"/><path style="opacity:1" fill="#bec57f" d="M279.5 44.5q-.885 9.564 4 18 3.576 5.333 4 12-1.437-.68-2.5-2-3.245-6.935-6.5-14a28.7 28.7 0 0 1-.5-13q.574-.835 1.5-1"/><path style="opacity:1" fill="#8ba400" d="M213.5 67.5a70 70 0 0 0 3.5 5q5.126-2.358-.5-3.5 2.5-1 5 0 3.893 4.745 6.5 8.5 3.656-1.95 7.5-3a5.73 5.73 0 0 1-1 3q-.407 1.578-.5 4a20.7 20.7 0 0 1-3.5 4q.119-3.249 2-6l-1-2a6.85 6.85 0 0 1-4 1q-.2 5.094-3 1-2.028 1.525-3 4-2.972-.907-1-3a495 495 0 0 0-4-3q-3.149-1.667-7-2 .18 4.057-3 2-1.695 2.595-6 4.5l4 1q-7.865 1.93 0 4.5a15.7 15.7 0 0 1 2 4 18.9 18.9 0 0 0-3 5q-4.23-1.725-1-3.5-2.906-2.193-6-3.5a2.43 2.43 0 0 0 2-.5 104 104 0 0 0-5-13 40.8 40.8 0 0 0 8-3.5 9.3 9.3 0 0 0 2.5 2q4.179-2.338 8.5-1 .47-3.03 1-6"/><path style="opacity:1" fill="#f2f2e4" d="M306.5 78.5q.65 4.043 4 6.5a20.9 20.9 0 0 1 6.5 3.5 99 99 0 0 0 1.5 13.5q2.5 1 5 0 3.793-3.962 3-9.5 1.442 2.648 1 6 3.497.003 6-2.5a25 25 0 0 1 7-.5q.673-4.712 3-9a259 259 0 0 1-1-17q1.587 1.65 2.5 4l.5 8q.937 19.176-14 30.5-3.64.873-8 1-2.854-3.608-7-5.5a40.7 40.7 0 0 1-3.5-7 192 192 0 0 1-5.5-5q-.029-5.034-4-8 1.477-2.397.5-5a44 44 0 0 0-5.5-3q-2.99-7.173-2.5-15 1.5 2 3 0 .74 2.45.5 5h6q-.465 4.82 1 9"/><path style="opacity:1" fill="#a5b913" d="M220.5 45.5q3.138.091 4.5 3l1-1a30 30 0 0 1-.5 5 4.93 4.93 0 0 1-3-.5q3.493-2.993-2-3.5a6.85 6.85 0 0 1-4 1q.282 2.405-1.5 4-1.774-1.847-3.5 0 1.71 4.548 4.5 9a26.7 26.7 0 0 0-6.5-1.5q-1.255-1.517-1-3.5a27.3 27.3 0 0 0-9 2 123 123 0 0 1-2 8q-.91-.186-1-1a64 64 0 0 1-5-5q-.616 1.143-.5 3-6.047.819-12.5 0-.255 1.983 1 3.5 2.633.303 5 1.5-1.623.373-4 .5 2 2 0 4-1.863.467-3-1.5l-3 3q-.467 1.466-2 2-4.581-4.325-7.5 1 2.916 2.541 3.5 7a5.6 5.6 0 0 1-3-1q-1.536 2.03-4 2v-4q-1.763.329-3-1-.984-3.177-1-6 1.02-2.835.5-6 1.517 1.255 3.5 1-.356-1.936 1-3a30.8 30.8 0 0 1 6.5 3q3.69-3.999 8.5-7 3.956-1.68 8 0a34 34 0 0 0 6.5-3 6.3 6.3 0 0 1 3 1q2.25-5.25 7.5-7.5 6.025-.168 11-3.5 2.03-1.536 2-4 2.872.427 5-1"/><path style="opacity:1" fill="#a2b60e" d="M154.5 74.5q-.258 2.988 2 5a28 28 0 0 0-1.5 5 87 87 0 0 1-6.5 4q2.38 2.51 2 6 5.185.25 1 3-2.382-2.42-3-6-2.453-1.77-5 0l-1-1a6.8 6.8 0 0 1 3-1.5q-3.91-3.157-9-2.5a16.6 16.6 0 0 0-7-.5q3.4 1.621.5 4.5-1.67-1.816-3.5 0l-1-1.5a36 36 0 0 0 3-7.5q1.983-.255 3.5 1 .279-3.6 1.5-7 3.808 2.032 0 4.5 2.64 6.3 8 2.5a14.7 14.7 0 0 0-5-1.5 3.65 3.65 0 0 0 1-1.5 61 61 0 0 0 8-4q-2.597-1.264-5.5 0-4.929-.302-1.5-2.5a32.4 32.4 0 0 0-8-.5q.255-1.983-1-3.5 11.716-3.565 0-4.5v-3a41.7 41.7 0 0 0 9-1q-.735 7.51 2.5 2l1 1q1.497 4.116 2 8 2.98-5.8 1.5 0 1.789 2.826 5 1.5-3.606-.762-2-3.5 2.313 1.87 4.5 0 1.513 2.288 1.5 5"/><path style="opacity:1" fill="#9eb112" d="M162.5 65.5v2q-1.356 1.064-1 3-1.983.255-3.5-1 .52 3.165-.5 6-1.083-.627-1.5-2-.575.835-1.5 1 .013-2.712-1.5-5-2.187 1.87-4.5 0-1.606 2.738 2 3.5-3.211 1.326-5-1.5 1.48-5.8-1.5 0-.503-3.884-2-8l-1-1q-3.235 5.51-2.5-2a41.7 41.7 0 0 1-9 1v3q11.716.935 0 4.5 1.255 1.517 1 3.5a32.4 32.4 0 0 1 8 .5q-3.429 2.198 1.5 2.5 2.903-1.264 5.5 0a61 61 0 0 1-8 4 3.65 3.65 0 0 1-1 1.5q2.633.303 5 1.5-5.36 3.8-8-2.5 3.808-2.468 0-4.5a26.7 26.7 0 0 0-1.5 7q-1.517-1.255-3.5-1h-4a25 25 0 0 1-.5 7q-2.766-4.55-3-10-2.468.29-5.5.5l-3 2.5q2.668 2.54 1.5 6a14.8 14.8 0 0 0-3.5-4q-3.308 3.12-8 4 2.114 9.015-4.5 2a217 217 0 0 0-2.5-6 5.73 5.73 0 0 1 1-3 25.9 25.9 0 0 0 10-1l2-4a16.14 16.14 0 0 0 7-5.5q-2.781-5.589 3-8a53 53 0 0 0 7 0q9.155-6.138 20-8.5a55 55 0 0 1 1 12z"/><path style="opacity:1" fill="#95ad04" d="M216.5 77.5q-2.514 1.025-2 4-2.405-.282-4 1.5l2 1q-5.763 6.765 2.5 2.5 1.892.39 3 2 2.014-2.103 2.5 1-1.003 3.792-2.5 0-2 6-8 8a4.46 4.46 0 0 1-1.5-2q3.616-.882 6-3.5-2.722-3.404-7-2a3.94 3.94 0 0 1 1 2q-2.057 3.478-2 7.5 3.07.016 6 1 3.103-.486 1-2.5 5.5-1 3 4-9.161.123-1 3.5a6.8 6.8 0 0 1 1.5 3l1-1a8.2 8.2 0 0 0 2.5 4l-6 1a10.8 10.8 0 0 0 6 2q-.354 3.71-4 5 1.966 1.137 1.5 3a19.6 19.6 0 0 0-3.5 2q-1 0-1 1-3.453 2.834-8 3.5a35 35 0 0 1-3 6q2.164 3.117-2 2.5-2.779 9.681-4 0 .805-3.96-2.5 0-.923-3.817 1.5-7 2 2 4 0l5-6a25.5 25.5 0 0 0-4-4.5q1.32-2.034 1-4.5a4.93 4.93 0 0 1 3 .5 31.1 31.1 0 0 0 7 8.5l1-1.5q.932-3.748-2-7 .353-1.31 1-2.5-1.497-3.248-3-6.5a48 48 0 0 0 1-5.5q-2.926-.41-4.5 2-.598-.342-.5-1a85 85 0 0 0-7.5-8 29 29 0 0 0-1.5-4q.186-.91 1-1 3.094 1.307 6 3.5-3.23 1.775 1 3.5a18.9 18.9 0 0 1 3-5 15.7 15.7 0 0 0-2-4q-7.865-2.57 0-4.5l-4-1q4.305-1.905 6-4.5 3.18 2.057 3-2 3.851.333 7 2"/><path style="opacity:1" fill="#9cb303" d="m276.5 72.5 6 3q.546 3.35 3 5.5 1.263 4.11 0 10.5 2.835-.306 4.5 2 .618 5.987 5.5 9.5a4.93 4.93 0 0 1-3 .5v3a27.6 27.6 0 0 0-5-1.5q.826-2.097 1-4a32 32 0 0 1-5-4.5q-2.571 1.663-2-2-3.245.333-5 3 .197-1.314-1-2-2.997 2.751-7 3a16.9 16.9 0 0 1 1-8q1.68 2 4 1.5-1.579-2.574 2-3a29.8 29.8 0 0 0-6-4.5q-2.78 3.053-5.5-1-2.589 1.614-5.5.5 1.386-1.275 0-3 1.894-3.948 6-3-5.239 1.632-.5 3.5a6.98 6.98 0 0 1 3.5-2.5q-2.432-3.249-2.5-7.5-2-2-4 0l-.5-1.5a324 324 0 0 0 5-7.5 10.1 10.1 0 0 0 6-1.5 6.8 6.8 0 0 1-3-1.5q5.724-2.555 5.5 4 3.974 1.567.5 5-1.329-1.238-1-3h-3v3q-5.508.817 1 1 .338 5.335-2.5 1-2.589 1.721.5 3.5 3.654.186 7-1.5"/><path style="opacity:1" fill="#97ae04" d="M196.5 66.5q-1.93 3.105-5 1-1.103 4.23-5 6 1.833 3.717 6 3v-3q4.455-.369 4-5h3q-.393-2.217 1-4 2.086 1.718 4 0a4.45 4.45 0 0 0-2-1.5q3.119-1.212 6 .5-3.116 3.477-2.5 8 2.503-2.505 2.5-6 3.02-.163 5 2a263 263 0 0 0-1 6q-4.321-1.338-8.5 1a9.3 9.3 0 0 1-2.5-2 40.8 40.8 0 0 1-8 3.5 104 104 0 0 1 5 13 2.43 2.43 0 0 1-2 .5q-.814.09-1 1a29 29 0 0 1 1.5 4 85 85 0 0 1 7.5 8h-7v-4h-4q-.644-2.18-.5-5-1.922.18-3.5-1.5 4.5-.5 5-5-2.133-4.52-5-.5-1.038-2.752 1-5.5a12.9 12.9 0 0 0-5-.5q.329-1.762-1-3a26.3 26.3 0 0 1-8 3 19.7 19.7 0 0 1 5 5.5q-2.68 2.732-4 6.5 2.016 4.734-3 4.5 2.413-7.135 1-14.5a9.4 9.4 0 0 1-3 3l-1-1.5a61 61 0 0 0 4.5-6.5l1 1a4.93 4.93 0 0 0 .5-3h-3l3-3q1.137 1.967 3 1.5 2-2 0-4 2.377-.127 4-.5a14.7 14.7 0 0 0-5-1.5q-1.255-1.517-1-3.5 6.453.819 12.5 0-.116-1.857.5-3a64 64 0 0 0 5 5"/><path style="opacity:1" fill="#91aa01" d="M254.5 58.5q1.438.68 2.5 2 .35 3.42 2.5 6 1.393-1.783 1-4h5a324 324 0 0 1-5 7.5l.5 1.5q2-2 4 0 .068 4.251 2.5 7.5a6.98 6.98 0 0 0-3.5 2.5q-4.739-1.868.5-3.5-4.106-.948-6 3 1.386 1.725 0 3 2.911 1.114 5.5-.5 2.72 4.053 5.5 1a29.8 29.8 0 0 1 6 4.5q-3.579.426-2 3-2.32.5-4-1.5a16.9 16.9 0 0 0-1 8q4.003-.249 7-3 1.197.686 1 2-.017 3.285-2 7 2.142 4.204 2 9a4.93 4.93 0 0 1-3-.5q1.255-1.517 1-3.5-4.938.57-7-4l3-4a15 15 0 0 1-4-3.5q1.673-1.853 0-3.5a81 81 0 0 1-7.5-4 4.93 4.93 0 0 1-.5-3 4.93 4.93 0 0 0-3 .5q1.32 2.034 1 4.5 4.648-.132 1 1.5 4.894 1.305 1 4 1.984 1.224 2 3.5-8.127-4.185-11.5 5a14.5 14.5 0 0 0-4.5-4.5l6-1q-3.637-.44-5.5-3.5a11.3 11.3 0 0 1-1.5 3.5 28 28 0 0 0-5 1.5 23.7 23.7 0 0 0-3-3.5q2.822-.297 5-1.5 1.496.045 2-1.5a11.1 11.1 0 0 0-2-3.5 82 82 0 0 0-7-2q-1.656 1.697 1 2.5-2.94 2.91-4-1.5 2.103-2.014-1-2.5 10.221-.113 8-10.5 5.512 0 8-5 2.668-1.754 3-5h5q-.37-5.158 1-10"/><path style="opacity:1" fill="#95ab04" d="M128.5 81.5a36 36 0 0 1-3 7.5l1 1.5q1.83-1.816 3.5 0 2.9-2.879-.5-4.5a16.6 16.6 0 0 1 7 .5q-.116 1.875-2 2.5l4 1q-1.255 1.517-1 3.5-4.718-.865-4.5-6a18.5 18.5 0 0 0-.5 6q-3.481.407-6-2-7.55 5.001 2 3 .593 4.533-3 3a4.45 4.45 0 0 0-2 1.5q3.457 1.343 7 2.5-3.424 3.236 2 4-.192 2.385-2 4-1.449-5.457-7-5-.571 3.663 2 2a42.8 42.8 0 0 1 1.5 9q4.687-1.93 9.5 0a13.1 13.1 0 0 1-7 1v4q6.375 1.948 12 5-.129 1.494 1 2.5 4.585.384 9 1.5a7.25 7.25 0 0 0 2.5-2q1.221-6.702 7-10 1 2.5 0 5-2.76-2-1.5.5a4.93 4.93 0 0 1-3 .5q-.303 6.39-3 12-2.328-2.303-4-5.5-9.561.318-18-4 1.794-.981 0-1.5l-2 1a16.1 16.1 0 0 0-5-2q.42-4.332-4-4.5a11.14 11.14 0 0 0-3.5-5.5q-2.862-4.059-2.5-9 2.24-1.218 2-4 .528-3.636-3-4.5a26.9 26.9 0 0 0 10-4.5q-2.399-3.343-6-1-3.673-.72-6.5-3-1.978 2.625-1.5 6a28.8 28.8 0 0 0-10 2q-.615 8.327-9 6.5 1.36-2.549 1-5.5-1.148-3.16 0-5.5 2.392.926 4-1l-2-2q2.306-1.665 2-4.5a217 217 0 0 1 2.5 6q6.614 7.015 4.5-2 4.692-.88 8-4a14.8 14.8 0 0 1 3.5 4q1.168-3.46-1.5-6l3-2.5q3.032-.21 5.5-.5.234 5.45 3 10 .745-3.465.5-7z"/><path style="opacity:1" fill="#c1cc63" d="M289.5 79.5q4.593 3.185 4.5 9a31.5 31.5 0 0 1 5.5 9 43.3 43.3 0 0 0-6.5-6q-.417-6.489-3.5-12"/><path style="opacity:1" fill="#9bb006" d="M157.5 75.5q.016 2.823 1 6 1.237 1.329 3 1v4q2.464.03 4-2 1.375.94 3 1 .413 3.388 3 1 3.839 2.935-1 5 1.817 1.83 0 3.5 6.326 3.906-1 4.5.393 2.217-1 4a3.65 3.65 0 0 1-1.5-1q-.669-4.743 1.5-9h-3q.1 2.188-1 4-5.082-.468-7.5 4-.94-7.308-5-1-3.63-2.445-6 1-.193-3.755-3-6-.466 1.864 1.5 3l-4 5a41.3 41.3 0 0 1 8 0 52 52 0 0 1 1 7q-2.835-.305-4.5 2a7.3 7.3 0 0 0-2-3q-1.513 11.943-6.5 2 2.46.52 5-1.5-.912-1.305-2.5-1.5-2.16.972-4.5 1v-4a4.93 4.93 0 0 0 3-.5q1.794-.981 0-1.5-2.754.791-5 2-5.424-.764-2-4a135 135 0 0 1-7-2.5 4.45 4.45 0 0 1 2-1.5q3.593 1.533 3-3-9.55 2.001-2-3 2.519 2.407 6 2a18.5 18.5 0 0 1 .5-6q-.218 5.135 4.5 6-.255-1.983 1-3.5l-4-1q1.884-.625 2-2.5 5.09-.657 9 2.5a6.8 6.8 0 0 0-3 1.5l1 1q2.547-1.77 5 0 .618 3.58 3 6 4.185-2.75-1-3 .38-3.49-2-6a87 87 0 0 0 6.5-4 28 28 0 0 1 1.5-5q-2.258-2.012-2-5 .925-.165 1.5-1 .417 1.373 1.5 2"/><path style="opacity:1" fill="#92a901" d="M174.5 75.5h3a4.93 4.93 0 0 1-.5 3l-1-1a61 61 0 0 1-4.5 6.5l1 1.5a9.4 9.4 0 0 0 3-3q1.413 7.365-1 14.5 5.016.234 3-4.5 1.32-3.768 4-6.5a19.7 19.7 0 0 0-5-5.5 26.3 26.3 0 0 0 8-3q1.329 1.238 1 3 2.549-.24 5 .5-2.038 2.748-1 5.5 2.867-4.02 5 .5-.5 4.5-5 5 1.578 1.68 3.5 1.5-.144 2.82.5 5-.436 1.725-2.5 2a42 42 0 0 1-7-2q-4.735 2.437-2.5 7a15.9 15.9 0 0 1-6.5-1q-2.217 3.223-6 4-4.595-.652-.5-2.5a10.5 10.5 0 0 0-3-2.5 19.2 19.2 0 0 0-1-6q1.1-1.812 1-4h3q-2.169 4.257-1.5 9 .621.686 1.5 1 1.393-1.783 1-4 7.327-.594 1-4.5 1.817-1.67 0-3.5 4.839-2.065 1-5-2.587 2.388-3-1-.584-4.459-3.5-7 2.919-5.325 7.5-1 1.533-.534 2-2"/><path style="opacity:1" fill="#e3e5c5" d="M283.5 62.5a31 31 0 0 0 4.5 3 30 30 0 0 0 2 9q2.191 2.074 2.5-1a31 31 0 0 0 2-2 34.3 34.3 0 0 1 1.5 8q3.77 7.953 2.5 0 2.83 1.29 5.5 3 .977 2.603-.5 5 3.971 2.966 4 8 2.74 7.855 5.5 16 3.067 1.908 5.5-1a25.9 25.9 0 0 1 1 10q5.428-2.01 13-4a47 47 0 0 1-7 6h-4a19.6 19.6 0 0 0-3.5-2q-1.893-4.885-6-8a4.93 4.93 0 0 1-.5-3 4.93 4.93 0 0 1-3-.5 2027 2027 0 0 0-5-5.5q.957-5.034-4-6a31.5 31.5 0 0 0-5.5-9q.093-5.815-4.5-9a16.3 16.3 0 0 1-2-5q-.424-6.667-4-12"/><path style="opacity:1" fill="#8ba200" d="M245.5 73.5q-2.488 5-8 5 2.221 10.387-8 10.5 3.103.486 1 2.5 1.06 4.41 4 1.5-2.656-.803-1-2.5a82 82 0 0 1 7 2 11.1 11.1 0 0 1 2 3.5q-.504 1.545-2 1.5-.467-1.466-2-2a59 59 0 0 0-8 2.5l3 2.5q-1.809 3.024 0 6 4.395-2.517 3 1.5 1.236 4.587 4 .5.894 1.824.5 4 2.42-1.683 3.5 1.5a31.8 31.8 0 0 0 7 2.5q-.255 1.983 1 3.5a79 79 0 0 1-6 6.5q-3.546-1.891-2-5.5a25.8 25.8 0 0 1-2.5-7.5q-1.313 3.02-4 0a12.9 12.9 0 0 0-.5 5h-4q.478-3.375-1.5-6-1.36 3.163-4.5 4-2.638.681-6 0a18.5 18.5 0 0 1 .5-6q1.24 2.858 3.5 1a3.65 3.65 0 0 0-1-1.5q4.452-1.894 3-6 2.41-1.574 2-4.5h-4a8.4 8.4 0 0 1-.5 4q-1.22-3.603.5-7 3.129 2.285 5-1a9.1 9.1 0 0 1-1-5h-5v-5q3.362.681 6 0a20.7 20.7 0 0 0 3.5-4q.093-2.422.5-4 5.115.175 4-5 3.839-.453 7 1"/><path style="opacity:1" fill="#98af04" d="M165.5 103.5a10.5 10.5 0 0 1 3 2.5q-4.095 1.848.5 2.5 3.783-.777 6-4a15.9 15.9 0 0 0 6.5 1q2 2 0 4-6.947-3.527-8 4-3.126 1.122-6 0a54.8 54.8 0 0 1-6 10.5l4 1-7 1a10.8 10.8 0 0 1 3 2.5q1.974-3.792 3.5 0-1.379 4.176-1.5 9h-6q-1.18-2.406-4-2 2.697-5.61 3-12a4.93 4.93 0 0 0 3-.5q-1.26-2.5 1.5-.5 1-2.5 0-5-5.779 3.297-7 10a7.25 7.25 0 0 1-2.5 2 55.6 55.6 0 0 0-9-1.5q-1.129-1.006-1-2.5 2.554 1.74 2-2-2.77.417-4-2a41 41 0 0 1 4-6.5q1.248-1.032 2 .5-1.635 1.663 0 3l2-4q4 2 2-2l1.5-1a11.8 11.8 0 0 1 3.5 4.5q1.65.648 2-1.5-3.376-5.492-4-12 3.71.354 5 4 1.428-2.318 1-5 3.868-.159 7 2"/><path style="opacity:.824" fill="#8ca302" d="M114.5 104.5a19.5 19.5 0 0 1-3-7 16.9 16.9 0 0 0-8 1q3.763 3.135-1 5.5a10.1 10.1 0 0 1-6 1.5 39.4 39.4 0 0 0-3 9.5q1.908 3.067-1 5.5 4.39 2.368 9.5 3a11.8 11.8 0 0 0 3.5-2v3q-4.292-.237-8 2a19 19 0 0 0-6-1q.803-4.108-3-5.5a7.3 7.3 0 0 1 3-2q-3.489-4.719-4-10.5-.165-.925-1-1.5 3.814-4.048 2-9a4.93 4.93 0 0 1 3-.5q.36 2.951-1 5.5 8.385 1.828 9-6.5a28.8 28.8 0 0 1 10-2q-.478-3.375 1.5-6 2.827 2.28 6.5 3 3.601-2.343 6 1a26.9 26.9 0 0 1-10 4.5q3.528.864 3 4.5.24 2.782-2 4"/><path style="opacity:1" fill="#8ba300" d="M196.5 137.5q-1.779 4.649-7 4 .544-3.617-3-4.5 2-1.5 0-3 4.15-1.004 1.5-3.5-2.35 2.034-5.5 2v4h-4a20.1 20.1 0 0 0-2-10q2.284-2.674 6-2 .782 4.828 5 3-2.963-1.163-2.5-4 3.692 3.014 7.5.5-3.763-3.135 1-5.5 1.817 1.16 0 2.5 1.423.926 2.5-.5a39.2 39.2 0 0 0 4.5-11q-5.43-1.569-11 0-3.661-.054-6-2.5 2.614-1.665 4-4.5 2.232 2.656 5.5 4l1-2q4.382 2.798 9.5.5-1.129-1.006-1-2.5h2q-.098.658.5 1 1.574-2.41 4.5-2a48 48 0 0 1-1 5.5q1.503 3.252 3 6.5a11.5 11.5 0 0 0-1 2.5q2.932 3.252 2 7l-1 1.5a31.1 31.1 0 0 1-7-8.5 4.93 4.93 0 0 0-3-.5q.32 2.466-1 4.5a25.5 25.5 0 0 1 4 4.5l-5 6q-2 2-4 0-2.423 3.183-1.5 7 3.305-3.96 2.5 0"/><path style="opacity:1" fill="#8fa701" d="M216.5 77.5q1.972 1.464 4 3-1.972 2.093 1 3 .972-2.475 3-4 2.8 4.094 3-1a6.85 6.85 0 0 0 4-1l1 2q-1.881 2.751-2 6-2.638.681-6 0v5h5a9.1 9.1 0 0 0 1 5q-1.871 3.285-5 1-1.72 3.397-.5 7a8.4 8.4 0 0 0 .5-4h4q.41 2.926-2 4.5 1.452 4.106-3 6 .686.621 1 1.5-2.26 1.858-3.5-1a18.5 18.5 0 0 0-.5 6q3.362.681 6 0 1.209 2.246 2 5-5.068 1.561-9 0 1.818 3.635 6 3-5.8 1.839-12 0a19.6 19.6 0 0 1 3.5-2q.466-1.863-1.5-3 3.646-1.29 4-5a10.8 10.8 0 0 1-6-2l6-1a8.2 8.2 0 0 1-2.5-4l-1 1a6.8 6.8 0 0 0-1.5-3q-8.161-3.377 1-3.5 2.5-5-3-4 2.103 2.014-1 2.5a19 19 0 0 0-6-1q-.057-4.022 2-7.5a3.94 3.94 0 0 0-1-2q4.278-1.404 7 2-2.384 2.618-6 3.5a4.46 4.46 0 0 0 1.5 2q6-2 8-8 1.497 3.792 2.5 0-.486-3.103-2.5-1-1.108-1.61-3-2-8.263 4.265-2.5-2.5l-2-1q1.595-1.782 4-1.5-.514-2.975 2-4"/><path style="opacity:1" fill="#d7de9a" d="M303.5 103.5a2027 2027 0 0 1 5 5.5 4.93 4.93 0 0 0 3 .5 4.93 4.93 0 0 0 .5 3q4.107 3.115 6 8a19.6 19.6 0 0 1 3.5 2 8.4 8.4 0 0 1-4-.5l-4-6a4.93 4.93 0 0 0-3-.5 2883 2883 0 0 0-7-12"/><path style="opacity:1" fill="#94ad03" d="M193.5 98.5h4v4h5q-.129 1.494 1 2.5-5.118 2.298-9.5-.5l-1 2q-3.268-1.344-5.5-4-1.386 2.835-4 4.5 2.339 2.446 6 2.5 5.57-1.569 11 0a39.2 39.2 0 0 1-4.5 11q-1.077 1.426-2.5.5 1.817-1.34 0-2.5-4.763 2.365-1 5.5-3.808 2.514-7.5-.5-.463 2.837 2.5 4-4.218 1.828-5-3-.413-3.388-3-1-4.359-.88-5-5 2.435-.21 3.5 2 2.375-2.553-1.5-4l4-1a3.65 3.65 0 0 0-1-1.5q2.305-1.665 2-4.5 2-2 0-4-2.235-4.563 2.5-7a42 42 0 0 0 7 2q2.064-.275 2.5-2"/><path style="opacity:1" fill="#95ad02" d="M240.5 97.5q-2.178 1.203-5 1.5a23.7 23.7 0 0 1 3 3.5 28 28 0 0 1 5-1.5 11.3 11.3 0 0 0 1.5-3.5q1.863 3.06 5.5 3.5l-6 1a14.5 14.5 0 0 1 4.5 4.5q3.373-9.185 11.5-5 2.146.572 2.5 3 1.61-1.17 3.5.5-3.655 2.063-8 1.5v11q5.514-.051 3.5-5 5.494 1.989 2.5 7h3q-.428 6.325 6 6.5-1.255 1.517-1 3.5 2.188.1 4-1v3a4.93 4.93 0 0 1 3 .5q-.291 2.355 0 4.5-1.952-1.467-3.5 1l-1.5-1q3.636-3.321-2-3 .22 4.581-4 6 .18-2.548 1-5l-3-5q1.392-1.902.5-4a45 45 0 0 0-5-5l-1 1q-2.066-5.648-3 1-1.638-4.78-5.5-8-1.197.686-1 2a31.8 31.8 0 0 1-7-2.5q-1.08-3.183-3.5-1.5.394-2.176-.5-4-2.764 4.087-4-.5 1.395-4.017-3-1.5-1.809-2.976 0-6l-3-2.5a59 59 0 0 1 8-2.5q1.533.534 2 2"/><path style="opacity:1" fill="#9eb604" d="M292.5 106.5a4.93 4.93 0 0 0 .5 3q2.857.987 4.5 3.5-.699 1.949 0 4.5 2.217-.393 4 1 3.205 2.203 4 6-7.011-1.727-1 2.5-4.59 1.225-8.5-1.5a24.3 24.3 0 0 0-3.5 3.5q1.18 1.97-1 3.5a11.1 11.1 0 0 0-3-4.5q1.827-2.157 2-5-1.36-2.866 0-5.5-1.855-6.357-8-9-1.536 2.03-4 2-.672-4.008-4-6 1.983-3.715 2-7 1.755-2.667 5-3-.571 3.663 2 2a32 32 0 0 0 5 4.5q-.174 1.903-1 4 2.62.528 5 1.5"/><path style="opacity:1" fill="#9ab202" d="M274.5 104.5q3.328 1.992 4 6 2.464.03 4-2 6.145 2.643 8 9-1.36 2.634 0 5.5-.173 2.843-2 5a11.1 11.1 0 0 1 3 4.5q-1.602 1.401-4 1-2.951-.359-5.5 1-1.764-3.992-5.5-6-1.812 1.1-4 1-.255-1.983 1-3.5-6.428-.175-6-6.5h-3q2.994-5.011-2.5-7 2.014 4.949-3.5 5v-11q4.345.563 8-1.5-1.89-1.67-3.5-.5-.354-2.428-2.5-3-.016-2.276-2-3.5 3.894-2.695-1-4 3.648-1.632-1-1.5.32-2.466-1-4.5a4.93 4.93 0 0 1 3-.5 4.93 4.93 0 0 0 .5 3 81 81 0 0 0 7.5 4q1.673 1.647 0 3.5a15 15 0 0 0 4 3.5l-3 4q2.062 4.57 7 4 .255 1.983-1 3.5a4.93 4.93 0 0 0 3 .5q.142-4.796-2-9"/><path style="opacity:1" fill="#7e9602" d="M114.5 104.5q-.362 4.941 2.5 9a11.14 11.14 0 0 1 3.5 5.5q4.42.168 4 4.5-10.963-5.979-13-18-6.54.76-13 2 .64 4.932-2 9a10.8 10.8 0 0 1 2.5 3 3.94 3.94 0 0 1 2-1 66 66 0 0 0 8.5 1.5q2.256 2.006 3 5 3.6.279 7 1.5.544 1.025 1 2l2 2q-1.494-.129-2.5 1-7.995-4.106-14.5-10a11.8 11.8 0 0 1-3.5 2q-5.11-.632-9.5-3 2.908-2.433 1-5.5a39.4 39.4 0 0 1 3-9.5 10.1 10.1 0 0 0 6-1.5q4.763-2.365 1-5.5a16.9 16.9 0 0 1 8-1 19.5 19.5 0 0 0 3 7"/><path style="opacity:1" fill="#9cb209" d="M164.5 97.5a19.2 19.2 0 0 1 1 6q-3.132-2.159-7-2 .428 2.682-1 5-1.29-3.646-5-4 .624 6.508 4 12-.35 2.148-2 1.5a11.8 11.8 0 0 0-3.5-4.5l-1.5 1q2 4-2 2l-2 4q-1.635-1.337 0-3-.752-1.532-2-.5a41 41 0 0 0-4 6.5q1.23 2.417 4 2 .554 3.74-2 2-5.625-3.052-12-5v-4a13.1 13.1 0 0 0 7-1q-4.813-1.93-9.5 0a42.8 42.8 0 0 0-1.5-9q-2.571 1.663-2-2 5.551-.457 7 5 1.808-1.615 2-4 2.246-1.209 5-2 1.794.519 0 1.5a4.93 4.93 0 0 1-3 .5v4q2.34-.028 4.5-1 1.588.195 2.5 1.5-2.54 2.02-5 1.5 4.987 9.943 6.5-2a7.3 7.3 0 0 1 2 3q1.665-2.305 4.5-2a52 52 0 0 0-1-7 41.3 41.3 0 0 0-8 0l4-5q-1.966-1.136-1.5-3 2.807 2.245 3 6 2.37-3.445 6-1 4.06-6.308 5 1 2.418-4.468 7.5-4"/><path style="opacity:1" fill="#93ab02" d="M246.5 126.5q.474 4.134 3 7.5-.778 3.28-4 4.5 1.61 3.927-2 6.5l2 2q-1.426 1.077-.5 2.5 8.07 9.066 5.5-2 2.951.359 5.5-1a4.46 4.46 0 0 1 1.5 2q-2.78 1.89-5.5 0-.106 2.414.5 5a22.2 22.2 0 0 1-2 5q-3.207-.722-5-3.5-1.926-1.026-3.5.5-2.376-2.042-5.5-2v-3q-1.763-.329-3 1a21 21 0 0 1 2 7q-8.692.11-4-7a26.7 26.7 0 0 0-7-1.5q1.374-.417 2-1.5 2.42-2.655 6-3.5-1.75-.75-2.5-2.5-3.1 3.772-4.5-.5a4.93 4.93 0 0 0-3-.5q-.163-6.58 5-1 .179-4.461-3.5-8-5.5 2.114-6.5-4l1-2q1.329 1.237 1 3 3.481-.406 6 2 3.058-1.165 2-3.5 2.21-1.065 2-3.5h-3q-4.182.635-6-3 3.932 1.561 9 0-.791-2.754-2-5 3.14-.837 4.5-4 1.978 2.625 1.5 6h4a12.9 12.9 0 0 1 .5-5q2.687 3.02 4 0a25.8 25.8 0 0 0 2.5 7.5q-1.546 3.609 2 5.5"/><path style="opacity:.22" fill="#b9c366" d="M339.5 111.5v3h5l1 6a5.6 5.6 0 0 1-3-1 72 72 0 0 0-4.5-6q-1.224 1.984-3.5 2 1.627-3.13 5-4"/><path style="opacity:1" fill="#bec95f" d="M299.5 97.5q4.957.966 4 6 3.53 6.02 7 12a4.93 4.93 0 0 1 3 .5l4 6a8.4 8.4 0 0 0 4 .5h4l-1 2-8-1a219 219 0 0 1-13-15q-.586-6.156-4-11"/><path style="opacity:1" fill="#617602" d="M124.5 123.5v1q-.128 1.494 1 2.5a43 43 0 0 1-5 1.5 44 44 0 0 0-1-2 26.7 26.7 0 0 0-7-1.5q-.744-2.994-3-5a66 66 0 0 1-8.5-1.5 3.94 3.94 0 0 0-2 1 10.8 10.8 0 0 0-2.5-3q2.64-4.068 2-9 6.46-1.24 13-2 2.037 12.021 13 18"/><path style="opacity:1" fill="#9eb605" d="M268.5 139.5q.058 1.884-1.5 3-.196-4.307-3.5-7 2.258-2.013 2-5-3.663.22-2-2.5-1.724-2.616-4-.5-2.024-2.175-5-2-1.906 4.381 0 7.5a32 32 0 0 0 4 2.5 30 30 0 0 1-2 5q-2.898-1.53-6-.5l2 1a6.8 6.8 0 0 0-3 1.5 14.8 14.8 0 0 0 4 3.5q-1.856.348-3 1.5 2.57 11.066-5.5 2-.926-1.423.5-2.5l-2-2q3.61-2.573 2-6.5 3.222-1.22 4-4.5-2.526-3.366-3-7.5a79 79 0 0 0 6-6.5q-1.255-1.517-1-3.5-.197-1.314 1-2 3.862 3.22 5.5 8 .934-6.648 3-1l1-1a45 45 0 0 1 5 5q.892 2.098-.5 4l3 5a20.2 20.2 0 0 0-1 5"/><path style="opacity:1" fill="#8ca300" d="M213.5 125.5q0-1 1-1 6.2 1.839 12 0h3q.21 2.435-2 3.5 1.058 2.335-2 3.5-2.519-2.406-6-2 .329-1.763-1-3l-1 2q1 6.114 6.5 4 3.679 3.539 3.5 8-5.163-5.58-5 1a4.93 4.93 0 0 1 3 .5q1.4 4.272 4.5.5.75 1.75 2.5 2.5-3.58.845-6 3.5-3.509-2.755-8-4a9.1 9.1 0 0 1-1-5q-3.124.042-5.5-2a30.9 30.9 0 0 0-3.5 9.5q1.722 2.589 3.5-.5a6.98 6.98 0 0 0 3.5 2.5q-6.856 3.087-2 9-.022 2.077-1 3.5-.925.166-1.5 1a62 62 0 0 0-10.5-1.5 4.93 4.93 0 0 1 3-.5q5.908.483 2.5-4-2.178-12.241 3.5-23a6.85 6.85 0 0 1 4-1q.904-3.811 0-7"/><path style="opacity:1" fill="#99b002" d="M268.5 139.5q4.22-1.419 4-6 5.636-.321 2 3l1.5 1q1.548-2.467 3.5-1-.291-2.145 0-4.5a4.93 4.93 0 0 0-3-.5v-3q3.736 2.008 5.5 6 2.549-1.359 5.5-1a9.1 9.1 0 0 0 1 5q-1.502 3.654-3.5 0-2.197-1.245-4.5-.5 2.556 3.356 3 7.5h-3q.329-1.763-1-3-3.863 4.05 1 4v4l-6-1q.393-2.217-1-4-2.041 3.726.5 7a4.46 4.46 0 0 1 1.5-2q1.817 1.16 0 2.5 3.893-.196 7.5.5 1.779-3.265 5.5-3-.531 3.061 0 5.5 2.895-1.389 7-1.5 4.201 1.092 7 4a7.3 7.3 0 0 0-2 3.5 41.5 41.5 0 0 1 1.5 5.5 42.5 42.5 0 0 0-7.5-4q-.266 3.721 3 5.5-2 1-4 0a87 87 0 0 0-6-6.5q-1.526 3.506-5 5 .544 1.025 1 2-.465 1.857-1.5 0l-1-3a7.25 7.25 0 0 1-2.5 2q-3.739-11.922-5.5 1-.476-2.701-2-1-.926 1.423.5 2.5-2.18 2.473-3 5.5-5.783-.8-2-4.5a300 300 0 0 0-2-12.5 7.3 7.3 0 0 1-3.5 2q-2.685-1.178-5 0-1.124 1.702.5 3 3.015 1.503 6 3-1.473 6.075-5.5 2a4.46 4.46 0 0 1-1.5 2q-2.15-2.58-2.5-6-1.648 3.816-3.5 0l3-2-.5-1q-2.281-.028-4.5-1 3 0 3-3a22.2 22.2 0 0 0 2-5q-.606-2.586-.5-5 2.72 1.89 5.5 0a4.46 4.46 0 0 0-1.5-2q-2.549 1.359-5.5 1 1.144-1.152 3-1.5a14.8 14.8 0 0 1-4-3.5 6.8 6.8 0 0 1 3-1.5l-2-1q3.102-1.03 6 .5a30 30 0 0 0 2-5 32 32 0 0 1-4-2.5q-1.906-3.119 0-7.5 2.976-.175 5 2 2.276-2.116 4 .5-1.663 2.72 2 2.5.258 2.987-2 5 3.304 2.693 3.5 7 1.558-1.116 1.5-3"/><path style="opacity:1" fill="#9bb204" d="M305.5 124.5q2.265-1.842 4.5 0a24.1 24.1 0 0 1 2.5 10h-3q.764-6.987-5.5-4-1.989-.383-3-2 .354 4.356 2.5 9 3.261.63 6 0 1.419 2.363 0 5 .476 1.721 2 2.5a12.14 12.14 0 0 1-4.5 3.5l-1-9a15.3 15.3 0 0 1-1.5 4.5q-1.423.926-2.5-.5l-1 3q-.552-2.115-2.5-3l-3 4q1.98 2.655 4.5 1 3.148 1.45 5.5 4 1.427-1.238 3-.5-4.038 4.256-2 9.5h6q-.628 4.237 3 6.5-1.701 1.524 1 2-1.75.75-2.5 2.5l-1-5q-.625 1.884-2.5 2 .393-2.217-1-4-1.182 7.966-7.5 3-3.83 1.084-6.5 4a28 28 0 0 0-5-1.5q-1.32-2.034-1-4.5h-3v3h-3a44 44 0 0 0-1-2q3.474-1.494 5-5a87 87 0 0 1 6 6.5q2 1 4 0-3.266-1.779-3-5.5a42.5 42.5 0 0 1 7.5 4 41.5 41.5 0 0 0-1.5-5.5 7.3 7.3 0 0 1 2-3.5q-2.799-2.908-7-4-4.105.111-7 1.5-.531-2.439 0-5.5-3.721-.265-5.5 3-3.607-.696-7.5-.5 1.817-1.34 0-2.5a4.46 4.46 0 0 0-1.5 2q-2.541-3.274-.5-7 1.393 1.783 1 4l6 1v-4q-4.863.05-1-4 1.329 1.237 1 3h3q-.444-4.144-3-7.5 2.303-.745 4.5.5 1.998 3.654 3.5 0a9.1 9.1 0 0 1-1-5q2.398.401 4-1 2.18-1.53 1-3.5a24.3 24.3 0 0 1 3.5-3.5q3.91 2.726 8.5 1.5-6.011-4.227 1-2.5"/><path style="opacity:1" fill="#a7bc17" d="M282.5 72.5a63 63 0 0 1 5 6.5 2.43 2.43 0 0 0 2 .5q3.083 5.511 3.5 12a43.3 43.3 0 0 1 6.5 6q3.414 4.844 4 11a219 219 0 0 0 13 15q-1.426 2.128-1 5 4.019.26 7 3-1.783 1.393-4 1a25 25 0 0 1-.5 7q-.636-5.535-2.5-.5a4.93 4.93 0 0 1-3 .5v-5q-.165-5.25-2.5-10-2.235-1.842-4.5 0-.795-3.797-4-6-1.783-1.393-4-1-.699-2.551 0-4.5-1.643-2.513-4.5-3.5a4.93 4.93 0 0 1-.5-3v-3a4.93 4.93 0 0 0 3-.5q-4.882-3.513-5.5-9.5-1.665-2.306-4.5-2 1.263-6.39 0-10.5-2.454-2.15-3-5.5l-6-3q-.045-1.496 1.5-2a27.6 27.6 0 0 1 4.5 2"/><path style="opacity:1" fill="#96ac06" d="M87.5 107.5q.511 5.781 4 10.5a7.3 7.3 0 0 0-3 2q3.803 1.392 3 5.5 3.07.016 6 1 3.708-2.237 8-2v-3q6.505 5.894 14.5 10 1.006-1.129 2.5-1l1 2v1q1.662 3.292-.5 6-3.66-4.755-8.5-3.5a7.25 7.25 0 0 1 2 2.5 123 123 0 0 1-2 8 18.5 18.5 0 0 0-4-5q-4.734 2.73-6 8a4.93 4.93 0 0 0 1 2.5q-4.242.636-4-4 .383-1.989 2-3-2.964-.282-6 .5-7.509-2.697 0-4.5a41 41 0 0 0-9-.5q-.026 6.115 4 10-4.533.108-6.5-4-.233-15.119-1.5-28-1.318-4.5 2-8.5-2.074-2.191 1-2.5"/><path style="opacity:1" fill="#93ab02" d="M181.5 109.5q.305 2.835-2 4.5.686.621 1 1.5l-4 1q3.875 1.447 1.5 4-1.065-2.21-3.5-2 .641 4.12 5 5 2.587-2.388 3 1-3.716-.674-6 2a20.1 20.1 0 0 1 2 10h4v-4q3.15.034 5.5-2 2.65 2.496-1.5 3.5 2 1.5 0 3 3.544.883 3 4.5 5.221.649 7-4 1.221 9.681 4 0 4.164.617 2-2.5a35 35 0 0 0 3-6q4.547-.666 8-3.5.904 3.189 0 7a6.85 6.85 0 0 0-4 1q-5.678 10.759-3.5 23 3.408 4.483-2.5 4-.853-5.571-4.5-2-3.03-.798-2.5-4 2.217.393 4-1l-7-2q3.162-4.146 1-9a51 51 0 0 1-4.5 5 3.65 3.65 0 0 0-1.5-1 22.8 22.8 0 0 1-5 4q-2.205-4.633 2-7.5-3.42-.35-6-2.5l-1 1q3.46 2.742-1 5-2.415-1.578-5 0a15.7 15.7 0 0 1-4-2.5q1.255-1.517 1-3.5-7.201 2.758-12-3h6q.121-4.824 1.5-9-1.526-3.792-3.5 0a10.8 10.8 0 0 0-3-2.5l7-1-4-1a54.8 54.8 0 0 0 6-10.5q2.874 1.122 6 0 1.053-7.527 8-4"/><path style="opacity:.055" fill="#9fb314" d="M345.5 120.5a168 168 0 0 0 4.5 5 168 168 0 0 0 6.5 15.5q-1.32 2.034-1 4.5-3.496.003-6-2.5l1-1a12.9 12.9 0 0 0-5-.5q-2.128-1.426-5-1 .572-2.146 3-2.5l-4-1a14.7 14.7 0 0 0 5-1.5q-3.063-1.505-1-4-2.316-1.683-4 1l-2-2q-2.552 1.953-3.5 5a20.4 20.4 0 0 0-1.5-5.5q6.238-.584 1.5-3.5-1.948.699-4.5 0a6.85 6.85 0 0 0 1 4q-1.834 1.671-4 .5 4.273-5.641-2-6.5l1-2a47 47 0 0 0 7-6l2-1q2.276-.016 3.5-2a72 72 0 0 1 4.5 6 5.6 5.6 0 0 0 3 1"/><path style="opacity:1" fill="#8a9f02" d="M124.5 124.5v-1a16.1 16.1 0 0 1 5 2l2-1q1.794.519 0 1.5 8.439 4.318 18 4 1.672 3.197 4 5.5 2.82-.406 4 2 4.799 5.758 12 3 .255 1.983-1 3.5a15.7 15.7 0 0 0 4 2.5q2.585-1.578 5 0 4.46-2.258 1-5l1-1q2.58 2.15 6 2.5-4.205 2.867-2 7.5a22.8 22.8 0 0 0 5-4 3.65 3.65 0 0 1 1.5 1 51 51 0 0 0 4.5-5q2.162 4.854-1 9l7 2q-1.783 1.393-4 1-.53 3.202 2.5 4 3.647-3.571 4.5 2a4.93 4.93 0 0 0-3 .5q5.384.31 10.5 1.5.575-.834 1.5-1 .572 2.146 3 2.5-.834.575-1 1.5-1.494-.129-2.5 1-5.639-3.487-12.5-4.5-4.465-3.567-10-4.5v-3q-2.34.028-4.5 1-1.753-.763-2.5-2.5a34.3 34.3 0 0 1-8-1.5q-2.29-3.659-6-6-2-2-4 0a98 98 0 0 1-7-5.5q-5.423-.92-8.5-5.5-3.403 3.831-7 0l-1 1-1-4-1 1a102 102 0 0 0-14.5-8"/><path style="opacity:.004" fill="#b2c03c" d="M362.5 137.5h3q-1.119 5.715-3 0"/><path style="opacity:1" fill="#92a906" d="M340.5 140.5q2.872-.426 5 1 .129 1.494-1 2.5 2.686.623 4 3a30.3 30.3 0 0 1-10-.5v3h4a6.85 6.85 0 0 1-1 4q3.114.9 6 1l-1 2a65 65 0 0 1 4 9 8.2 8.2 0 0 0 4 2.5q-1.25 2.25-3.5 3.5.478 3.142-.5 6 .571 3.31 4 4 1.428 2.318 1 5-2 .5 0 1h1q.166.925 1 1.5a74 74 0 0 0-5 10.5q-1.938.393-4-1.5 1.17-1.61-.5-3.5-1.937 1.937-3.5 0 5.2-5.077 4-12 .965-7.797 2-15.5a25.9 25.9 0 0 1-6.5-5.5q-.977-2.603.5-5a14.7 14.7 0 0 0-5-1.5q-1.957-3.222-5.5-4.5l-1 3q-1.574-1.526-3.5-.5a20.7 20.7 0 0 1 4 3.5l-1 16q-1.792 4.78-7 5 .643-2.857-2-5.5 2.32-.5 4 1.5 1.858-2.26-1-3.5a4.46 4.46 0 0 0 2-1.5q-3.872-3.26 0-6.5-4-1 0-2-3.369-1.885-4-5.5h4q-.237-4.568 2-8.5-1.32-2.034-1-4.5h-3q.255-1.983-1-3.5 6.314-1.784 11 2.5 2.047-1.96 4 0"/><path style="opacity:1" fill="#798c01" d="M124.5 124.5a102 102 0 0 1 14.5 8l1-1 1 4 1-1q3.597 3.831 7 0 3.077 4.58 8.5 5.5a98 98 0 0 0 7 5.5q2-2 4 0 3.71 2.341 6 6a34.3 34.3 0 0 0 8 1.5q.747 1.737 2.5 2.5 2.16-.972 4.5-1v3q5.535.933 10 4.5 6.861 1.013 12.5 4.5 1.006-1.129 2.5-1 .5 1.5 2 2l8 4q3.52 2.295.5 6a8.4 8.4 0 0 1-.5-4h-3q.329 1.763-1 3a20 20 0 0 1-4-3.5q-3.684.096-7-.5-3.558-3.25-8-1v-4a82.7 82.7 0 0 0-15.5-4 53.3 53.3 0 0 1-6.5-7q-1.147 1.668-3 2-1.523-5.762-7-3.5-2.289-2.83-6-3 1.701-1.524-1-2a99.4 99.4 0 0 0-17-6.5 37 37 0 0 0-2-4.5q-7.51-3.659-15-6a12.9 12.9 0 0 0-5-.5l-1-2-2-2a43 43 0 0 0 5-1.5q-1.128-1.006-1-2.5"/><path style="opacity:1" fill="#8ea401" d="M123.5 133.5q4.569.337 7 4 .116 1.875 2 2.5-1.237 2.532-3 .5-3.071 2.273-7 2.5-2.064 3.814-3.5.5-3.831 7.03-.5 14-.544 1.025-1 2a12.9 12.9 0 0 1-.5 5l-1-8q-2.776 3.846-5.5 0a9.3 9.3 0 0 1-2.5 2l-4-4-1.5.5q3.561 3.295 0 7 3.784 3.861-.5 6.5-.575-.834-1.5-1 .29-5.27-2-10-3.132-2.892-4.5-7-1.035 1.857-1.5 0-4.026-3.885-4-10a41 41 0 0 1 9 .5q-7.509 1.803 0 4.5 3.036-.782 6-.5-1.617 1.011-2 3-.242 4.636 4 4a4.93 4.93 0 0 1-1-2.5q1.266-5.27 6-8a18.5 18.5 0 0 1 4 5 123 123 0 0 0 2-8 7.25 7.25 0 0 0-2-2.5q4.84-1.255 8.5 3.5 2.162-2.708.5-6"/><path style="opacity:1" fill="#99b003" d="M226.5 148.5q-.626 1.083-2 1.5 3.6.279 7 1.5-4.692 7.11 4 7a21 21 0 0 0-2-7q1.237-1.329 3-1v3q3.124-.042 5.5 2 1.574-1.526 3.5-.5 1.793 2.778 5 3.5 0 3-3 3-.99 2.196-1 5 .255 3.322 2 6l-1.5 1q-3.612-3.797-4.5-9-3.138-.091-4.5-3-.375 7.126-7.5 8-1.187-.849 0-2.5-.956-2.569-3-.5-1.329-1.237-1-3-5.458.99-10 4-1.5-.5-2-2 .166-.925 1-1.5-2.428-.354-3-2.5.978-1.423 1-3.5-4.856-5.913 2-9a6.98 6.98 0 0 1-3.5-2.5q-1.778 3.089-3.5.5a30.9 30.9 0 0 1 3.5-9.5q2.376 2.042 5.5 2a9.1 9.1 0 0 0 1 5q4.491 1.245 8 4"/><path style="opacity:1" fill="#8ea502" d="M163.5 163.5q-2.6-.762-1-2.5-1.574-.738-3 .5a4.46 4.46 0 0 1-1.5-2 10.3 10.3 0 0 1-1.5 4l-1-4a13 13 0 0 1-4 4.5q1.752 2.108-1 3.5a14.7 14.7 0 0 0 5 1.5q-3.229.149-5 2.5a44 44 0 0 0-4.5 4q-.81-2.452 1.5-4l-3-2q1.752-2.108-1-3.5 4.244-.072 7-3h-6q5.8-3.512-1-1.5a2.43 2.43 0 0 1-2-.5q1.794-.981 0-1.5a57 57 0 0 0-7 2.5q4.248 2.862-1 2.5-3.107-3.788-4-9-1.834-1.671-4-.5l2 1q-1.595 1.782-4 1.5.406-3.481-2-6 3.089-5.046 2-.5 1.64 2.048 4 1-2.454-2.15-3-5.5 1.828.586 2.5-2 2.093 1.972 3-1 5.047 2.886 2.5 8 1.594 3.392 4 6 1.475-6.878 0-13 2.994-2.75 7-1.5-1.5 3.136 2 5a66 66 0 0 1 8.5 1.5q3.003 3.012 6 6a4.46 4.46 0 0 1 1.5-2q2.735 1.87 6 2v6q-2.975-.514-4 2"/><path style="opacity:1" fill="#90a504" d="M117.5 159.5q1.337 1.142 2 3a55.6 55.6 0 0 0-1.5 9q-1.077 1.426-2.5.5 1.846-1.774 0-3.5-2 2.5 0 5a37 37 0 0 1 4.5 2q1.66-1.818 1.5 1-.863 10.72-8 2-1.382 2.297 1 4-.442 2.885-2.5 5-6.343-4.437-9.5-11.5a11.1 11.1 0 0 1 2-3.5q-1.526-3.792-3.5 0-2.181-2.616-.5-5 .925.166 1.5 1 4.284-2.639.5-6.5 3.561-3.705 0-7l1.5-.5 4 4a9.3 9.3 0 0 0 2.5-2q2.724 3.846 5.5 0l1 8a12.9 12.9 0 0 0 .5-5"/><path style="opacity:1" fill="#89a102" d="M123.5 133.5v-1a12.9 12.9 0 0 1 5 .5q7.49 2.341 15 6a37 37 0 0 1 2 4.5 99.4 99.4 0 0 1 17 6.5q2.701.476 1 2 3.711.17 6 3 5.477-2.262 7 3.5 1.354 4.337 1 9-5.056-1.957-4 4h-4q.059 2.788-2.5 4a8.4 8.4 0 0 1-.5-4q-4.483-1.703-9-.5.957-2.56 0-5.5h3a8.4 8.4 0 0 0 .5 4 23.4 23.4 0 0 0 2.5-6q1.025-2.514 4-2v-6q-3.265-.13-6-2a4.46 4.46 0 0 0-1.5 2 2828 2828 0 0 1-6-6 66 66 0 0 0-8.5-1.5q-3.5-1.864-2-5-4.006-1.25-7 1.5-3.388-.413-1-3-1.91-2.927-5-4-2.431-3.663-7-4"/><path style="opacity:1" fill="#97ad03" d="m316.5 123.5 8 1q6.273.859 2 6.5 2.166 1.171 4-.5a6.85 6.85 0 0 1-1-4q2.552.699 4.5 0 4.738 2.916-1.5 3.5a20.4 20.4 0 0 1 1.5 5.5q.948-3.047 3.5-5l2 2q1.684-2.683 4-1-2.063 2.495 1 4a14.7 14.7 0 0 1-5 1.5l4 1q-2.428.354-3 2.5-1.953-1.96-4 0-4.686-4.284-11-2.5 1.255 1.517 1 3.5h3q-.32 2.466 1 4.5-2.237 3.932-2 8.5h-4q.631 3.615 4 5.5-4 1 0 2-3.872 3.24 0 6.5a4.46 4.46 0 0 1-2 1.5q2.858 1.24 1 3.5-1.68-2-4-1.5 2.643 2.643 2 5.5h-3q-.255-1.983 1-3.5-3.818-3.097-7 .5a12 12 0 0 0-6-1v-4q1.875-.116 2.5-2l1 5q.75-1.75 2.5-2.5-2.701-.476-1-2-3.628-2.263-3-6.5h-6q-2.038-5.244 2-9.5-1.573-.738-3 .5a17.2 17.2 0 0 0-5.5-4q-2.52 1.655-4.5-1l3-4q1.948.885 2.5 3l1-3q1.077 1.426 2.5.5a15.3 15.3 0 0 0 1.5-4.5l1 9a12.14 12.14 0 0 0 4.5-3.5q-1.524-.779-2-2.5 1.419-2.637 0-5-2.739.63-6 0-2.146-4.644-2.5-9 1.011 1.617 3 2 6.264-2.987 5.5 4h3v5a4.93 4.93 0 0 0 3-.5q1.864-5.035 2.5.5.745-3.465.5-7 2.217.393 4-1-2.981-2.74-7-3-.426-2.872 1-5"/><path style="opacity:.055" fill="#a9b353" d="M84.5 153.5q0 1 1 1 3.59 1.065 7.5 2 1.345 4.485 2.5 9-1.905 1.182-3 3-1 0-1 1v1q-.09.814-1 1a75 75 0 0 0-7-12q.63-3.063 1-6"/><path style="opacity:1" fill="#8ba201" d="M130.5 137.5q3.09 1.073 5 4-2.388 2.587 1 3 1.475 6.122 0 13-2.406-2.608-4-6 2.547-5.114-2.5-8-.907 2.972-3 1-.672 2.586-2.5 2 .546 3.35 3 5.5-2.36 1.048-4-1 1.089-4.546-2 .5 2.406 2.519 2 6 2.405.282 4-1.5l-2-1q2.166-1.171 4 .5.893 5.212 4 9 5.248.362 1-2.5a57 57 0 0 1 7-2.5q1.794.519 0 1.5a2.43 2.43 0 0 0 2 .5q-4.115 6.06-3.5 13-1.448-2.64-3.5-1 5.023 4.813 7 11.5-2 1-4 0a3.65 3.65 0 0 0 1-1.5q-2.175-2.024-2-5-5.208.561-9-3-2.024 2.175-5 2-.255-7.162-3-1 .16-2.818-1.5-1a37 37 0 0 0-4.5-2q-2-2.5 0-5 1.846 1.726 0 3.5 1.423.926 2.5-.5a55.6 55.6 0 0 1 1.5-9q-.663-1.858-2-3 .456-.975 1-2-3.331-6.97.5-14 1.436 3.314 3.5-.5 3.929-.227 7-2.5 1.763 2.032 3-.5-1.884-.625-2-2.5"/><path style="opacity:1" fill="#738a02" d="M343.5 168.5h-3a4.93 4.93 0 0 0-.5-3q-2.014 1.07-4-1l-1 7q-1.006 1.129-2.5 1l1-16q2.618-1.077 2-3a56 56 0 0 0 6.5 6q.068 4.251 2.5 7.5-.834.575-1 1.5"/><path style="opacity:.429" fill="#9cb207" d="M345.5 141.5a12.9 12.9 0 0 1 5 .5l-1 1q2.504 2.503 6 2.5a106.6 106.6 0 0 1 5 20q-3.946 2.257 1 2-.27 6.114-3 11.5.826 2.097 1 4a11.1 11.1 0 0 0-3 4.5h-1v-1q.428-2.682-1-5-3.429-.69-4-4 .978-2.858.5-6 2.25-1.25 3.5-3.5a8.2 8.2 0 0 1-4-2.5 65 65 0 0 0-4-9l1-2q-2.886-.1-6-1a6.85 6.85 0 0 0 1-4h-4v-3q4.976 1.082 10 .5-1.314-2.377-4-3 1.129-1.006 1-2.5"/><path style="opacity:1" fill="#617a01" d="M343.5 168.5q2.914 3.707 3 9-4.322 5.488-6 12.5-3.347.933-5.5 3.5-1.115-3.648-4.5-5h5q.366-4.144 4-6.5l-2-1q8.945.748 3-4.5 3.731-4.155-2.5-5a3.94 3.94 0 0 0-2 1q-.217 4.487-4.5 5 1.201-2.312 1-5 1.494.129 2.5-1l1-7q1.986 2.07 4 1 .724 1.422.5 3z"/><path style="opacity:1" fill="#95ab02" d="M283.5 169.5h3v-3h3q-.32 2.466 1 4.5 2.577.519 5 1.5 2.67-2.916 6.5-4 6.318 4.966 7.5-3 1.393 1.783 1 4v4a12 12 0 0 1 6 1q3.182-3.597 7-.5-1.255 1.517-1 3.5a89 89 0 0 1-9 4 47 47 0 0 1-6.5-5q-1.904 2.107-1.5 5h-3v-3q-7.86-1.37 0-3-1.98-2.655-4.5-1a18.5 18.5 0 0 0-.5 6 12 12 0 0 0-6 1q-1.466-5.38-6.5-8a21.3 21.3 0 0 1-4 5q-2.376-2.042-5.5-2 .282-2.405-1.5-4-.939 3.737-2.5 0l-1 1a32.4 32.4 0 0 1 3 4.5q-3.261 1.316-5-1.5.82-3.027 3-5.5-1.426-1.077-.5-2.5 1.524-1.701 2 1 1.761-12.922 5.5-1a7.25 7.25 0 0 0 2.5-2l1 3q1.035 1.857 1.5 0"/><path style="opacity:1" fill="#899f01" d="M163.5 163.5a23.4 23.4 0 0 1-2.5 6 8.4 8.4 0 0 1-.5-4h-3q.957 2.94 0 5.5 4.517-1.203 9 .5a8.4 8.4 0 0 0 .5 4q2.559-1.212 2.5-4h4q-1.056-5.957 4-4 3.855 1.824 7.5 2a4.93 4.93 0 0 0 .5-3q3.46.729 6.5 0 3.134 4.778-2.5 7 3.894 2.696-1 4l-2-1q-3.679 5.556-6.5 0 .358-5.526-4.5-8-1.329 1.237-1 3h3a32.5 32.5 0 0 1-.5 8q-2.838 1.676-2.5 5h3q-1.279 3.64-5 2.5 3.442-1.476-1-2.5-2.072 4.227 3 4.5-.343.598-1 .5-1.935.837-2.5 3a3.65 3.65 0 0 0-1.5-1 18.5 18.5 0 0 0-4 5q.635-3.816 0-8a9.4 9.4 0 0 0 3-3q-1.918-10.803-11.5-5-1.418-.608-2-2a8.4 8.4 0 0 0-.5 4 12 12 0 0 1 6 1q.858 3.626 3 6.5-2.244 4.458-4-1-5.591-.009-8.5-4.5l-4.5 3q2.24.962 3 3a25 25 0 0 1-.5 7q-2.23.983-3.5 3 2.575 4.972-2 3-1.951-2.076 0-4.5a4.93 4.93 0 0 0-3-.5q.2-5.094 3-1 1.329-1.237 1-3 3.726-3.227 0-7 2.242-4.02 7-4.5-3.14-2.028-7-1.5-.059-2.788 2.5-4 .677 3.644 2.5.5a3.65 3.65 0 0 1-1-1.5q2.175-2.024 2-5 1.771-2.351 5-2.5a14.7 14.7 0 0 1-5-1.5q2.752-1.392 1-3.5a13 13 0 0 0 4-4.5l1 4a10.3 10.3 0 0 0 1.5-4 4.46 4.46 0 0 0 1.5 2q1.426-1.238 3-.5-1.6 1.738 1 2.5"/><path style="opacity:1" fill="#859b01" d="M209.5 172.5q-.129 1.494 1 2.5 2.316.434 4 2-2.305 1.665-2 4.5h5v3q-2.464.03-4-2-2.682 1.174-3 4-2.554-1.74-2 2h-3l-1-6h-5q.492-7.352-4-2.5a52 52 0 0 0-10 1.5 51 51 0 0 1-2 6q-1.772-2.453 0-5l-2-1a14.2 14.2 0 0 0-4 3h-3q-.338-3.324 2.5-5 .746-3.97.5-8h-3q-.329-1.763 1-3 4.858 2.474 4.5 8 2.821 5.556 6.5 0l2 1q4.894-1.304 1-4 5.634-2.222 2.5-7-3.04.729-6.5 0a4.93 4.93 0 0 1-.5 3q-3.645-.176-7.5-2 .354-4.663-1-9 1.853-.332 3-2a53.3 53.3 0 0 0 6.5 7 82.7 82.7 0 0 1 15.5 4v4q4.442-2.25 8 1"/><path style="opacity:1" fill="#839801" d="M224.5 171.5q.353-.895 1.5-1 3.719 3.182 8 2 .582 1.392 2 2 6.322-2.726 10 3 4.949-1.455 10.5 0-.1 2.188 1 4-4.535-.132-8-3-2.028 3.204-5 1-1.329 1.237-1 3-2.375 1.69-6 0a8.4 8.4 0 0 0-.5-4q-2.631 2.625-6.5 2 1.556 7.705-5.5 5a24.3 24.3 0 0 0-3.5 3.5q2.822.297 5 1.5-1.473.961-3 2a9.3 9.3 0 0 0-2.5-2q-4.5.429-9.5 0-1.1-1.812-1-4 .318-2.826 3-4 1.536 2.03 4 2v-3h-5q-.305-2.835 2-4.5-1.684-1.566-4-2-1.129-1.006-1-2.5 3.316.596 7 .5a20 20 0 0 0 4 3.5q1.329-1.237 1-3h3a8.4 8.4 0 0 0 .5 4q3.02-3.705-.5-6"/><path style="opacity:1" fill="#879c00" d="M150.5 171.5q.175 2.976-2 5a3.65 3.65 0 0 0 1 1.5q-1.823 3.144-2.5-.5-2.559 1.212-2.5 4 3.86-.528 7 1.5-4.758.48-7 4.5 3.726 3.773 0 7-1.188-2.636-4-4 2.03-1.536 2-4-6.496 1.573-6.5-5l-1 1q-2.198-3.623-6.5-3v3q3.644.127 6 3a4.46 4.46 0 0 1-1.5 2q-5.371-2.115-11.5-5a4.93 4.93 0 0 0 .5 3q3.82 1.036 6.5 4a71 71 0 0 0-4 2q-.975.456-2 1-1.962-5.734-8-6.5 1.936-1.937 0-3.5-2.382-1.703-1-4 7.137 8.72 8-2 2.745-6.162 3 1 2.976.175 5-2 3.792 3.561 9 3-.175 2.976 2 5a3.65 3.65 0 0 1-1 1.5q2 1 4 0-1.977-6.687-7-11.5 2.052-1.64 3.5 1-.615-6.94 3.5-13 6.8-2.012 1 1.5h6q-2.756 2.928-7 3 2.752 1.392 1 3.5l3 2q-2.31 1.548-1.5 4a44 44 0 0 1 4.5-4"/><path style="opacity:1" fill="#8fa402" d="M268.5 176.5q1.739 2.816 5 1.5a32.4 32.4 0 0 0-3-4.5l1-1q1.561 3.737 2.5 0 1.782 1.595 1.5 4 3.124-.042 5.5 2a21.3 21.3 0 0 0 4-5q5.034 2.62 6.5 8a12 12 0 0 1 6-1 18.5 18.5 0 0 1 .5-6q2.52-1.655 4.5 1-7.86 1.63 0 3v3h3q-1.144 1.152-3 1.5-9.271.258-18 1.5a77.4 77.4 0 0 1-16-3q-5.454 1.545-11 0-1.1-1.812-1-4-5.551-1.455-10.5 0-3.678-5.726-10-3-1.418-.608-2-2-4.281 1.182-8-2-1.147.105-1.5 1l-8-4q4.542-3.01 10-4-.329 1.763 1 3 2.044-2.069 3 .5-1.187 1.651 0 2.5 7.125-.874 7.5-8 1.362 2.909 4.5 3 .888 5.203 4.5 9l1.5-1q-1.745-2.678-2-6 .01-2.804 1-5 2.219.972 4.5 1l.5 1-3 2q1.852 3.816 3.5 0 .35 3.42 2.5 6a4.46 4.46 0 0 0 1.5-2q4.027 4.075 5.5-2-2.985-1.497-6-3-1.624-1.298-.5-3 2.315-1.178 5 0a7.3 7.3 0 0 0 3.5-2 300 300 0 0 1 2 12.5q-3.783 3.7 2 4.5"/><path style="opacity:1" fill="#4e6402" d="M330.5 188.5q-1.052-3.316-4-5.5a26 26 0 0 1 5-5.5q4.283-.513 4.5-5a3.94 3.94 0 0 1 2-1q6.231.845 2.5 5 5.945 5.248-3 4.5l2 1q-3.634 2.356-4 6.5z"/><path style="opacity:1" fill="#899f02" d="M243.5 182.5q1.467.467 2 2 6.492.768 12 4 11.324-2.548 22 2 5.48-1.608 11 0a172 172 0 0 0 24-2 13.6 13.6 0 0 1 6-3q2.241.314 3.5 2a4.46 4.46 0 0 1 1.5-2 514 514 0 0 1 3 4q-6.704.049-12.5-1-2.532 2.317-6 3l-1-1a4.93 4.93 0 0 0-.5 3q-2.816-.6-4.5-3-1.418.608-2 2-1.334-2.083-1.5.5.608 1.418 2 2a8.2 8.2 0 0 0-4 2.5q-1.817-1.16 0-2.5-1.116-1.558-3-1.5-.926-.166-1.5-1a6.55 6.55 0 0 1-3 3q-.76-2.244 1.5-3.5-4.795-1.48-8.5 1.5-1.77-1.776-4.5-2-5.292.903-11 0-8.836-2.902-18-1 1.237 1.329 3 1a34.2 34.2 0 0 1-1 10q-2.312-1.201-5-1a4.93 4.93 0 0 1-.5 3l-3-7q-1.135 4.856-1.5 9-5.248.362-1-2.5a20.5 20.5 0 0 0-3.5-2.5q-1.35 1.905-3.5 3 4.02 6.447-4 7v-3q4.91.044 2-4a30.6 30.6 0 0 1 3-5q3.628-.498 7 0-2.404-4.106-2.5-9a6.55 6.55 0 0 1-3-3l-4 4q-.582-1.392-2-2-1.95 1.81-4.5 2-2.178-1.203-5-1.5a24.3 24.3 0 0 1 3.5-3.5q7.056 2.705 5.5-5 3.869.625 6.5-2a8.4 8.4 0 0 1 .5 4q3.625 1.69 6 0"/><path style="opacity:.004" fill="#98a521" d="M100.5 184.5q-1 1 0 0"/><path style="opacity:1" fill="#748601" d="M257.5 181.5q5.546 1.545 11 0a77.4 77.4 0 0 0 16 3q9.196-.888 17.5 2a101 101 0 0 1 7.5-3 22.5 22.5 0 0 0 3 4l-1.5 1a18.2 18.2 0 0 0-3.5-2q-8.442 3.082-17 4-5.52-1.608-11 0-10.676-4.548-22-2-5.508-3.232-12-4-.533-1.533-2-2-.329-1.763 1-3 2.972 2.204 5-1 3.465 2.868 8 3"/><path style="opacity:1" fill="#8b9e02" d="M144.5 194.5q.329 1.763-1 3-2.8-4.094-3 1a4.93 4.93 0 0 1 3 .5q-1.951 2.424 0 4.5-.042 3.124 2 5.5-3.147-1.476-6 1 1.752 2.109-1 3.5a4.46 4.46 0 0 0 2 1.5q-6.215.57-2-4a6.98 6.98 0 0 1-2.5-3.5q-3.643-.794-5.5-4-5.8 1.48 0-1.5l-4-2q.834-.575 1-1.5-.572-2.146-3-2.5.699-1.949 0-4.5a71 71 0 0 1 4-2q-2.68-2.964-6.5-4a4.93 4.93 0 0 1-.5-3q6.129 2.885 11.5 5a4.46 4.46 0 0 0 1.5-2q-2.356-2.873-6-3v-3q4.302-.623 6.5 3l1-1q.004 6.573 6.5 5 .03 2.464-2 4 2.812 1.364 4 4"/><path style="opacity:1" fill="#7e9402" d="M332.5 172.5q.201 2.688-1 5a26 26 0 0 0-5 5.5q2.948 2.184 4 5.5 3.385 1.352 4.5 5 2.153-2.567 5.5-3.5 1.678-7.012 6-12.5 1.763 5.475 0 11h-3v5a12.9 12.9 0 0 0-5 .5q1.846 1.774 0 3.5a6.85 6.85 0 0 0-4-1q.174 5.115-5 4v-4q-1.763.329-3-1 2.38-2.51 2-6a514 514 0 0 0-3-4 4.46 4.46 0 0 0-1.5 2q-1.259-1.686-3.5-2a13.6 13.6 0 0 0-6 3 172 172 0 0 1-24 2q8.558-.918 17-4a18.2 18.2 0 0 1 3.5 2l1.5-1a22.5 22.5 0 0 1-3-4 101 101 0 0 0-7.5 3q-8.304-2.888-17.5-2a159 159 0 0 1 18-1.5q1.856-.348 3-1.5-.404-2.893 1.5-5a47 47 0 0 0 6.5 5 89 89 0 0 0 9-4h3q5.208-.22 7-5"/><path style="opacity:1" fill="#93a904" d="M268.5 191.5q5.708.903 11 0 2.73.224 4.5 2 3.705-2.98 8.5-1.5-2.26 1.256-1.5 3.5a6.55 6.55 0 0 0 3-3q.574.834 1.5 1 .129 1.494-1 2.5a64 64 0 0 1 4 4.5 4.46 4.46 0 0 0 1.5-2 20.4 20.4 0 0 0 1.5 5.5q-3.79 2.96-2-1.5-10.43-4.826-4 4.5 7.476-.535 1 5 2.701.476 1 2a4.93 4.93 0 0 0 3 .5v3a52.2 52.2 0 0 0-11-5q-.768-4.276-5-4 .674-3.716-2-6l-1 6a9.3 9.3 0 0 0-2.5-2 29.8 29.8 0 0 1-8.5 3q-4.839-7.812-7-15 4.178.876 5-3"/><path style="opacity:1" fill="#99a816" d="M92.5 150.5q.465 1.857 1.5 0 1.368 4.108 4.5 7 2.29 4.73 2 10-1.681 2.384.5 5 1.974-3.792 3.5 0a11.1 11.1 0 0 0-2 3.5q3.157 7.063 9.5 11.5 2.058-2.115 2.5-5 1.936 1.563 0 3.5 6.038.766 8 6.5 1.025-.544 2-1 .699 2.551 0 4.5 2.428.354 3 2.5-4.164-.617-2 2.5-.398 3.227-3 5 4.19 2.632 5 7.5 7.24-.978 5.5 6a8.4 8.4 0 0 1-.5-4h-5a66 66 0 0 1 3.5 10q1 2 2 0 4.653 5.199-2.5 7 .359 6.514 1 13-2.267-4.567-2-10h-3q.114-3.634-3-5.5 1.471-7.745-.5-15.5l-3.5-3.5q-1.405-1.038-2 .5h-2a162 162 0 0 1-12-17.5 13.6 13.6 0 0 0-3-6q1.523-2.06 0-3.5-2.256-3.504-6-5.5 2.098-.892 4 .5.096-1.298 1-3-3.044-4.337-8-6v-1q1 0 1-1 1.095-1.818 3-3-1.155-4.515-2.5-9a144 144 0 0 1-7.5-2q-1 0-1-1a31 31 0 0 0-2-2q2.103-2.014-1-2.5.423-3.417 3-7-.453-2.482-4-4l2-2q-2.39-8.84 2-17.5 1.267 12.881 1.5 28 1.967 4.108 6.5 4"/><path style="opacity:1" fill="#859a01" d="M173.5 189.5q3.147 2.153 7 3.5-4.18 3.031 1 4.5a29 29 0 0 1 1.5 4 11.3 11.3 0 0 1 1.5-3.5 31.3 31.3 0 0 0 9-3.5l1 1-2 2a41 41 0 0 1 4 6q-3.832 4.126-3-2h-4q.329-1.763-1-3-2.38 2.51-2 6-1.724.436-2 2.5 1.155 3.103 4 5l-2 1a12.5 12.5 0 0 0-7 0l1-1q-3.076-.624-4-3.5-.38-3.49 2-6-2.318-1.428-5-1 .471 7.114-4 1-4.893 5.7-6-1a9.1 9.1 0 0 1-5 1v-4a4.93 4.93 0 0 0 3-.5 12.9 12.9 0 0 0-5-.5q-.679-1.437-2-2.5a12.9 12.9 0 0 1 5-.5v-4q-2.682-.428-5 1 1.509 1.344 0 2-3.014-.847-5-3-.76-2.038-3-3l4.5-3q2.909 4.491 8.5 4.5 1.756 5.458 4 1-2.142-2.874-3-6.5a12 12 0 0 0-6-1 8.4 8.4 0 0 1 .5-4q.582 1.392 2 2 9.582-5.803 11.5 5a9.4 9.4 0 0 1-3 3q.635 4.184 0 8a18.5 18.5 0 0 1 4-5 3.65 3.65 0 0 1 1.5 1q.565-2.163 2.5-3"/><path style="opacity:1" fill="#8aa301" d="M344.5 194.5h-3v6q-4.796-.142-9 2a31 31 0 0 0-2 2q-1.83-1.817-3.5 0-4.011-3.556-9.5-3.5a26.7 26.7 0 0 0 6.5-1.5q-1.22-4.5.5-9-5.027.045-8 4-1.185 1.663-1.5 4-1.517-1.255-3.5-1 .609 2.414 0 4a25.8 25.8 0 0 0-7-5q2.748-.835 4-3a4.93 4.93 0 0 1 .5-3l1 1q3.468-.683 6-3 5.796 1.049 12.5 1 .38 3.49-2 6 1.237 1.329 3 1v4q5.174 1.115 5-4a6.85 6.85 0 0 1 4 1q1.846-1.726 0-3.5a12.9 12.9 0 0 1 5-.5v-5h3q1.763-5.525 0-11-.086-5.293-3-9 .166-.925 1-1.5-2.432-3.249-2.5-7.5a56 56 0 0 1-6.5-6q.618 1.923-2 3a20.7 20.7 0 0 0-4-3.5q1.926-1.026 3.5.5l1-3q3.543 1.278 5.5 4.5a14.7 14.7 0 0 1 5 1.5q-1.477 2.397-.5 5a25.9 25.9 0 0 0 6.5 5.5q-1.035 7.703-2 15.5 1.2 6.923-4 12"/><path style="opacity:1" fill="#7d9300" d="M211.5 190.5q5 .429 9.5 0a9.3 9.3 0 0 1 2.5 2 186 186 0 0 1 3-2q1.144 1.152 3 1.5-1.87 2.187 0 4.5a30 30 0 0 0-5 6.5 39.5 39.5 0 0 0 4 6.5l-1 1a54 54 0 0 0-9-4q-.768-4.276-5-4v3q1.763-.329 3 1a54.6 54.6 0 0 0 3.5 8q.399-5.226 4.5-3-2.44 3.11-4.5 7-2.928-3.747-7-2-.991-3.624-4.5-5 2.968-2.646 1-6-1.84-1.294-4 0-4.061-2.34-5-7h3v-8q2.55.19 4.5 2 1.616-1.394 3.5-2"/><path style="opacity:1" fill="#809501" d="M210.5 186.5q-.1 2.188 1 4-1.884.606-3.5 2-1.95-1.81-4.5-2v8h-3q.939 4.66 5 7v6q-.975.456-2 1-1.428-2.318-1-5-2.405-.282-4 1.5 1.526 1.574.5 3.5-2.874-2.142-6.5-3a13.3 13.3 0 0 1-3-4.5 4.93 4.93 0 0 0-3-.5q-.38-3.49 2-6 1.329 1.237 1 3h4q-.832 6.126 3 2a41 41 0 0 0-4-6l2-2-1-1a31.3 31.3 0 0 1-9 3.5 11.3 11.3 0 0 0-1.5 3.5 29 29 0 0 0-1.5-4q-5.18-1.469-1-4.5-3.853-1.347-7-3.5.657.098 1-.5-5.072-.273-3-4.5 4.442 1.024 1 2.5 3.721 1.14 5-2.5a14.2 14.2 0 0 1 4-3l2 1q-1.772 2.547 0 5a51 51 0 0 0 2-6 52 52 0 0 1 10-1.5q4.492-4.852 4 2.5h5l1 6h3q-.554-3.74 2-2"/><path style="opacity:1" fill="#8c9f02" d="M149.5 190.5q1.986 2.153 5 3 1.509-.656 0-2 2.318-1.428 5-1v4a12.9 12.9 0 0 0-5 .5q1.321 1.063 2 2.5-1.983.255-3.5-1-2.514 3.808.5 7.5a4.93 4.93 0 0 1-3 .5l3 9q3.384-5.256 9-8a21.74 21.74 0 0 0 7 6.5 8.4 8.4 0 0 1-4 .5v5q-7.57 2.451-2-2-1.68-3.144-4.5-1v-4q-5.134 4.374-8.5 10 2.718-.111 6 1 2.855 1.996 3.5 4 3.512-.76 1.5-3.5a11.3 11.3 0 0 0 3.5-1.5q2.708 3.933 0 8-1.21-5.923-1.5 0-1.763-.329-3 1a29 29 0 0 1 5 2.5q-4 1 0 2-1.13 2.23-2.5.5-3.686 3.888-.5 7.5a42.8 42.8 0 0 0-9 1.5q2.284 2.674 6 2 .063 3.523-3.5 4-2.167-1.863-4.5-2.5a52 52 0 0 1-4 6.5 6.8 6.8 0 0 1-1.5-3l-1 1q-.173 2.41 1.5 4.5-.92 4.074-3 7.5a211 211 0 0 1-12-18 300 300 0 0 1-1-13q7.153-1.801 2.5-7-1 2-2 0a66 66 0 0 0-3.5-10h5a8.4 8.4 0 0 0 .5 4q1.74-6.978-5.5-6-.81-4.868-5-7.5 2.602-1.773 3-5-2.164-3.117 2-2.5-.166.925-1 1.5l4 2q-5.8 2.98 0 1.5 1.857 3.206 5.5 4a6.98 6.98 0 0 0 2.5 3.5q-4.215 4.57 2 4a4.46 4.46 0 0 1-2-1.5q2.752-1.391 1-3.5 2.853-2.476 6-1-2.042-2.376-2-5.5 4.575 1.972 2-3 1.27-2.017 3.5-3 .745-3.465.5-7"/><path style="opacity:1" fill="#8ea501" d="M308.5 193.5q-1.252 2.165-4 3a25.8 25.8 0 0 1 7 5q.609-1.586 0-4 1.983-.255 3.5 1 .315-2.337 1.5-4-.356 4.068 0 7.5-2.452 2.622-6 3.5 1.924 4.538 1 9a7.25 7.25 0 0 1-2.5-2 69 69 0 0 1-4.5 6.5l2 1q-4.227 3.172.5 5.5 2.148-3.753.5-7.5a18.5 18.5 0 0 1 6-.5q-.359 2.951 1 5.5a4.93 4.93 0 0 1-3 .5q.393 2.217-1 4a4.46 4.46 0 0 1-1.5-2q-1.627 1.854-2.5 4a16.5 16.5 0 0 1-5-2.5q1.817-1.67 0-3.5 3.597-3.182.5-7-.574.834-1.5 1v-3a4.93 4.93 0 0 1-3-.5q1.701-1.524-1-2 6.476-5.535-1-5-6.43-9.326 4-4.5-1.79 4.46 2 1.5a20.4 20.4 0 0 1-1.5-5.5 4.46 4.46 0 0 1-1.5 2 64 64 0 0 0-4-4.5q1.129-1.006 1-2.5 1.884-.058 3 1.5-1.817 1.34 0 2.5a8.2 8.2 0 0 1 4-2.5q-1.392-.582-2-2 .166-2.583 1.5-.5.582-1.392 2-2 1.684 2.4 4.5 3"/><path style="opacity:1" fill="#8ca002" d="M268.5 191.5q-.822 3.876-5 3 2.161 7.188 7 15a29.8 29.8 0 0 0 8.5-3 9.3 9.3 0 0 1 2.5 2l1-6q2.674 2.284 2 6 4.232-.276 5 4a52.2 52.2 0 0 1 11 5q.926-.166 1.5-1 3.097 3.818-.5 7 1.817 1.83 0 3.5a16.5 16.5 0 0 0 5 2.5q-.467 1.467-2 2-1.794-.519 0-1.5a18.5 18.5 0 0 0-6-.5q-.268-3.159-3-5l-1 1q1.175 5.83 5 10-1.272 2.39-4 3.5-5.97.61-11 3.5a23.3 23.3 0 0 1-7-3q1.144-1.152 3-1.5a26.7 26.7 0 0 1-7-1.5q1.827-1.954-1-3 4.85-1.602 11-3l-3-6q1.237-1.329 3-1-.26 6.782 2 1-1.23-2.417-4-2v-4q-3.618-.384-5 3-3.144-1.68-1-4.5 7.041.507 14-.5-3.515-1.09-8-3 5.635-2.172 0-3a688 688 0 0 1-7 2q.22 3.663-2.5 2-1.35 1.905-3.5 3a9.3 9.3 0 0 1 2 2.5 43 43 0 0 0-3 5.5l3 3q1.137-1.966 3-1.5-1.752 2.109 1 3.5a43 43 0 0 0-7.5 1q.876-7.403-5.5-5-.318-2.826-3-4l3-2a29 29 0 0 1-5-2.5q1.712-1.112 0-2.5l3-4q-4.769-.306-10-2a4.46 4.46 0 0 1 2-1.5q-2.26-1.256-1.5-3.5a8.9 8.9 0 0 1 3 4l2.5-2.5q-1.922-2.113-3-4.5a34.2 34.2 0 0 0 1-10q-1.763.329-3-1 9.164-1.902 18 1"/><path style="opacity:1" fill="#879c01" d="M242.5 205.5q2.608-1.401 5 0a3.65 3.65 0 0 1-1 1.5 169 169 0 0 1 5 11.5q2.217-.393 4 1-3.337 5.298-5.5 0-2.199 2.737-5.5 1.5 1.426-1.077.5-2.5-1.782 1.595-1.5 4h4q.255 1.983-1 3.5 4.25 2.75-.5 4.5l-1-3q-2.254.502-4 0v-4q-5.687 4.466-12.5 2v3q-3.126 1.122-6 0a4.46 4.46 0 0 0-1.5 2l-2.5-5 3-3q1.536 2.03 4 2v-4q-5.403-.267 0-3-3.32-2.607-1-6-4.101-2.226-4.5 3a54.6 54.6 0 0 1-3.5-8q-1.237-1.329-3-1v-3q4.232-.276 5 4a54 54 0 0 1 9 4l1-1a39.5 39.5 0 0 1-4-6.5 30 30 0 0 1 5-6.5q-1.87-2.313 0-4.5-1.856-.348-3-1.5 2.55-.19 4.5-2 1.418.608 2 2l4-4a6.55 6.55 0 0 0 3 3q.096 4.894 2.5 9-3.372-.498-7 0a30.6 30.6 0 0 0-3 5q2.91 4.044-2 4v3q8.02-.553 4-7 2.15-1.095 3.5-3a20.5 20.5 0 0 1 3.5 2.5q-4.248 2.862 1 2.5"/><path style="opacity:1" fill="#95ab05" d="M332.5 202.5v3q-4.67 1.711-8.5-2 .028 2.41.5 5a25 25 0 0 0-7 .5 8.4 8.4 0 0 0 4 .5 9.1 9.1 0 0 0 1 5l-2 2q3.477 2.001 5-2 1.427 1.238 3 .5-1.794-.981 0-1.5a7.3 7.3 0 0 1 3 2 36.3 36.3 0 0 1-7 4 9.4 9.4 0 0 0 3 3q-7.275 2.307-12 8-7.194 4.803-16 5-3.825-4.17-5-10l1-1q2.732 1.841 3 5a18.5 18.5 0 0 1 6 .5q-1.794.981 0 1.5 1.533-.533 2-2 .873-2.146 2.5-4a4.46 4.46 0 0 0 1.5 2q1.393-1.783 1-4a4.93 4.93 0 0 0 3-.5q-1.359-2.549-1-5.5a18.5 18.5 0 0 0-6 .5q1.648 3.747-.5 7.5-4.727-2.328-.5-5.5l-2-1a69 69 0 0 0 4.5-6.5 7.25 7.25 0 0 0 2.5 2q.924-4.462-1-9 3.548-.878 6-3.5-.356-3.432 0-7.5 2.973-3.955 8-4-1.72 4.5-.5 9a26.7 26.7 0 0 1-6.5 1.5q5.489-.056 9.5 3.5 1.67-1.817 3.5 0a31 31 0 0 1 2-2"/><path style="opacity:1" fill="#809400" d="M186.5 204.5a4.93 4.93 0 0 1 3 .5 13.3 13.3 0 0 0 3 4.5q3.626.858 6.5 3 1.026-1.926-.5-3.5 1.595-1.782 4-1.5-.428 2.682 1 5 1.025-.544 2-1a21.5 21.5 0 0 1 0 5.5 25.7 25.7 0 0 1 7.5 2.5q1.088-2.487 3.5-1.5 2.256 2.006 3 5-2.332-.637-4.5-2.5-2.643.666.5 1.5-1.91 4.137-5 1a3.65 3.65 0 0 0 1-1.5q-3.046-2.967-5 0 1.237 1.329 3 1v3h-4v3q-2.435.21-3.5-2-2.328-.733-3.5 2 .013 2.712-1.5 5-1.056-3.34-3.5-1a87 87 0 0 1-5.5-3 8.4 8.4 0 0 1-.5-4q-4.061 1.423-8 0 3.014-.847 5-3a8.9 8.9 0 0 1-3-4q6.818-.17 14-1.5l-16-1q-2.15-4.063-6-6.5 1.237-1.329 3-1 .924 2.876 4 3.5l-1 1a12.5 12.5 0 0 1 7 0l2-1q-2.845-1.897-4-5 .276-2.064 2-2.5"/><path style="opacity:.2" fill="#9eb210" d="M344.5 194.5q1.563 1.937 3.5 0 1.67 1.89.5 3.5 2.062 1.893 4 1.5.375 4.507-4 5.5l1 1q-4.874 2.994-6.5 8.5-2.684 3.548-6.5 5 .003 3.496-2.5 6a3.65 3.65 0 0 0-1.5-1q-2.822 3.35-7 4 .393-2.217-1-4-.836 2.748-3 4-1 0-1 1-2.687-.201-5 1 4.725-5.693 12-8a9.4 9.4 0 0 1-3-3 36.3 36.3 0 0 0 7-4 7.3 7.3 0 0 0-3-2q-1.794.519 0 1.5-1.573.738-3-.5-1.523 4.001-5 2l2-2a9.1 9.1 0 0 1-1-5 8.4 8.4 0 0 1-4-.5 25 25 0 0 1 7-.5 30 30 0 0 1-.5-5q3.83 3.711 8.5 2v-3q4.204-2.142 9-2v-6z"/><path style="opacity:1" fill="#879b01" d="M165.5 217.5q.984-1.166 2.5-2 2.624 4.581 8 5 2.566.091 4-2a29 29 0 0 0 1.5 4q-2.052 1.64-3.5-1a12.5 12.5 0 0 0 0 7q1.604-.21 3.5 0-3.615 2.437-8 4-3.378 2.695-3 7a19.2 19.2 0 0 1-6 1q.077 3.105-2 5.5l1 1.5q1.663-1.635 3 0 3.106 3.645 1 8l-4-4q-2.403 1.575-2.5-2a4.46 4.46 0 0 0-1.5 2q2.484 2.868 1 6a35 35 0 0 1-4.5-1 31.4 31.4 0 0 0-1 13q1.517 1.255 3.5 1 .426 2.872-1 5-2.47-2.74-6-4-.5-1.5-2-2a19.6 19.6 0 0 1-6-6q2.08-3.426 3-7.5-1.673-2.09-1.5-4.5l1-1a6.8 6.8 0 0 0 1.5 3 52 52 0 0 0 4-6.5q2.333.637 4.5 2.5 3.563-.477 3.5-4-3.716.674-6-2a42.8 42.8 0 0 1 9-1.5q-3.186-3.612.5-7.5 1.37 1.73 2.5-.5-4-1 0-2a29 29 0 0 0-5-2.5q1.237-1.329 3-1 .29-5.923 1.5 0 2.708-4.067 0-8a11.3 11.3 0 0 1-3.5 1.5q2.012 2.74-1.5 3.5-.645-2.004-3.5-4-3.282-1.111-6-1 3.366-5.626 8.5-10v4q2.82-2.144 4.5 1-5.57 4.451 2 2"/><path style="opacity:1" fill="#869b01" d="M156.5 197.5a12.9 12.9 0 0 1 5 .5 4.93 4.93 0 0 1-3 .5v4a9.1 9.1 0 0 0 5-1q1.107 6.7 6 1 4.471 6.114 4-1 2.682-.428 5 1-2.38 2.51-2 6-1.763-.329-3 1 3.85 2.437 6 6.5l16 1q-7.182 1.33-14 1.5a8.9 8.9 0 0 0 3 4q-1.986 2.153-5 3 3.939 1.423 8 0a8.4 8.4 0 0 0 .5 4 87 87 0 0 0 5.5 3q1.917 1.873 3 4.5-2.13 2.295-3 5.5l-2-3a3.65 3.65 0 0 0-1.5 1q-2.853-3.576-3.5-8-8.753 1.258-2 7a125 125 0 0 0-9-3q-1.656 1.697 1 2.5-2.73 1.469-3.5 4.5-.575-.834-1.5-1a5.73 5.73 0 0 1-1-3q-.378-4.305 3-7 4.385-1.563 8-4-1.896-.21-3.5 0a12.5 12.5 0 0 1 0-7q1.448 2.64 3.5 1a29 29 0 0 1-1.5-4q-1.434 2.091-4 2-5.376-.419-8-5-1.516.834-2.5 2v-5a8.4 8.4 0 0 0 4-.5 21.74 21.74 0 0 1-7-6.5q-5.616 2.744-9 8l-3-9a4.93 4.93 0 0 0 3-.5q-3.014-3.692-.5-7.5 1.517 1.255 3.5 1"/><path style="opacity:1" fill="#879b02" d="M224.5 211.5q-2.32 3.393 1 6-5.403 2.733 0 3v4q-2.464.03-4-2l-3 3 2.5 5a4.46 4.46 0 0 1 1.5-2q2.874 1.122 6 0 1.853-.332 3-2a32 32 0 0 1 3 7q-2.682-.428-5 1-.885 1.948-3 2.5l2 1q-1.926 1.026-3.5-.5-1.124 1.702.5 3 1.426-1.238 3-.5-1.4 2.412 2 2.5 3.103-.486 1-2.5 1-1 2 0-.016 4.137 1 8.5a4.46 4.46 0 0 1 1.5 2 4.46 4.46 0 0 1 1.5-2q2.082 1.636 1.5 4-4.606-1.008-9.5-.5a6.8 6.8 0 0 0 3-1.5q-1.654-.747-3-1 1.817-1.83 0-3.5 2.19-1.217 1-2.5-2.45.702-3.5 3-1.05-2.298-3.5-3-1.83 1.817-3.5 0a12.9 12.9 0 0 0-.5 5 8.4 8.4 0 0 0-4 .5q1.255 1.517 1 3.5-2.322-.481-2-3-1.983-.255-3.5 1-1.35-1.905-3.5-3 1.346-.253 3-1a6.8 6.8 0 0 0-3-1.5q3.068-1.395 2.5-4.5-1.362 2.909-4.5 3 .41-2.926-2-4.5 3.684-3.238-2-2.5a12.9 12.9 0 0 0-.5-5q-.981 1.794-1.5 0 4.399-3.357-1-3 1.172-2.733 3.5-2 1.065 2.21 3.5 2v-3h4v-3q-1.763.329-3-1 1.954-2.967 5 0a3.65 3.65 0 0 1-1 1.5q3.09 3.137 5-1-3.143-.834-.5-1.5 2.168 1.863 4.5 2.5-.744-2.994-3-5-2.412-.987-3.5 1.5a25.7 25.7 0 0 0-7.5-2.5 21.5 21.5 0 0 0 0-5.5v-6q2.16-1.294 4 0 1.968 3.354-1 6 3.509 1.376 4.5 5 4.072-1.747 7 2 2.06-3.89 4.5-7"/><path style="opacity:1" fill="#859900" d="M252.5 201.5q1.078 2.387 3 4.5l-2.5 2.5a8.9 8.9 0 0 0-3-4q-.76 2.244 1.5 3.5a4.46 4.46 0 0 0-2 1.5q5.231 1.694 10 2l-3 4q1.712 1.388 0 2.5a29 29 0 0 0 5 2.5l-3 2q2.682 1.174 3 4-4.68 2.04-10 2.5 2.344 1.139 3 3.5h-5v5q-3.49-.38-6 2 2.254.828 5 0a8.4 8.4 0 0 0 .5 4q4.49-1.473 9.5-1 3.916.221 7 2.5a290 290 0 0 0-11 3.5q-2.241-.314-3.5-2l-3 5a8.4 8.4 0 0 1-.5-4q-6.785-.71-7 6-5.4-.037-11 1a31 31 0 0 0-6-2.5q1.426-1.077.5-2.5a13.9 13.9 0 0 0-4 3l-1-2q-.625 1.884-2.5 2 .255-1.983-1-3.5a8.4 8.4 0 0 1 4-.5 12.9 12.9 0 0 1 .5-5q1.67 1.817 3.5 0 2.45.702 3.5 3 1.05-2.298 3.5-3 1.19 1.283-1 2.5 1.817 1.67 0 3.5 1.346.253 3 1a6.8 6.8 0 0 1-3 1.5q4.894-.508 9.5.5.582-2.364-1.5-4a4.46 4.46 0 0 0-1.5 2 4.46 4.46 0 0 0-1.5-2q-1.016-4.363-1-8.5-1-1-2 0 2.103 2.014-1 2.5-3.4-.088-2-2.5-1.574-.738-3 .5-1.624-1.298-.5-3 1.574 1.526 3.5.5l-2-1q2.115-.552 3-2.5 2.318-1.428 5-1a32 32 0 0 0-3-7q-1.147 1.668-3 2v-3q6.813 2.466 12.5-2v4q1.746.502 4 0l1 3q4.75-1.75.5-4.5 1.255-1.517 1-3.5h-4q-.282-2.405 1.5-4 .926 1.423-.5 2.5 3.301 1.237 5.5-1.5 2.163 5.298 5.5 0-1.783-1.393-4-1a169 169 0 0 0-5-11.5 3.65 3.65 0 0 0 1-1.5q-2.392-1.401-5 0 .365-4.144 1.5-9l3 7a4.93 4.93 0 0 0 .5-3q2.688-.201 5 1"/><path style="opacity:1" fill="#869a00" d="M277.5 239.5a4.93 4.93 0 0 1-.5 3q-4.473-2.237-7.5-6-3.312 1.158 1 2.5l-2.5 2.5q-3.277-3.747-7 0-5.05-1.228-9.5 0-.59-3.045-1-7a9.86 9.86 0 0 0 6-1q-.686-1.197-2-1-.656-2.361-3-3.5 5.32-.46 10-2.5 6.376-2.403 5.5 5a43 43 0 0 1 7.5-1q-2.752-1.391-1-3.5-1.863-.466-3 1.5l-3-3a43 43 0 0 1 3-5.5 9.3 9.3 0 0 0-2-2.5q2.15-1.095 3.5-3 2.72 1.663 2.5-2a688 688 0 0 0 7-2q5.635.828 0 3 4.485 1.91 8 3a65 65 0 0 1-14 .5q-2.144 2.82 1 4.5 1.382-3.384 5-3v4q2.77-.417 4 2-2.26 5.782-2-1-1.763-.329-3 1l3 6q-6.15 1.398-11 3 2.827 1.046 1 3 3.4 1.221 7 1.5-1.856.348-3 1.5"/><path style="opacity:1" fill="#849900" d="M198.5 228.5q5.399-.357 1 3 .519 1.794 1.5 0a12.9 12.9 0 0 1 .5 5q5.684-.738 2 2.5 2.41 1.574 2 4.5 3.138-.091 4.5-3 .568 3.105-2.5 4.5a6.8 6.8 0 0 1 3 1.5q-1.654.747-3 1 2.15 1.095 3.5 3 1.517-1.255 3.5-1-.322 2.519 2 3 .116 1.875 2 2.5-2.654 1.529-5 .5-1.5.5-2 2l4 3q-.544 1.025-1 2a8.4 8.4 0 0 1-4-.5l1-1-3-2.5a37.8 37.8 0 0 0-3-9 7.5 7.5 0 0 0-.5 5l-1-1a20.4 20.4 0 0 1-5.5 1.5q9.862.798 0 2.5a23.7 23.7 0 0 0-6-2.5q-1.827-3.668 1.5-6.5 2.284 3.18 5.5 1l2.5 3a29 29 0 0 1 1.5-4 29.2 29.2 0 0 0-8-.5q1.87-2.241 0-4.5-3.594 2.35-8 2.5 4.035.502 2 3.5a8.4 8.4 0 0 1-4-.5q1.73-1.37-.5-2.5a25 25 0 0 1-.5-7q-6.753-5.742 2-7 .647 4.424 3.5 8a3.65 3.65 0 0 1 1.5-1l2 3q.87-3.205 3-5.5-1.083-2.627-3-4.5 2.444-2.34 3.5 1 1.513-2.288 1.5-5"/><path style="opacity:1" fill="#b3b346" d="M115.5 212.5q3.409 3.462 1 8-11.193 9.13-18 20a52 52 0 0 0-7 1q-7.197 1.866-14 0 2.847-.884 6-.5a29.6 29.6 0 0 1-4.5-2.5q-.925-1.423.5-2.5a87.6 87.6 0 0 0 22-8q7.908-7.147 14-15.5"/><path style="opacity:1" fill="#afb129" d="M117.5 211.5q.595-1.538 2-.5l3.5 3.5q1.971 7.755.5 15.5 3.114 1.866 3 5.5h3q-.267 5.433 2 10a211 211 0 0 0 12 18 19.6 19.6 0 0 0 6 6 21.6 21.6 0 0 0-5 6q-2.21.857-5 0 2.124-3.54-1-6l-3 2q-1.147-1.668-3-2a15 15 0 0 1 3-6.5q-3.144-1.105-5-3.5.42-7.22-2-14a51 51 0 0 1-4 5.5q-3.129.94-1-1.5a4.46 4.46 0 0 0-2-1.5q3.09-2.24 1-5.5a3.65 3.65 0 0 0-1.5 1 8.4 8.4 0 0 1-.5-4h-3a25.9 25.9 0 0 1 5-9.5 80 80 0 0 1 0-8q-3.026-5.187-5-10.5"/><path style="opacity:1" fill="#91a303" d="M254.5 232.5q1.314-.197 2 1a9.86 9.86 0 0 1-6 1q.41 3.955 1 7 4.45-1.228 9.5 0 3.723-3.747 7 0l2.5-2.5q-4.312-1.342-1-2.5 3.027 3.763 7.5 6a4.93 4.93 0 0 0 .5-3 23.3 23.3 0 0 0 7 3 19.2 19.2 0 0 0-1 6 74 74 0 0 1-9 3.5 41.3 41.3 0 0 0-9 6.5q-3.882-.065-7 3.5l-7.5 1.5q-1.989-.383-3-2a9.5 9.5 0 0 1-1.5 2.5q2.949 2.02 1 4.5-.436-1.724-2.5-2a36.3 36.3 0 0 1-13.5.5 3.65 3.65 0 0 1-1-1.5 227 227 0 0 0 7-6.5q1.423-.926 2.5.5.743-2.96.5-6 .215-6.71 7-6a8.4 8.4 0 0 0 .5 4l3-5q1.259 1.686 3.5 2a290 290 0 0 1 11-3.5q-3.084-2.279-7-2.5-5.01-.473-9.5 1a8.4 8.4 0 0 1-.5-4q-2.746.828-5 0 2.51-2.38 6-2v-5z"/><path style="opacity:.014" fill="#b7b95a" d="M317.5 234.5q.5 2 1 0 2.329 8.907 3.5 19-4.94-8.355-9.5-18-.024-2.976 3-3 1.82.37 2 2"/><path style="opacity:1" fill="#b5b431" d="M115.5 212.5v-1h2q1.974 5.313 5 10.5a80 80 0 0 0 0 8 25.9 25.9 0 0 0-5 9.5h3a8.4 8.4 0 0 0 .5 4 3.65 3.65 0 0 1 1.5-1q2.09 3.26-1 5.5a4.46 4.46 0 0 1 2 1.5q-2.129 2.44 1 1.5a51 51 0 0 0 4-5.5 35.1 35.1 0 0 1 2 14q-1.451-1.96-4-2 .362 5.248-2.5 1-3.062 1.384-6.5 1-.5 8-1 0-5.667-1.04 0-3.5-1.91-12.516-5-.5a2.43 2.43 0 0 0-.5-2 7.3 7.3 0 0 1-2 3 225 225 0 0 1 2-8q2.335 1.058 3.5-2a20.2 20.2 0 0 0-3-4.5q.798-3.03 4-2.5v-4a37.7 37.7 0 0 1-8-2.5 9.5 9.5 0 0 0 1.5-2.5l1 1q1.24-2.858 3.5-1a4.46 4.46 0 0 1-2 1.5q2.96.743 6 .5.269-3.681-2-6.5a13 13 0 0 0 1-5.5q2.409-4.538-1-8"/><path style="opacity:.023" fill="#deddd6" d="M53.5 242.5a4.93 4.93 0 0 1 3 .5 23.6 23.6 0 0 1-1 3.5q-1.808-1.615-2-4"/><path style="opacity:1" fill="#8da002" d="M184.5 239.5q-.245 3.535.5 7 2.23 1.13.5 2.5a8.4 8.4 0 0 0 4 .5q-.846 4.124 2 8-2.928 3.705-1 8-2.024 2.175-5 2a12.9 12.9 0 0 0-.5-5q-1.512-.723-3.5-.5-.711-2.593-3-3.5 1.937-1.46 0-3-1.953-2.552-5-3.5a15.3 15.3 0 0 1-1.5-4.5q-3.206-1.468-5.5-4l1-1 3 2q1.197-.686 1-2 .925.166 1.5 1 .77-3.031 3.5-4.5-2.656-.803-1-2.5 4.61 1.347 9 3"/><path style="opacity:1" fill="#c5c341" d="M303.5 238.5q-.538 2.807-3 4.5 5.966 3.65 8 10.5 1.237 1.329 3 1a8.4 8.4 0 0 0 .5 4 116.4 116.4 0 0 1 13.5 14q-2.621 2.909-5 6.5-3.465.745-7 .5-.179-5.77 2-11-4.185.417-7.5-2-5.621 3.957-12.5 4v-8h4a12.9 12.9 0 0 0 .5 5q5.98-4.212.5-8.5 1.858-6.997 2.5.5 7.86-7.638-3.5-8a41.5 41.5 0 0 0-1-8q1.84-3.121 4.5-6 .598.343.5 1"/><path style="opacity:.992" fill="#859609" d="M98.5 240.5a502 502 0 0 1-18.5 24 862 862 0 0 0-23.5 26q-1.238 1.329-3 1a8.44 8.44 0 0 0 .5 4q5.266-3.264 11.5-4.5 6.05-6.639 15-7.5v1l-8 4q-10.382 5.973-21 9a20.7 20.7 0 0 0-5-1v-3q1.144-3.1 4-5-1.808-2.242 0-4 .09.814 1 1 1.329-1.237 1-3 4.178-.65 7-4a6.8 6.8 0 0 1 1.5 3l3-6 1 2q5.19-2.438 6.5-8a14.1 14.1 0 0 1 5.5-4 42.5 42.5 0 0 1 4.5-8l-2-2q4.483-2.418 7-7-.534-1.533-2-2a60 60 0 0 0 7-5 52 52 0 0 1 7-1"/><path style="opacity:1" fill="#b9b838" d="M116.5 220.5a13 13 0 0 1-1 5.5q2.269 2.819 2 6.5a18.5 18.5 0 0 1-6-.5 4.46 4.46 0 0 0 2-1.5q-2.26-1.858-3.5 1l-1-1a9.5 9.5 0 0 1-1.5 2.5 37.7 37.7 0 0 0 8 2.5v4q-3.202-.53-4 2.5a20.2 20.2 0 0 1 3 4.5q-1.165 3.058-3.5 2a225 225 0 0 0-2 8 7.3 7.3 0 0 0 2-3q.696.891.5 2-.653 2.696 0 6-5.187-.129-6 5a275 275 0 0 0-25 17q-8.95.861-15 7.5-6.234 1.236-11.5 4.5a8.44 8.44 0 0 1-.5-4q1.762.329 3-1a862 862 0 0 1 23.5-26 502 502 0 0 0 18.5-24q6.807-10.87 18-20"/><path style="opacity:1" fill="#c1bf3a" d="M298.5 243.5a41.5 41.5 0 0 1 1 8q11.36.362 3.5 8-.642-7.497-2.5-.5 5.48 4.288-.5 8.5a12.9 12.9 0 0 1-.5-5h-4v8q6.879-.043 12.5-4 3.315 2.417 7.5 2-2.179 5.23-2 11 3.535.245 7-.5 2.379-3.591 5-6.5 3.222-.166 1 2 1.635 2.254 3.5 1a50.3 50.3 0 0 1 6 9q1.426.576 2.5 1a514 514 0 0 0 3 4l-3 3q-1.997-.999-4.5 0a168.5 168.5 0 0 0-25.5-9q-11.446-2.56-19.5-11 .028 2.41.5 5v4a38 38 0 0 0-6-9.5q1.686-1.259 2-3.5a25.3 25.3 0 0 0-2.5-6q-1.604.21-3.5 0l4.5-7 1 1q.868-2.143-.5-4 1.37-2.379 0-4.5l2-2a232 232 0 0 1 12-2.5"/><path style="opacity:.008" fill="#a5a936" d="M314.5 245.5q2.755.209 1 2-1.197-.686-1-2"/><path style="opacity:.004" fill="#9cac14" d="M66.5 251.5q-1-2 0 0"/><path style="opacity:1" fill="#b2b929" d="M325.5 228.5q-1.78 1.55-2.5 4-3.751-2.334-4.5 2h-1q-.18-1.63-2-2-3.024.024-3 3-2.143 3.285-6 4a5.73 5.73 0 0 1-3-1q.098-.657-.5-1-2.66 2.879-4.5 6a232 232 0 0 0-12 2.5l-2 2q1.37 2.121 0 4.5a27.7 27.7 0 0 0-8 6 94 94 0 0 0-9 3 255 255 0 0 0-25 11q-4.637 4.818-12 4-.045-1.496 1.5-2a34 34 0 0 0 7.5-3 592 592 0 0 0 8-3q1.949-2.48-1-4.5a9.5 9.5 0 0 0 1.5-2.5q1.011 1.617 3 2l7.5-1.5q3.118-3.565 7-3.5a41.3 41.3 0 0 1 9-6.5 74 74 0 0 0 9-3.5 19.2 19.2 0 0 1 1-6q5.03-2.89 11-3.5 2.728-1.11 4-3.5 8.806-.197 16-5 2.313-1.201 5-1 1 0 1-1 2.164-1.252 3-4 1.393 1.783 1 4"/><path style="opacity:1" fill="#8ea103" d="M170.5 239.5a5.73 5.73 0 0 0 1 3q.197 1.314-1 2l-3-2-1 1q2.294 2.532 5.5 4a15.3 15.3 0 0 0 1.5 4.5q3.047.948 5 3.5 1.937 1.54 0 3a4.93 4.93 0 0 1-.5 3q-1.5-2-3 0l-1 6q-3.585-1.836-7.5-1.5 2.995 1.73 1 4.5-1.29-3.646-5-4a6.85 6.85 0 0 1-1 4 26.6 26.6 0 0 0 3.5 2 14.4 14.4 0 0 0 2.5 5.5q-2.86.197-4.5-2.5-1.545.504-1.5 2-2.243-.733-4-2 1.426-2.128 1-5-1.983.255-3.5-1a31.4 31.4 0 0 1 1-13q2.3.666 4.5 1 1.484-3.132-1-6a4.46 4.46 0 0 1 1.5-2q.097 3.575 2.5 2l4 4q2.106-4.355-1-8-1.337-1.635-3 0l-1-1.5q2.077-2.395 2-5.5a19.2 19.2 0 0 0 6-1"/><path style="opacity:1" fill="#8d9f02" d="M240.5 253.5a18.5 18.5 0 0 1-.5 6q-1.077-1.426-2.5-.5a227 227 0 0 1-7 6.5 3.65 3.65 0 0 0 1 1.5 36.3 36.3 0 0 0 13.5-.5q2.064.276 2.5 2a592 592 0 0 1-8 3q-.974-1.659-1-3-4.925 3.36-11 3.5-1.083 4.111-5 3-2.26-1.256-1.5-3.5a55 55 0 0 0 4.5-2 391 391 0 0 0-7-.5 23.7 23.7 0 0 0 6-2.5 39.5 39.5 0 0 1 5-5.5l-5.5-3.5q-3.74 2.21 1.5 4.5-4.89 3.258-11 1.5v-1q.456-.975 1-2l-4-3q.5-1.5 2-2 2.346 1.029 5-.5-1.884-.625-2-2.5 1.875-.116 2.5-2l1 2a13.9 13.9 0 0 1 4-3q.926 1.423-.5 2.5a31 31 0 0 1 6 2.5q5.6-1.037 11-1"/><path style="opacity:1" fill="#889d03" d="M84.5 246.5q1.466.467 2 2-2.517 4.582-7 7l2 2a42.5 42.5 0 0 0-4.5 8 14.1 14.1 0 0 0-5.5 4q-1.31 5.562-6.5 8l-1-2-3 6a6.8 6.8 0 0 0-1.5-3q-2.822 3.35-7 4 .329 1.763-1 3-.91-.186-1-1 1.653-2.388-2-2.5 3.575-.097 2-2.5 2.53-3.915 4 1a32 32 0 0 1 2.5-4 36.3 36.3 0 0 1 .5-7l2 2a37.4 37.4 0 0 0 6-7q-1.718-2.086 0-4a4.45 4.45 0 0 1 1.5 2q-.399-5.101 4.5-5 .715-3.74 4-6l-2-3 1-1a28.8 28.8 0 0 1 10-1"/><path style="opacity:.004" fill="#b5b547" d="M322.5 254.5q2 1 0 0"/><path style="opacity:.004" fill="#c0be4e" d="M322.5 256.5q2 1 0 0"/><path style="opacity:.004" fill="#bab942" d="M329.5 256.5q2 1 0 0"/><path style="opacity:.004" fill="#bebe45" d="M329.5 257.5q-1 1 0 0"/><path style="opacity:1" fill="#b6b631" d="M130.5 259.5q1.856 2.395 5 3.5a15 15 0 0 0-3 6.5 9.86 9.86 0 0 1-1 6q-1.509-.656 0-2-4.532-6.16-3.5 1-1.757-2.09-4.5-2.5l3-1q-2.969-1.026-5 1.5a35 35 0 0 1-4.5-1q-1.026 1.926.5 3.5-2.41 1.573-2 4.5h-3q-.333-4.335 2-8-4.113-6.18-2 1a38.3 38.3 0 0 1-6 10q-1.238 1.329-3 1a240 240 0 0 0-3 10.5 27 27 0 0 1 1 3.5 43 43 0 0 0-3 9 6.8 6.8 0 0 1-1.5-3q-1.155 1.647-2.5 0-1.8-.219-2 3 2.237 3.062-1 4.5.685.621 1 1.5a35.6 35.6 0 0 1-4.5 4 160 160 0 0 1-2.5 10q-1.025.544-2 1a208.8 208.8 0 0 1 19.5-47l1.5-6a49.5 49.5 0 0 0 4-7q-1.238-1.329-3-1 .813-5.129 6-5-.653-3.304 0-6 3.09-12.016 5 .5-5.667 2.46 0 3.5.5 8 1 0 3.438.384 6.5-1 2.862 4.248 2.5-1 2.549.04 4 2"/><path style="opacity:1" fill="#889d01" d="M214.5 262.5v1q-6.032-2.115-11.5 1-1.585-1.69-3.5-1.5l1 1q-3.578.426-2 3-3.628 3.465 2 2 1.635-3.824 5-2.5l-4 4q1.863 5.145-4 4 .117 4.128.5 8-3.404-1.642-6.5 1a66 66 0 0 0-10-1.5q-3.522-.52-5.5-3.5a7.3 7.3 0 0 0-2 3q-2.631-2.625-6.5-2-3.597.361-6-2-.045-1.496 1.5-2 1.64 2.697 4.5 2.5a14.4 14.4 0 0 1-2.5-5.5 26.6 26.6 0 0 1-3.5-2 6.85 6.85 0 0 0 1-4q3.71.354 5 4 1.995-2.77-1-4.5 3.915-.336 7.5 1.5l1-6q1.5-2 3 0a4.93 4.93 0 0 0 .5-3q2.289.907 3 3.5 1.988-.223 3.5.5a12.9 12.9 0 0 1 .5 5q2.976.175 5-2-1.928-4.295 1-8-2.846-3.876-2-8 2.035-2.998-2-3.5 4.406-.15 8-2.5 1.87 2.259 0 4.5a29.2 29.2 0 0 1 8 .5 29 29 0 0 0-1.5 4l-2.5-3q-3.216 2.18-5.5-1-3.327 2.832-1.5 6.5a23.7 23.7 0 0 1 6 2.5q9.862-1.702 0-2.5a20.4 20.4 0 0 0 5.5-1.5l1 1a7.5 7.5 0 0 1 .5-5 37.8 37.8 0 0 1 3 9l3 2.5-1 1a8.4 8.4 0 0 0 4 .5"/><path style="opacity:.037" fill="#c3c04c" d="M303.5 238.5q1.337.915 3 1 5.124 3.74 6.5 10a359 359 0 0 1 10.5 12 29 29 0 0 1 1.5 4q6.04 3.19 11.5 9a3.65 3.65 0 0 0 1.5-1q7.098 4.144 12.5 10 .063 3.523-3.5 4-3.892-2.47-8.5-2-1.074-.424-2.5-1a50.3 50.3 0 0 0-6-9q-1.865 1.254-3.5-1 2.222-2.166-1-2a116.4 116.4 0 0 0-13.5-14 8.4 8.4 0 0 1-.5-4q-1.763.329-3-1-2.034-6.85-8-10.5 2.462-1.693 3-4.5"/><path style="opacity:1" fill="#859801" d="M239.5 271.5a34 34 0 0 1-7.5 3q-1.545.504-1.5 2-14.955 2.225-28 9a311 311 0 0 1-28-1.5q-4.03-1.573-7-4.5 3.869-.625 6.5 2a7.3 7.3 0 0 1 2-3q1.978 2.98 5.5 3.5a66 66 0 0 1 10 1.5q3.096-2.642 6.5-1a114 114 0 0 1-.5-8q5.863 1.145 4-4l4-4q-3.365-1.324-5 2.5-5.628 1.465-2-2-1.578-2.574 2-3l-1-1q1.915-.19 3.5 1.5 5.468-3.115 11.5-1 6.11 1.758 11-1.5-5.24-2.29-1.5-4.5l5.5 3.5a39.5 39.5 0 0 0-5 5.5 23.7 23.7 0 0 1-6 2.5q3.742.236 7 .5a55 55 0 0 1-4.5 2q-.76 2.244 1.5 3.5 3.917 1.111 5-3 6.075-.14 11-3.5.026 1.341 1 3"/><path style="opacity:1" fill="#c7c43c" d="M284.5 252.5q1.368 1.857.5 4l-1-1-4.5 7q1.896.21 3.5 0a25.3 25.3 0 0 1 2.5 6q-.314 2.241-2 3.5a38 38 0 0 1 6 9.5 2.43 2.43 0 0 1-.5 2q-1.574-1.526-3.5-.5a16.7 16.7 0 0 0-4 4 10.94 10.94 0 0 1 3 4.5q-1.684 2.973 0 6a2.43 2.43 0 0 1-.5 2 16.1 16.1 0 0 1-2.5-5q-1.16-1.817-2.5 0l-3-3q-1.023 3.657 1.5 6.5-.275 2.526-2 .5-1.025-2.514-4-2 .13-4.648-1.5-1-.244-3.264-2.5-5.5 1.32-2.034 1-4.5 3.663.22 2-2.5 3.75-1.25.5-3.5-1.064 2.21-3.5 2-3.231 3.798-4.5 12-.177-4.307-2-8a43.4 43.4 0 0 1-2.5 10q-1.152-1.144-1.5-3l-1 1q-1.083-4.141.5-8 1.096-.663 3-.5a9.5 9.5 0 0 1-1.5-2.5q-1.524 1.701-2-1a4.93 4.93 0 0 0-.5 3 12.9 12.9 0 0 1-5-.5 20.5 20.5 0 0 1 3.5-6.5l1 2q.746-3.97.5-8 4.276-3.38 10-4.5 2.042-2.376 2-5.5a94 94 0 0 1 9-3 27.7 27.7 0 0 1 8-6"/><path style="opacity:1" fill="#c2c133" d="M267.5 261.5q.042 3.124-2 5.5-5.724 1.12-10 4.5a32.5 32.5 0 0 1-.5 8l-1-2a20.5 20.5 0 0 0-3.5 6.5 12.9 12.9 0 0 0 5 .5 4.93 4.93 0 0 1 .5-3q.476 2.701 2 1 .561 1.375 1.5 2.5-1.904-.163-3 .5h-6a25 25 0 0 1-.5 7l-1-8a32 32 0 0 0-2.5 4q-1.388-1.712-2.5 0a17.8 17.8 0 0 0-3.5-5l-1 1q6.563 5.842-1.5 9l-1-1a18.5 18.5 0 0 0-5 7 42 42 0 0 0-7.5-3.5q5.094-.595 6.5-5.5l1 3q1.405-4.278-2-7l-1 2v-6q-.417 1.374-1.5 2 .729-3.46 0-6.5 1.856-.348 3-1.5 7.363.818 12-4a255 255 0 0 1 25-11"/><path style="opacity:.004" fill="#b5b935" d="M374.5 276.5q2 1 0 0"/><path style="opacity:1" fill="#bdba33" d="M149.5 269.5q1.5.5 2 2 1.94 6.211 5 13 2.054-2.436 5-1.5 4.036 3.466 3.5 8.5a14.5 14.5 0 0 1-4-5 10.8 10.8 0 0 1-2.5 3q1.497 3.792 2.5 0 1.904 2.107 1.5 5-.663 1.096-.5 3a84 84 0 0 1-6.5-3l-2 2q.21-2.634-1.5-5a14.7 14.7 0 0 1-1.5 5l-2-1a34.3 34.3 0 0 1-1.5-8 8.4 8.4 0 0 0-.5 4q-2.464.03-4-2a28 28 0 0 1-5 1.5q2.157.328 4 1.5a23.8 23.8 0 0 0-1 9h-4a124 124 0 0 1 3-3.5q-.666-2.643-1.5.5-.964-3.645-4.5-5 3.063-1.505 1-4 2.243.033 5-1l-2-3a15.86 15.86 0 0 1 6-5.5q-1.817-1.34 0-2.5a4.46 4.46 0 0 1 1.5 2q1.147-2.349-.5-4a21.6 21.6 0 0 1 5-6"/><path style="opacity:.477" fill="#94a70b" d="M77.5 241.5q6.803 1.866 14 0a60 60 0 0 1-7 5 28.8 28.8 0 0 0-10 1l-1 1 2 3q-3.285 2.26-4 6-4.899-.101-4.5 5a4.45 4.45 0 0 0-1.5-2q-1.718 1.914 0 4a37.4 37.4 0 0 1-6 7l-2-2a36.3 36.3 0 0 0-.5 7 32 32 0 0 0-2.5 4q-1.47-4.915-4-1 1.575 2.403-2 2.5 3.653.112 2 2.5-1.808 1.758 0 4-2.856 1.9-4 5-1.93 1.902-3-1a75 75 0 0 0-3 11l-2-1a10.3 10.3 0 0 0-1.5 4 31.5 31.5 0 0 1 .5-11.5q-.162-2.522-2-3.5v-1q1.203-2.178 1.5-5 1-2 2 0 1.85-4.851 6-8 .414-3.4 2-7 4.703-1.851 4.5-6a25.5 25.5 0 0 0 4.5-4l1 2q.75-2.75 3.5-3.5-3.103-.486-1-2.5a38.3 38.3 0 0 0 7-5 84 84 0 0 1 7.5-8l1 2q.183-2.905 2.5-4"/><path style="opacity:.004" fill="#b6ba36" d="M374.5 279.5q2 1 0 0"/><path style="opacity:1" fill="#697602" d="M105.5 266.5q1.762-.329 3 1a49.5 49.5 0 0 1-4 7q-5.507-.476-5 5-4.305-.378-7 3-3.677 4.35-4 10-4.597.389-7.5-3-.417 1.374-1.5 2-.38-3.49 2-6-.186-.91-1-1v-1a275 275 0 0 1 25-17"/><path style="opacity:1" fill="#636b01" d="M308.5 283.5q1.874 2.454 5 3.5a570 570 0 0 0 13 4.5v2a251 251 0 0 1-15-4 177 177 0 0 0-22-12 30 30 0 0 1-.5-5q8.054 8.44 19.5 11"/><path style="opacity:1" fill="#bcb935" d="M132.5 269.5q1.853.332 3 2l3-2q3.124 2.46 1 6-1.143 2.213-1 4l-2-4q-1.509.656 0 2a20 20 0 0 0-3 4.5q.608 1.418 2 2a18.8 18.8 0 0 0-5.5 4.5q1.779-4.728 1-10l-1 5-2-8a11.3 11.3 0 0 1-1.5 3.5q-5.775-1.716-2.5 3.5l3-2q-.051 4.014-1.5 8 8.487 4.006-1 4 .342-2.782-2-2.5-2.274 2.846-2 6.5-10.546.029-1 1.5-6.084.527-11 4a53 53 0 0 0 7 1.5v3q-4.51 2.553-8.5 2-1.174 2.9 1.5 4.5-1.128 1.006-1 2.5-5.114 2.215-9-2l-2 2a75 75 0 0 1-2-12q1.345 1.647 2.5 0a6.8 6.8 0 0 0 1.5 3 43 43 0 0 1 3-9 27 27 0 0 0-1-3.5 240 240 0 0 1 3-10.5q1.762.329 3-1a38.3 38.3 0 0 0 6-10q-2.113-7.18 2-1-2.333 3.665-2 8h3q-.41-2.927 2-4.5-1.526-1.574-.5-3.5 2.3.666 4.5 1 2.031-2.526 5-1.5l-3 1q2.743.41 4.5 2.5-1.032-7.16 3.5-1-1.509 1.344 0 2a9.86 9.86 0 0 0 1-6"/><path style="opacity:.004" fill="#93a618" d="M35.5 290.5q-1-2 0 0"/><path style="opacity:1" fill="#b8b936" d="M308.5 283.5a168.5 168.5 0 0 1 25.5 9q2.503-.999 4.5 0l3-3a514 514 0 0 1-3-4q4.608-.47 8.5 2 3.563-.477 3.5-4 2.133.313 4 1.5l11.5 11.5 1-2q2.018 3.495.5 7-4.695.627-9 1-3.913.15-7-2a317 317 0 0 1-25-9 570 570 0 0 1-13-4.5q-3.126-1.046-5-3.5"/><path style="opacity:1" fill="#bcba2f" d="M151.5 271.5q3.53 1.26 6 4 1.757 1.267 4 2 2.403 2.361 6 2 2.97 2.927 7 4.5a311 311 0 0 0 28 1.5 2.43 2.43 0 0 1-.5 2q-3.732-1.368-5 3a32.4 32.4 0 0 1-1.5 6 4.46 4.46 0 0 0 2 1.5q-8.097 3.422-16 3.5v10q-1.203 2.178-1.5 5l-1-1a12.9 12.9 0 0 0-.5 5q1.763-.329 3 1l-3 3q-.83-5.934-2.5 0-4.138-1.731-8 1l-1.5-1q1.668-1.147 2-3 1.763.329 3-1a49.4 49.4 0 0 1-1.5-12q1.595 1.407 3.5 3a21.6 21.6 0 0 0 4-5q-.885-1.948-3-2.5a445 445 0 0 1-11-2.5 10.3 10.3 0 0 1 4-1.5 95 95 0 0 1-5-5.5q.404-2.893-1.5-5-1.003 3.792-2.5 0a10.8 10.8 0 0 0 2.5-3 14.5 14.5 0 0 0 4 5q.536-5.034-3.5-8.5-2.946-.936-5 1.5-3.06-6.789-5-13"/><path style="opacity:1" fill="#c2bf33" d="M256.5 285.5q-1.583 3.859-.5 8l1-1q.348 1.856 1.5 3 .426 2.872-1 5-3.027 1.684-6 0-2.075 3.864-2.5-1-2.003 3.157-5 2a150 150 0 0 0-10.5 5 10.8 10.8 0 0 0-2.5 3 20 20 0 0 1-3.5-4q-1.046 2.827-3 1-1.393 1.783-1 4a4.93 4.93 0 0 0-3 .5q2.995 1.73 1 4.5-1.948-.885-2.5-3-1.774 1.846-3.5 0 5.721-2.654-.5-5-1.258 4.761-5 8-1.647-2.201-4.5-2.5a4.93 4.93 0 0 1 3-.5v-5q-1.725-4.905 3-8-3.737-1.561 0-2.5-2.21-1.064-2-3.5-2.835.305-4.5-2l-1 2a8.4 8.4 0 0 1-.5-4q8.641-.313 1-2.5a25.7 25.7 0 0 0 7.5-2.5q2.558 1.895 5 1 .81 2.452-1.5 4 1.783 1.393 4 1 .144-5.716 2 0 1.329-1.237 1-3h5v-3q1.083-.626 1.5-2v6l1-2q3.405 2.722 2 7l-1-3q-1.406 4.905-6.5 5.5a42 42 0 0 1 7.5 3.5 18.5 18.5 0 0 1 5-7l1 1q8.063-3.158 1.5-9l1-1a17.8 17.8 0 0 1 3.5 5q1.112-1.712 2.5 0a32 32 0 0 1 2.5-4l1 8q.745-3.465.5-7z"/><path style="opacity:1" fill="#b7b52f" d="M139.5 275.5q2.79.857 5 0 1.647 1.651.5 4a4.46 4.46 0 0 0-1.5-2q-1.817 1.16 0 2.5a15.86 15.86 0 0 0-6 5.5l2 3q-2.757 1.033-5 1 2.063 2.495-1 4 3.536 1.355 4.5 5 .834-3.143 1.5-.5a124 124 0 0 0-3 3.5v2h-1q-.353-.895-1.5-1a152.5 152.5 0 0 0-26.5 13q-.128-1.494 1-2.5-2.674-1.6-1.5-4.5 3.99.553 8.5-2v-3a53 53 0 0 1-7-1.5q4.916-3.473 11-4-9.546-1.471 1-1.5-.274-3.654 2-6.5 2.342-.282 2 2.5 9.487.006 1-4 1.449-3.986 1.5-8l-3 2q-3.275-5.216 2.5-3.5a11.3 11.3 0 0 0 1.5-3.5l2 8 1-5q.779 5.272-1 10a18.8 18.8 0 0 1 5.5-4.5q-1.392-.582-2-2a20 20 0 0 1 3-4.5q-1.509-1.344 0-2l2 4q-.143-1.787 1-4"/><path style="opacity:1" fill="#c5c238" d="M275.5 298.5q-7.28.3-1 4.5-7.912 4.182-3 11.5-4.693.479-8.5 3-1.861-2.157-1.5-4.5l-7-3.5-1-2v-3a4.93 4.93 0 0 0 3-.5l-1-1q1.884-.625 2-2.5 1.426-2.128 1-5a43.4 43.4 0 0 0 2.5-10q1.823 3.693 2 8 1.269-8.202 4.5-12 2.436.21 3.5-2 3.25 2.25-.5 3.5 1.663 2.72-2 2.5.32 2.466-1 4.5 2.256 2.236 2.5 5.5 1.63-3.648 1.5 1 2.975-.514 4 2"/><path style="opacity:1" fill="#bfbd30" d="M230.5 276.5q-1.144 1.152-3 1.5.729 3.04 0 6.5v3h-5q.329 1.763-1 3-1.856-5.716-2 0-2.217.393-4-1 2.31-1.548 1.5-4-2.442.895-5-1a25.7 25.7 0 0 1-7.5 2.5q7.641 2.187-1 2.5a8.4 8.4 0 0 0 .5 4l1-2q1.665 2.305 4.5 2-.21 2.436 2 3.5-3.737.939 0 2.5-4.725 3.095-3 8-4.942.435-6 5-2.976-.175-5 2-5.835-2.331-11 1-1.393-1.783-1-4h-3q-.5 8-1 0v-10q7.903-.078 16-3.5a4.46 4.46 0 0 1-2-1.5 32.4 32.4 0 0 0 1.5-6q1.268-4.368 5-3a2.43 2.43 0 0 0 .5-2q13.045-6.775 28-9"/><path style="opacity:1" fill="#8ba202" d="M311.5 289.5a251 251 0 0 0 15 4 58 58 0 0 1 8 4q-2.289.907-3 3.5 3.103.486 1 2.5 2.094 2.028 2 5a28.2 28.2 0 0 0-2 7q-2.464-.03-4 2-4.109-2.596 1-3.5-2-2 0-4-1.162-1.845-3-1a20.5 20.5 0 0 1-2.5 4.5 4.46 4.46 0 0 0-1.5-2q-1.964 8.175 0 10 3.8-.361 5 3a8.4 8.4 0 0 0 .5 4l1-1q1.35 1.905 3.5 3 2.03 1.536 2 4a58.6 58.6 0 0 1-8.5-4q-.561-3.597 0-7a7.3 7.3 0 0 0-2 3q-2.092-1.972-3 1-5.13-2.622-8.5-7.5-1.863-.466-3 1.5-3.504-2.36 0-4-3.981-3.592 1-6l-6-2q1.717-2.086 0-4-2.806-1.9-5.5 0a248 248 0 0 0-1.5-11q.425-4.864 2.5-1-.327-3.13.5-6 2.417 1.23 2 4a57.4 57.4 0 0 1 9-2"/><path style="opacity:1" fill="#748a01" d="M72.5 288.5a5.73 5.73 0 0 0-1 3q-2.186 2.613-3.5 6l-1-1a10.94 10.94 0 0 1-4.5 3 55 55 0 0 0-2 4.5h-4q4.44 5.172-2.5 6.5-1.934 8.247-7.5 15-5.245-1.466-1 2.5a31.2 31.2 0 0 1-3 9.5l2 2q-1.672 1.647 0 3.5-2.53 2.779-3 6.5l1.5 9q.981-1.794 1.5 0a4.46 4.46 0 0 1-2 1.5q2.578 5.265 2.5 11.5a171 171 0 0 0 6.5 17 810 810 0 0 1 2-11 91 91 0 0 1-1 19 3.65 3.65 0 0 1-1.5-1q-1.627 1.854-2.5 4-2.818.16-1-1.5a327 327 0 0 0-6-11.5 45.3 45.3 0 0 0-3.5-11q-1.062-1.321-2.5-2 2.876-3.765 7-3.5-.387-7.437-4.5-13.5-.939-3.32.5-6.5-2.808-4.965 0-10.5-1.467-4.693 0-10h3v-6q1.711-4.04-2.5-2-.987-2.412 1.5-3.5a4.94 4.94 0 0 1-1-2.5 13 13 0 0 1 3-4l-2-2a64.5 64.5 0 0 0 3.5-8q1.476 3.442 2.5-1 3.066-.847 5-3 10.619-3.027 21-9"/><path style="opacity:1" fill="#778702" d="M326.5 291.5a317 317 0 0 0 25 9q.166.926 1 1.5-3 1-6 0-6.681-1.381-12-4.5a58 58 0 0 0-8-4z"/><path style="opacity:1" fill="#c0be36" d="M162.5 294.5a95 95 0 0 0 5 5.5 10.3 10.3 0 0 0-4 1.5q5.481 1.319 11 2.5 2.115.552 3 2.5a21.6 21.6 0 0 1-4 5 174 174 0 0 1-3.5-3 49.4 49.4 0 0 0 1.5 12q-1.237 1.329-3 1v-1q1.12-5.4-.5-11-2.72 1.89-5.5 0a14.7 14.7 0 0 1 5-1.5 4.93 4.93 0 0 0-3-.5q-.273-7.941-1.5 0a27 27 0 0 0-4-8q-1.363 7.845-3.5 0-3.027-1.97-3.5 2l-1-3-1 6a34.2 34.2 0 0 1-6.5 3q1.635 1.663 0 3-3.888-7.687-3 2 3.74-.554 2 2h-3q-2.036 14.433-2.5 39a113 113 0 0 1-.5-15v-37h4a23.8 23.8 0 0 1 1-9 10.3 10.3 0 0 0-4-1.5 28 28 0 0 0 5-1.5q1.536 2.03 4 2a8.4 8.4 0 0 1 .5-4 34.3 34.3 0 0 0 1.5 8l2 1a14.7 14.7 0 0 0 1.5-5q1.71 2.366 1.5 5l2-2a84 84 0 0 0 6.5 3q-.163-1.904.5-3"/><path style="opacity:.004" fill="#bbbf54" d="M365.5 303.5q2 1 0 0"/><path style="opacity:.008" fill="#b0b73d" d="M365.5 304.5q1.368 1.857.5 4-2.445-1.977-.5-4"/><path style="opacity:1" fill="#596d00" d="M80.5 284.5q.814.09 1 1-2.38 2.51-2 6-3.372 5.463-4 12h-3a8.44 8.44 0 0 1-.5 4 89 89 0 0 1-6 9q-1.87 12.668-7 24a49.5 49.5 0 0 0-1.5 13q-1.356 1.063-1 3-3.378 1.25-6.5 0-.75-9.487-.5-19h3q.594-3.816-2-6.5 3.19-1.577 5-4.5a10.3 10.3 0 0 0-1-3 32.5 32.5 0 0 0 5-10q-2.121-7.403 6-6a18.4 18.4 0 0 1 .5-6q1.112 1.712 2.5 0 .972-2.473 3-4 1.94-3.204 0-6a5.73 5.73 0 0 1 1-3z"/><path style="opacity:1" fill="#5c6901" d="M135.5 303.5q.442 3.353-1 6a14 14 0 0 1-3.5-1 35.7 35.7 0 0 1-9 3 55.2 55.2 0 0 0-10.5 6.5 42.6 42.6 0 0 1-5.5 11.5l-1-2a8.4 8.4 0 0 0-.5 4q-6.23.397-10-4a55.5 55.5 0 0 1 13-12 152.5 152.5 0 0 1 26.5-13q1.147.105 1.5 1"/><path style="opacity:1" fill="#c5c239" d="M284.5 297.5q5.393 4.335 0 9.5.765 2.74 2.5.5 1 2 0 4-3.526-3.815-3.5 2a18.5 18.5 0 0 1 6 .5q-2.909 1.362-3 4.5a81 81 0 0 1-2.5-3q-1.016 3.283-3.5 1-2.254 1.635-1 3.5-2.538 2.101-2 5.5-4.596.534-8-2.5 4.728-2.716 4-7.5-2.257 3.946-2-1-4.912-7.318 3-11.5-6.28-4.2 1-4.5 1.725 2.026 2-.5-2.523-2.843-1.5-6.5l3 3q1.34-1.817 2.5 0a16.1 16.1 0 0 0 2.5 5 2.43 2.43 0 0 0 .5-2"/><path style="opacity:1" fill="#c3c137" d="M289.5 281.5q1.927 4.633 2 10-1.185-1.663-1.5-4-.141 16.346 4 32l1 18a11.3 11.3 0 0 0 3.5 1.5 25.3 25.3 0 0 1-6 2.5l-3-1q1.509-1.344 0-2-2.543 1.011-5 2 .062-3.368 2.5-2a12.5 12.5 0 0 0 0-7 4.46 4.46 0 0 1-1.5 2q-1.72-3.397-.5-7 1.843 3.67 4.5.5-9.786-3.797 0-7-1.116-1.558-3-1.5.091-3.138 3-4.5a18.5 18.5 0 0 0-6-.5q-.026-5.815 3.5-2 1-2 0-4-1.735 2.24-2.5-.5 5.393-5.165 0-9.5-1.684-3.027 0-6a10.94 10.94 0 0 0-3-4.5 16.7 16.7 0 0 1 4-4q1.926-1.026 3.5.5a2.43 2.43 0 0 0 .5-2"/><path style="opacity:1" fill="#849b02" d="M75.5 303.5q-.635 4.904 0 9.5l1 1.5 2-3q5.404 6.567 1 14-1.296 1.445-2 0 1.809-3.024 0-6l-2 4a21.5 21.5 0 0 1-5-6q-.67 5.675-3 10.5l1 1q-1.802 3.87-5 6.5a41.7 41.7 0 0 1 2.5 13 4.45 4.45 0 0 1 1.5-2q1.83 1.817 3.5 0 .725 1.422.5 3-2.745-.828-5 0a27 27 0 0 1-3.5 6 10.3 10.3 0 0 0-1.5-4q-4.809 4.812 0 8.5-3.817 1.912-6 5.5a40.9 40.9 0 0 1 2-12 49.5 49.5 0 0 1 1.5-13q5.13-11.332 7-24a89 89 0 0 0 6-9 8.44 8.44 0 0 0 .5-4z"/><path style="opacity:1" fill="#8a9d03" d="m104.5 274.5-1.5 6a208.8 208.8 0 0 0-19.5 47q-4.32 5.871-7 13 .423-4.27 0-8 3.408-2.329 4-6.5-.342-.598-1-.5 4.404-7.433-1-14l-2 3-1-1.5q-.635-4.596 0-9.5.628-6.537 4-12 1.083-.626 1.5-2 2.903 3.389 7.5 3 .323-5.65 4-10 2.695-3.378 7-3-.507-5.476 5-5"/><path style="opacity:1" fill="#556001" d="M249.5 304.5q1.144 2.643 4 3l1 2a45.6 45.6 0 0 1 7 9q-.09.814-1 1-3.682-6.678-11-9-1.94-3.204 0-6"/><path style="opacity:1" fill="#bcba2f" d="M257.5 300.5q-.116 1.875-2 2.5l1 1a4.93 4.93 0 0 1-3 .5v3q-2.856-.357-4-3-.09-.814-1-1a31.1 31.1 0 0 0-1 11h-1a25 25 0 0 0-.5-7 165 165 0 0 1-1.5 17q-1.203 2.178-1.5 5-5.315-3.546-9.5 2-1.329-1.237-1-3-4.915-.426-9.5 1l-.5-1a6.8 6.8 0 0 1 3-1.5 18.3 18.3 0 0 1-6-4q2.21-1.064 2-3.5 1.763.329 3-1l-1-1q-2.401.998-4.5 1-1.5-4-3 0a32 32 0 0 1-2.5-4q.306 1.543-1.5 3-1.112-1.712-2.5 0-1.329-1.237-1-3-8.11-1.598-12 5.5 1.129 1.006 1 2.5a10.7 10.7 0 0 0-5-4 25.8 25.8 0 0 0-7 5q-4.572 1.125-5 6a6.85 6.85 0 0 1-4 1q.21-2.436-2-3.5a12.6 12.6 0 0 1 4-2.5l3-3q-1.237-1.329-3-1a12.9 12.9 0 0 1 .5-5l1 1q.297-2.822 1.5-5 .5 8 1 0h3q-.393 2.217 1 4 5.165-3.331 11-1 2.024-2.175 5-2 1.058-4.565 6-5v5a4.93 4.93 0 0 0-3 .5q2.853.299 4.5 2.5 3.742-3.239 5-8 6.221 2.346.5 5 1.726 1.846 3.5 0 .552 2.115 2.5 3 1.995-2.77-1-4.5a4.93 4.93 0 0 1 3-.5q-.393-2.217 1-4 1.954 1.827 3-1a20 20 0 0 0 3.5 4 10.8 10.8 0 0 1 2.5-3q5.19-2.721 10.5-5 2.997 1.157 5-2 .425 4.864 2.5 1 2.973 1.684 6 0"/><path style="opacity:.004" fill="#a0b117" d="M377.5 314.5q1.538.595.5 2a2.43 2.43 0 0 1-.5-2"/><path style="opacity:1" fill="#bdbb32" d="M168.5 320.5q-1.083.626-1.5 2-1.5-4-3 0l-1-3q-1.065 2.21-3.5 2v-3q-3.755-1.709-8-2v7q1.763.329 3-1-1.238-1.427-.5-3 1.904 2.107 1.5 5-1.884.606-3.5 2a15.9 15.9 0 0 0-4.5-2.5 17.4 17.4 0 0 1 3-6.5q-4.024-1.515-8-3 1.74-2.554-2-2-.888-9.687 3-2 1.635-1.337 0-3a34.2 34.2 0 0 0 6.5-3l1-6 1 3q.473-3.97 3.5-2 2.137 7.845 3.5 0a27 27 0 0 1 4 8q1.227-7.941 1.5 0a4.93 4.93 0 0 1 3 .5 14.7 14.7 0 0 0-5 1.5q2.78 1.89 5.5 0 1.62 5.6.5 11"/><path style="opacity:1" fill="#bec035" d="M291.5 291.5q.33 4.52 4 7-1.61 6.78 1 13 1.423 11.23 6 21v3a128.5 128.5 0 0 1 6 17 144 144 0 0 0 6 15q-.309 3.074-2.5 1a12.9 12.9 0 0 1-.5-5h-6a35 35 0 0 0-8-6l-5-4q-5.415-5.922-11-12 1.937.356 3-1 2.457-.989 5-2 1.509.656 0 2l3 1a25.3 25.3 0 0 0 6-2.5 11.3 11.3 0 0 1-3.5-1.5l-1-18q-4.141-15.654-4-32 .315 2.337 1.5 4"/><path style="opacity:1" fill="#c1c139" d="m254.5 309.5 7 3.5q-.361 2.343 1.5 4.5 3.807-2.521 8.5-3-.257 4.946 2 1 .728 4.784-4 7.5 3.404 3.034 8 2.5-.538-3.399 2-5.5-1.254-1.865 1-3.5 2.484 2.283 3.5-1a81 81 0 0 0 2.5 3q1.884-.058 3 1.5-9.786 3.203 0 7-2.657 3.17-4.5-.5-1.22 3.603.5 7a4.46 4.46 0 0 0 1.5-2 12.5 12.5 0 0 1 0 7q-2.438-1.368-2.5 2-1.063 1.356-3 1a30059 30059 0 0 1-20-23 45.6 45.6 0 0 0-7-9"/><path style="opacity:1" fill="#8da103" d="M79.5 325.5q.658-.098 1 .5-.592 4.171-4 6.5.423 3.73 0 8a267 267 0 0 1-7 17l-6.5 13q-3.315-2.417-7.5-2v-3q2.183-3.588 6-5.5-4.809-3.688 0-8.5a10.3 10.3 0 0 1 1.5 4 27 27 0 0 0 3.5-6q2.255-.828 5 0a4.93 4.93 0 0 0-.5-3q-1.67 1.817-3.5 0a4.45 4.45 0 0 0-1.5 2 41.7 41.7 0 0 0-2.5-13q3.198-2.63 5-6.5l-1-1q2.33-4.825 3-10.5a21.5 21.5 0 0 0 5 6l2-4q1.809 2.976 0 6 .704 1.445 2 0"/><path style="opacity:1" fill="#b9b931" d="M142.5 314.5q3.976 1.485 8 3a17.4 17.4 0 0 0-3 6.5 15.9 15.9 0 0 1 4.5 2.5q1.616-1.394 3.5-2a4.93 4.93 0 0 0 .5 3q1.077-1.426 2.5-.5-1.1 4.946-5.5 7.5l-1 5-1-3q-5.912 4.682-3.5-2a4.93 4.93 0 0 1-3-.5q6.953-1.188 4-7.5.075 2.022-1 4.5-1.574.739-3-.5a6.8 6.8 0 0 0-1.5 3 6.8 6.8 0 0 0-1.5-3l-1 1.5q2.055 3.376 3.5 6.5-2.583.692-5.5.5 1.126 15.035.5 30.5a4.46 4.46 0 0 0-1.5-2 55.4 55.4 0 0 0-2 11q-1.549 2.117-2 5 .401 6.405.5 13 3.304.282 6.5-2-5.507 8.005-11 16l-1-5q2.78-9.687 3-20a21.4 21.4 0 0 0 1-7 102.6 102.6 0 0 0 2-21 323 323 0 0 1 2-19 113 113 0 0 0 .5 15q.464-24.567 2.5-39z"/><path style="opacity:1" fill="#bebc31" d="M244.5 324.5a357 357 0 0 1 0 31q-1.919-4.69-5-9a4.46 4.46 0 0 1 2-1.5 3.94 3.94 0 0 1-1-2q3.403-2.637.5-5.5-.097 3.575-2.5 2a9.4 9.4 0 0 1-3 3q-.835-3.327 0-7-1.983-.255-3.5 1l-2.5-3q-5.937 1.398-11 6a12.9 12.9 0 0 0-.5-5l-1 2q-.779-1.524-2.5-2l-6 2a15 15 0 0 0-5-3 1079 1079 0 0 0-6-11q.129-1.494-1-2.5 3.89-7.098 12-5.5-.329 1.763 1 3 1.388-1.712 2.5 0 1.806-1.457 1.5-3a32 32 0 0 0 2.5 4q1.5-4 3 0 2.099-.002 4.5-1l1 1q-1.237 1.329-3 1 .21 2.436-2 3.5a18.3 18.3 0 0 0 6 4 6.8 6.8 0 0 0-3 1.5l.5 1q4.585-1.426 9.5-1-.329 1.763 1 3 4.185-5.546 9.5-2 .297-2.822 1.5-5"/><path style="opacity:1" fill="#627301" d="M249.5 304.5q-1.94 2.796 0 6 7.318 2.322 11 9 .91-.186 1-1 9.957 11.462 20 23 5.585 6.078 11 12-3.015 3.12-7 1a34.3 34.3 0 0 1-1.5-8q-7.067-.818-7.5-8-1.237-1.329-3-1a4.93 4.93 0 0 0-.5-3q-5.325-2.311-9-7a34.2 34.2 0 0 0-6.5-9q-1.954 1.827-3-1-4.57 1.244-5 6-2.443-4.149-2-9a31.1 31.1 0 0 1 1-11q.91.186 1 1"/><path style="opacity:.315" fill="#8ea112" d="M35.5 291.5q1.838.978 2 3.5a31.5 31.5 0 0 0-.5 11.5 10.3 10.3 0 0 1 1.5-4l2 1a75 75 0 0 1 3-11q1.07 2.902 3 1v3q2.574.193 5 1-1.934 2.153-5 3-1.024 4.442-2.5 1a64.5 64.5 0 0 1-3.5 8l2 2a13 13 0 0 0-3 4 4.94 4.94 0 0 0 1 2.5q-2.487 1.088-1.5 3.5 4.211-2.04 2.5 2a99 99 0 0 1-5 0l1 2a13 13 0 0 0-3 4q-1.313 3.672-1 7l-2-9q-1.77 2.453 0 5a146 146 0 0 0-4 8 65 65 0 0 1-3-10.5l2-1a7.25 7.25 0 0 1-2-2.5q1.452-4.16 4-7.5a37 37 0 0 1-2-4.5q2.205-11.243 5.5-21l1 1q.565-2.163 2.5-3"/><path style="opacity:.005" fill="#9eae19" d="M383.5 321.5q2.399 1.948 1.5 5-1.904-2.107-1.5-5"/><path style="opacity:1" fill="#6f8201" d="M316.5 334.5a168 168 0 0 0-11.5-10q-2.373 1.193-.5 3.5a15.4 15.4 0 0 0-2 4.5q-4.577-9.77-6-21a10.1 10.1 0 0 1 1.5-4q1.524 1.701 2-1l1 7a74 74 0 0 1 6.5 8 135 135 0 0 1 7 2.5 4.93 4.93 0 0 1-3 .5q.959 3.39 3.5 6 1.34-1.817 2.5 0a6.85 6.85 0 0 0-1 4"/><path style="opacity:1" fill="#6a8101" d="M134.5 309.5a65 65 0 0 1-1 16q-.858 3.626-3 6.5-1.475.886-2-1 1.245-2.197.5-4.5-.803 2.656-2.5 1 1.593-5.107-3-1.5.686.621 1 1.5a22.8 22.8 0 0 0-5 4 3.65 3.65 0 0 1-1.5-1q-.368 3.245-3 5-2.187-1.87-4.5 0a525 525 0 0 1-6-4 8.4 8.4 0 0 1 .5-4l1 2a42.6 42.6 0 0 0 5.5-11.5 55.2 55.2 0 0 1 10.5-6.5q4.701-.9 9-3 1.707.733 3.5 1"/><path style="opacity:1" fill="#91a802" d="M334.5 297.5q5.319 3.119 12 4.5 3 1 6 0-.834-.574-1-1.5 3.087 2.15 7 2a44 44 0 0 0 3 2 26 26 0 0 1 0 5.5q2.262 5.404 6 7.5.729 3.46 0 6.5 1.374.417 2 1.5-7.622 2.719-7 11h-3q-.255-1.983 1-3.5a4.93 4.93 0 0 0-3-.5q.828 2.746 0 5-1.83-1.817-3.5 0 .028-2.41.5-5a16.9 16.9 0 0 1-8-1q1.712-1.388 0-2.5 3.283-1.016 1-3.5a20.6 20.6 0 0 0-9-.5 3.65 3.65 0 0 1-1-1.5q2.258-2.012 2-5-2.893-.404-5 1.5 2.36 4.137 1 8.5-2.674-2.285-2-6-1.983.255-3.5-1a81 81 0 0 1-2.5 3q-1.2-3.361-5-3-1.964-1.825 0-10a4.46 4.46 0 0 1 1.5 2 20.5 20.5 0 0 0 2.5-4.5q1.838-.845 3 1-2 2 0 4-5.109.904-1 3.5 1.536-2.03 4-2a28.2 28.2 0 0 1 2-7q.094-2.972-2-5 2.103-2.014-1-2.5.711-2.593 3-3.5"/><path style="opacity:1" fill="#657600" d="M197.5 322.5q3.062 5.537 6 11a874 874 0 0 1 25 50q-.026 2.292-2 1.5a20.2 20.2 0 0 0-4-8.5 33.3 33.3 0 0 1-8-5 4.93 4.93 0 0 1-3-.5q1.884-.625 2-2.5 1.954-3.385-2-5.5a10.3 10.3 0 0 0 4-1.5q-2.453-1.771-5 0 1.334-5.17 0-10-2.425-4.926-7-8a44.6 44.6 0 0 1-4-12q-4.623-3.769-6.5-8-2.033 4.848-5 2-10.053 9.912-10.5 24a6.85 6.85 0 0 1-4-1 898 898 0 0 1-12 23 39.7 39.7 0 0 0-1 11q-5.944-.945-7-7a503 503 0 0 1 21-35 140 140 0 0 0 11-17 25.8 25.8 0 0 1 7-5 10.7 10.7 0 0 1 5 4"/><path style="opacity:1" fill="#849902" d="M289.5 277.5a177 177 0 0 1 22 12 57.4 57.4 0 0 0-9 2q.417-2.77-2-4-.827 2.87-.5 6-2.075-3.864-2.5 1a248 248 0 0 1 1.5 11q2.694-1.9 5.5 0 1.717 1.914 0 4l6 2q-4.981 2.408-1 6-3.504 1.64 0 4 1.137-1.966 3-1.5 3.37 4.878 8.5 7.5.908-2.972 3-1a7.3 7.3 0 0 1 2-3q-.561 3.403 0 7a58.6 58.6 0 0 0 8.5 4 5.73 5.73 0 0 0 1 3q.915 1.337 1 3-1.884.606-3.5 2a37 37 0 0 1-6-8q-2.07 3.06-1.5 6a19.2 19.2 0 0 0-5-3h-1q0-3-3-3a6.85 6.85 0 0 1 1-4q-1.16-1.817-2.5 0-2.541-2.61-3.5-6a4.93 4.93 0 0 0 3-.5 135 135 0 0 0-7-2.5 74 74 0 0 0-6.5-8l-1-7q-.476 2.701-2 1a10.1 10.1 0 0 0-1.5 4q-2.61-6.22-1-13-3.67-2.48-4-7-.073-5.367-2-10z"/><path style="opacity:1" fill="#677d01" d="M71.5 291.5q1.94 2.796 0 6-2.028 1.527-3 4-1.388 1.712-2.5 0a18.4 18.4 0 0 0-.5 6q-8.121-1.403-6 6a32.5 32.5 0 0 1-5 10q.73 1.419 1 3-1.81 2.923-5 4.5 2.594 2.684 2 6.5h-3q-.25 9.513.5 19 3.122 1.25 6.5 0-.356-1.937 1-3a40.9 40.9 0 0 0-2 12h-1q-1.475 5.697-1 12a810 810 0 0 0-2 11 171 171 0 0 1-6.5-17q.078-6.235-2.5-11.5a4.46 4.46 0 0 0 2-1.5q-.519-1.794-1.5 0l-1.5-9q.47-3.721 3-6.5-1.672-1.853 0-3.5l-2-2a31.2 31.2 0 0 0 3-9.5q-4.245-3.966 1-2.5 5.566-6.753 7.5-15 6.94-1.328 2.5-6.5h4a55 55 0 0 1 2-4.5 10.94 10.94 0 0 0 4.5-3l1 1q1.314-3.387 3.5-6"/><path style="opacity:1" fill="#738b01" d="M213.5 368.5q-1.451-1.96-4-2v-3a4.93 4.93 0 0 1-3-.5q2.877-2.655 1.5-6.5-6.93-1.995-4.5-8.5-.404-2.41-2.5-3.5.585-14.161-13.5-12-1.226 4.468-2.5 9a35.6 35.6 0 0 0-5.5 9q-1.442 2.648-1 6h-4q.028 2.34 1 4.5-4.709 6.089-3.5 13.5-4.398-5.015-8.5 1a6.8 6.8 0 0 0 3 1.5l-2.5 2.5a244 244 0 0 0-4.5 11.5q.686.621 1 1.5a1963 1963 0 0 0-4 7q-4.182-3.862-8 0 4.936.829 7 5-.481 2.322-3 2 .255-1.983-1-3.5a14.7 14.7 0 0 1-5-1.5l2-7q-3.233-.751-5-3a100 100 0 0 1 10-16q1.056 6.055 7 7a39.7 39.7 0 0 1 1-11 898 898 0 0 0 12-23 6.85 6.85 0 0 0 4 1q.447-14.088 10.5-24 2.967 2.848 5-2 1.877 4.231 6.5 8a44.6 44.6 0 0 0 4 12q4.575 3.074 7 8 1.334 4.83 0 10 2.547-1.771 5 0a10.3 10.3 0 0 1-4 1.5q3.954 2.115 2 5.5"/><path style="opacity:1" fill="#768d01" d="M246.5 314.5h1q-.443 4.851 2 9 .43-4.756 5-6 1.046 2.827 3 1a34.2 34.2 0 0 1 6.5 9q3.675 4.689 9 7 .724 1.422.5 3 1.763-.329 3 1 .433 7.182 7.5 8a34.3 34.3 0 0 0 1.5 8q3.985 2.12 7-1l5 4v2a24.3 24.3 0 0 1-5.5 2 161 161 0 0 0-12.5-9l-6-4q1.477-2.397.5-5-1-2-2 0l-1-3q-1.077 1.426-2.5.5a32.5 32.5 0 0 1-5-4.5l-2 3a48 48 0 0 1-2.5-5q-2.785 5.82-4-1-5.313 8.317 1.5 15.5a20 20 0 0 1-3 4.5l3 2a4.46 4.46 0 0 1-2 1.5l5 1a10.3 10.3 0 0 0-4 1.5 3.65 3.65 0 0 0 1 1.5q-2.42 3.342-2 7.5 4.632-.455 5 4-3.321-.314-5 2.5 1.525 3.71 0 7.5h-7q.401-2.398-1-4z"/><path style="opacity:1" fill="#738901" d="M135.5 303.5h1v35a323 323 0 0 0-2 19q-.925.166-1.5 1-1.381-8.424-8.5-13-4.25 1.974-2.5 6a3.94 3.94 0 0 1-2 1 14.3 14.3 0 0 0-5.5-2.5q-.439-4.641-5-4.5-.865 5.37 3 9-6.773.38-6 7h-3q-1.208-6.312 0-12.5-1.469-17.09-16-8.5a29.8 29.8 0 0 0 2-7q2.516-3.06 5-6 3.77 4.397 10 4a525 525 0 0 0 6 4q2.313-1.87 4.5 0 2.632-1.755 3-5 .621.686 1.5 1a22.8 22.8 0 0 1 5-4 3.65 3.65 0 0 0-1-1.5q4.593-3.607 3 1.5 1.697 1.656 2.5-1 .745 2.303-.5 4.5.525 1.886 2 1 2.142-2.874 3-6.5a65 65 0 0 0 1-16q1.442-2.647 1-6"/><path style="opacity:1" fill="#b2b432" d="M94.5 303.5a75 75 0 0 0 2 12l2-2q3.886 4.215 9 2a55.5 55.5 0 0 0-13 12 576 576 0 0 0-5 6 669 669 0 0 1-16 22 20.2 20.2 0 0 0-1.5 6q-1 2-2 0a8.44 8.44 0 0 1-.5-4 267 267 0 0 0 7-17q2.68-7.129 7-13 .975-.456 2-1 1.41-4.954 2.5-10a35.6 35.6 0 0 0 4.5-4 3.65 3.65 0 0 0-1-1.5q3.237-1.438 1-4.5.2-3.219 2-3"/><path style="opacity:1" fill="#809900" d="M209.5 366.5q-5.99.75-5-4.5a88 88 0 0 1-8-.5l5-3.5a39.5 39.5 0 0 1-6-4.5q-2.495 2.063-4-1l-1 4q1.674-.147 3 1.5-1.473 3.014-4 1.5-2.965-5.364-1-11h-3q.22 3.663-2.5 2-1.076.816-1.5 2l-2-2a35.6 35.6 0 0 1 5.5-9q1.274-4.532 2.5-9 14.085-2.161 13.5 12 2.096 1.09 2.5 3.5-2.43 6.505 4.5 8.5 1.377 3.845-1.5 6.5a4.93 4.93 0 0 0 3 .5z"/><path style="opacity:1" fill="#98ae04" d="M369.5 325.5q.759 2.893 3 5-.363 4.229 0 9 1.983-.255 3.5 1 .324 2.826 2 5a30.3 30.3 0 0 1-.5 10q-2.464-.03-4 2-1.89-2.78 0-5.5l-1-.5-2 3a20.2 20.2 0 0 1-1.5-6l-1 1q-4.573-9.795-3.5 2h-3q.957-2.94 0-5.5-2.17 1.758-5 3.5-3.623-.328-3-4-7.97.723-7-7a39.7 39.7 0 0 1-11-1 5.73 5.73 0 0 1-1-3q.03-2.464-2-4-2.15-1.095-3.5-3l-1 1a8.4 8.4 0 0 1-.5-4 81 81 0 0 0 2.5-3q1.517 1.255 3.5 1-.674 3.715 2 6 1.36-4.363-1-8.5 2.107-1.904 5-1.5.258 2.988-2 5a3.65 3.65 0 0 0 1 1.5 20.6 20.6 0 0 1 9 .5q2.283 2.484-1 3.5 1.712 1.112 0 2.5a16.9 16.9 0 0 0 8 1 30 30 0 0 0-.5 5q1.67-1.817 3.5 0 .828-2.254 0-5a4.93 4.93 0 0 1 3 .5q-1.255 1.517-1 3.5h3q-.622-8.281 7-11"/><path style="opacity:1" fill="#7e9701" d="M279.5 352.5a16.6 16.6 0 0 0-2.5 4q-2.014-11.968-4 0-3 1-4 4-.181-3.34-2.5-5.5a7.3 7.3 0 0 1 3-2 3.65 3.65 0 0 1-1-1.5l3-2.5q-6.678-5.641-7 3.5-1.948-.885-2.5-3a164 164 0 0 1-1.5 9.5q.462 4.43 4 7l-3 5q.903 2.344 2.5-.5.853 5.792 2.5 0 2.288 9.58-2 18.5 1.815 4.761 0 9.5a20.5 20.5 0 0 1-4.5 5q-2.57-2.182-3.5-5.5a4.93 4.93 0 0 0-3-.5q-8.328-4.994-6-15h7q1.525-3.79 0-7.5 1.679-2.814 5-2.5-.368-4.455-5-4-.42-4.158 2-7.5a3.65 3.65 0 0 1-1-1.5 10.3 10.3 0 0 1 4-1.5l-5-1a4.46 4.46 0 0 0 2-1.5l-3-2a20 20 0 0 0 3-4.5q-6.813-7.183-1.5-15.5 1.215 6.82 4 1a48 48 0 0 0 2.5 5l2-3a32.5 32.5 0 0 0 5 4.5q1.423.926 2.5-.5l1 3q1-2 2 0 .977 2.603-.5 5z"/><path style="opacity:1" fill="#bbb831" d="M168.5 320.5v1q-.332 1.853-2 3l1.5 1q3.862-2.731 8-1 1.67-5.934 2.5 0a12.6 12.6 0 0 0-4 2.5q2.21 1.064 2 3.5a6.85 6.85 0 0 0 4-1q.428-4.875 5-6a140 140 0 0 1-11 17q-1.494-.129-2.5 1a53 53 0 0 1-9-6q-.173 2.41 1.5 4.5a6.2 6.2 0 0 1-2 3.5q-1.87-2.259 0-4.5-1.423-.926-2.5.5-2.426-1.554-.5-3.5-2.098-.892-4 .5a97 97 0 0 0-3 8q2.464-.03 4 2 4.048-2.733-.5-4-1-2 0-4a70.4 70.4 0 0 0 9.5 10.5q-2.469 8.197-9 13.5-1.87-2.259 0-4.5a4.93 4.93 0 0 0-3-.5q.362 5.248-2.5 1l-1 1q-.534-5.78-4.5-10a41 41 0 0 1 0 7.5 32.4 32.4 0 0 1 4 5.5l-5 2q2.935 2.36 1 6 .314 2.241 2 3.5-7.017.295-1 3-1.87 2.187 0 4.5-1.223 2.664-3.5 2-1-8-2 0a38 38 0 0 1-2.5-7 186 186 0 0 1-3 2 55.4 55.4 0 0 1 2-11 4.46 4.46 0 0 1 1.5 2q.626-15.465-.5-30.5 2.917.192 5.5-.5-1.445-3.124-3.5-6.5l1-1.5a6.8 6.8 0 0 1 1.5 3 6.8 6.8 0 0 1 1.5-3q1.426 1.239 3 .5 1.075-2.478 1-4.5 2.953 6.312-4 7.5a4.93 4.93 0 0 0 3 .5q-2.412 6.682 3.5 2l1 3 1-5q4.4-2.554 5.5-7.5-1.423-.926-2.5.5a4.93 4.93 0 0 1-.5-3q.404-2.893-1.5-5-.738 1.573.5 3-1.237 1.329-3 1v-7q4.245.291 8 2v3q2.435.21 3.5-2l1 3q1.5-4 3 0 .417-1.374 1.5-2"/><path style="opacity:1" fill="#8ea501" d="M335.5 337.5a39.7 39.7 0 0 0 11 1q-.97 7.723 7 7-.623 3.672 3 4 2.83-1.742 5-3.5.957 2.56 0 5.5h3q-1.073-11.795 3.5-2l1-1a20.2 20.2 0 0 0 1.5 6l2-3 1 .5q-1.89 2.72 0 5.5 1.536-2.03 4-2a6.85 6.85 0 0 0 1 4q-3.157 2.639-5 6v-4q-6.36.689-10.5-4a7.3 7.3 0 0 1-2 3q-1.948-.699-4.5 0h-3a8.4 8.4 0 0 0-.5-4l-1 1q-1.743-3.508-5.5-4 .276-4.232-4-5l-2-5a6.8 6.8 0 0 0-1.5 3 219 219 0 0 0-2.5-6 5.73 5.73 0 0 0-1-3"/><path style="opacity:.193" fill="#a2af33" d="M381.5 333.5q2.622 1.933 1 4.5 1.684 1.565 4 2l-2 1q5.038 3.111 7.5 9.5 1 2 2 0 .926 1.423-.5 2.5 2.404 4.375 6 7.5-1.602 1.401-4 1-4.716-5.233-7-12-5.79-5.759-9-13 .21-2.295 2-3"/><path style="opacity:1" fill="#bcbc33" d="M246.5 314.5v64a227 227 0 0 0 2 30q-.084 3.21-2 5-.5-1.5-2-2-5.993-11.496-12-23l-4-5a874 874 0 0 0-25-50q2.765.99 5 3l6-2q1.721.476 2.5 2l1-2a12.9 12.9 0 0 1 .5 5 5.6 5.6 0 0 0-3 1q-1.96 4.745 1 9 1.447-3.875 4-1.5-2.567 1.7-1.5 4.5 2.752-3.696 6 0a3.94 3.94 0 0 0 2-1 8.4 8.4 0 0 1 .5 4 8.4 8.4 0 0 0-4 .5q2.401.41 4.5 1.5l1-3q.951 8.452 1.5 1 8.087-.303 5 7-3.261-.63-6 0a115 115 0 0 1 8.5 10q-1.817-5.039 3-8-.485 6.295 2.5 1a25.9 25.9 0 0 0 1-10 357 357 0 0 0 0-31 165 165 0 0 0 1.5-17q.745 3.465.5 7"/><path style="opacity:1" fill="#bfbb32" d="M244.5 355.5a25.9 25.9 0 0 1-1 10q-2.985 5.295-2.5-1-4.817 2.961-3 8a115 115 0 0 0-8.5-10q2.739-.63 6 0 3.087-7.303-5-7-.549 7.452-1.5-1l-1 3a15.3 15.3 0 0 0-4.5-1.5 8.4 8.4 0 0 1 4-.5 8.4 8.4 0 0 0-.5-4 3.94 3.94 0 0 1-2 1q-3.248-3.696-6 0-1.067-2.8 1.5-4.5-2.553-2.375-4 1.5-2.96-4.255-1-9a5.6 5.6 0 0 1 3-1q5.063-4.602 11-6l2.5 3q1.517-1.255 3.5-1-.835 3.673 0 7a9.4 9.4 0 0 0 3-3q2.403 1.575 2.5-2 2.903 2.863-.5 5.5a3.94 3.94 0 0 0 1 2 4.46 4.46 0 0 0-2 1.5q3.081 4.31 5 9"/><path style="opacity:1" fill="#4a5d01" d="M319.5 337.5h1a377 377 0 0 1 5 10l8 5 2 7q5.355.601 6.5 6 1.006 1.129 2.5 1 1.52 1.76 4 2v3q3.615.631 5.5 4 .915 4.991 1.5 10h-3a301 301 0 0 0-5-8 148 148 0 0 1-15-18 44.7 44.7 0 0 0-9-7 72 72 0 0 0-4.5-9 13.9 13.9 0 0 0-4-3q-.466-6.83 4.5-3"/><path style="opacity:1" fill="#7a9101" d="M134.5 357.5q.075 10.673-2 21-3.904-3.515-9-5.5l-1-1q4.092-.959 1.5-3.5-2.241 1.87-4.5 0-2.856-7.32-5 1h4q-.165 4.33-3.5 7-2.926-.582-4.5-3 .657-1.509 2 0 1.329-1.237 1-3-2.976.175-5-2-1.816 1.16 0 2.5a16.6 16.6 0 0 1-7-.5l-1-8q-5.453-.233-1-3.5-3.11-4.754-3.5-10.5-2.241-1.87-4.5 0a15.7 15.7 0 0 1-2 4q-6.017-2.534-5-9 .534-2.535 3-3 14.531-8.59 16 8.5-1.208 6.188 0 12.5h3q-.773-6.62 6-7-3.865-3.63-3-9 4.561-.141 5 4.5a14.3 14.3 0 0 1 5.5 2.5 3.94 3.94 0 0 0 2-1q-1.75-4.026 2.5-6 7.119 4.576 8.5 13 .575-.834 1.5-1"/><path style="opacity:1" fill="#859e01" d="M279.5 352.5a161 161 0 0 1 12.5 9 24.3 24.3 0 0 0 5.5-2q2.371 2.463 2 6a76 76 0 0 1-8-2.5 20.8 20.8 0 0 0 3 7.5q-1.337 1.635-3 0l-6-10q-.724 4.754-4 8 3.942 4.923-3 5.5 6.46 1.299 1 3 1.75.75 2.5 2.5a3.65 3.65 0 0 1 1.5-1q1.817 1.16 0 2.5 1.5 1 3 0-1.536-2.05 1.5-2.5 1.353.365 2.5 1 0 1-1 1-3.328 4.056-1 9 6.785-.785 3 4-2.031-2.526-5-1.5a4.46 4.46 0 0 1 2 1.5 23.5 23.5 0 0 0-7 3.5q2.1 2.157 5 2.5-.417-2.77 2-4 4.808 1.358 9 1a55 55 0 0 1-4.5 2q-2.082-1.58-4.5-.5 2 1.5 0 3 4.93 3.198 2 8.5a71 71 0 0 0-2 4q-3.572-2.398-3.5-7-3.518-.934-6-5-.75 1.75-2.5 2.5 10.215 3.942-1 4.5v3a8.4 8.4 0 0 1 4 .5q-7.172 1.69-5-5-1.729-2.824-5-2.5.872 7.94-3 2 4.214-11.972-3-2a11.3 11.3 0 0 1 1 3.5 27.2 27.2 0 0 0-2 5.5q-1.763-.329-3 1 3.626 4.173 3.5 9 1.5-2 3 0 .747-3.04-.5-6 1.794-3.957 6-2.5-1.816 2.767-4 2.5a42.8 42.8 0 0 0 9 1.5l-2 2a22.5 22.5 0 0 0 7 2q-2.429.354-3 2.5h-21q.444-4.144 3-7.5a9.5 9.5 0 0 1-2.5-1.5 18.5 18.5 0 0 1-.5-6h4q.32-2.466-1-4.5a60 60 0 0 1-6-.5 16.9 16.9 0 0 1-1-8 4.93 4.93 0 0 1 3 .5q.93 3.318 3.5 5.5a20.5 20.5 0 0 0 4.5-5q1.815-4.739 0-9.5 4.288-8.92 2-18.5-1.647 5.792-2.5 0-1.597 2.844-2.5.5l3-5q-3.538-2.57-4-7a164 164 0 0 0 1.5-9.5q.552 2.115 2.5 3 .322-9.141 7-3.5l-3 2.5a3.65 3.65 0 0 0 1 1.5 7.3 7.3 0 0 0-3 2q2.319 2.16 2.5 5.5 1-3 4-4 1.986-11.968 4 0a16.6 16.6 0 0 1 2.5-4"/><path style="opacity:1" fill="#748b02" d="M336.5 340.5a22.1 22.1 0 0 0 2 9.5q-.608 1.418-2 2a45.3 45.3 0 0 0 8 2.5 6.85 6.85 0 0 1 1 4q3.159.268 5 3-2.03 1.536-2 4-2.398-.401-4 1-1.494.129-2.5-1-1.145-5.399-6.5-6l-2-7-8-5a377 377 0 0 0-5-10 19.2 19.2 0 0 1 5 3q-.57-2.94 1.5-6a37 37 0 0 0 6 8q1.616-1.394 3.5-2"/><path style="opacity:.91" fill="#85990a" d="M41.5 323.5v6h-3q-1.467 5.307 0 10-2.808 5.535 0 10.5-1.439 3.18-.5 6.5 4.113 6.063 4.5 13.5-4.124-.265-7 3.5-1.902-3.608-2-8-3.662.571-2-2-9.5-10.442-4-23a146 146 0 0 1 4-8q-1.77-2.547 0-5l2 9q-.313-3.328 1-7a13 13 0 0 1 3-4l-1-2q2.532.063 5 0"/><path style="opacity:1" fill="#7f9901" d="m179.5 350.5 2 2h3q-.507 3.985 2 7l-2 2q2.03 1.536 2 4-4.713.449-6.5 5 1.202 4.809 1 10-3.455-4.1-8.5-4.5a168 168 0 0 0 2 6.5q2.733-1.393 4 2-2.613 3.476-7 4.5a26.6 26.6 0 0 1-2 3.5q-2.691-3.255-6 0 3.523 1.564 6 4a16.3 16.3 0 0 1-2 4q-2.434-1.486-3 2 16.857 8.732-2 6 .822-1.782 1-3.5a110 110 0 0 0-7-3.5q-1.329 1.237-1 3-2.064-4.171-7-5 3.818-3.862 8 0 1.986-3.492 4-7a3.65 3.65 0 0 0-1-1.5 244 244 0 0 1 4.5-11.5l2.5-2.5a6.8 6.8 0 0 1-3-1.5q4.102-6.015 8.5-1-1.209-7.411 3.5-13.5a11.2 11.2 0 0 1-1-4.5h4q-.442-3.352 1-6"/><path style="opacity:1" fill="#b8b72f" d="M174.5 340.5a503 503 0 0 0-21 35 100 100 0 0 0-10 16 6.55 6.55 0 0 0-3 3q-3.196 2.282-6.5 2-.099-6.595-.5-13 .451-2.883 2-5 1.473-.961 3-2a38 38 0 0 0 2.5 7q1-8 2 0 2.277.664 3.5-2-1.87-2.313 0-4.5-6.017-2.705 1-3-1.686-1.259-2-3.5 1.935-3.64-1-6l5-2a32.4 32.4 0 0 0-4-5.5 41 41 0 0 0 0-7.5q3.966 4.22 4.5 10l1-1q2.862 4.248 2.5-1a4.93 4.93 0 0 1 3 .5q-1.87 2.241 0 4.5 6.531-5.303 9-13.5a70.4 70.4 0 0 1-9.5-10.5q-1 2 0 4 4.548 1.267.5 4-1.536-2.03-4-2a97 97 0 0 1 3-8q1.902-1.392 4-.5-1.926 1.946.5 3.5 1.077-1.426 2.5-.5-1.87 2.241 0 4.5a6.2 6.2 0 0 0 2-3.5q-1.673-2.09-1.5-4.5 4.2 3.466 9 6 1.006-1.129 2.5-1"/><path style="opacity:1" fill="#8ca004" d="M89.5 333.5a29.8 29.8 0 0 1-2 7q-2.466.465-3 3a2296 2296 0 0 1-4 8.5q3.67 1.843.5 4.5-2.583.692-5.5.5a7.3 7.3 0 0 0 3 2q-2 1 0 2 6.801-.062 3 4.5a24.7 24.7 0 0 1 2.5 9 20.95 20.95 0 0 0 3.5-8q.871 4.313 3 8-2.318 1.428-5 1-1.083.626-1.5 2-1.949-.699-4.5 0 .316 6.88-4.5 2-1.989.383-3 2l-1-1-1 3q-.981-1.794-1.5 0a18 18 0 0 0 2 3.5q-3.707.255-7 2.5a82 82 0 0 0-3-2.5q2.882-.54 5-2.5 2.257 3.946 2-1-2.835.305-4.5-2-2.532 1.237-.5 3-1.783 1.393-4 1v4q6.59-.064 1 1.5 4.452 3.267-1 3.5 2.103 6.602-.5 13-2.8-3.462-3-8a578 578 0 0 1-.5-34h1v3q4.185-.417 7.5 2l6.5-13a8.44 8.44 0 0 0 .5 4q1 2 2 0a20.2 20.2 0 0 1 1.5-6 669 669 0 0 0 16-22"/><path style="opacity:1" fill="#859e01" d="M209.5 366.5a174 174 0 0 1-4 3q4.016 2.645-1 3.5 8.761.845 1 3.5a9.3 9.3 0 0 1 2 2.5q-1.61 2.125 0 4a11.3 11.3 0 0 0-1.5 3.5l-1-2q-4.377 3.351-4.5 9-1.083-.626-1.5-2-1.098 3.635-4.5 3.5-3.432-4.216-4.5-8.5a8.4 8.4 0 0 0-.5 4h-3a4.93 4.93 0 0 0-.5-3q-1.793 4.03-.5 8-4.037 3.387-7 7.5a4.46 4.46 0 0 1 2 1.5 10.8 10.8 0 0 0-2.5 3q-3.045-3.356-7.5-4.5a4.46 4.46 0 0 0 2-1.5q-2.03-1.536-2-4 3.535.245 7-.5-4.157-1.413-8-.5-2.477-2.436-6-4 3.309-3.255 6 0a26.6 26.6 0 0 0 2-3.5q4.387-1.024 7-4.5-1.267-3.393-4-2a168 168 0 0 1-2-6.5q5.045.4 8.5 4.5.202-5.191-1-10 1.787-4.551 6.5-5 .03-2.464-2-4l2-2q-2.507-3.015-2-7h-3q.424-1.184 1.5-2 2.72 1.663 2.5-2h3q-1.965 5.636 1 11 2.527 1.514 4-1.5-1.326-1.647-3-1.5l1-4q1.505 3.063 4 1a39.5 39.5 0 0 0 6 4.5l-5 3.5q3.835.423 8 .5-.99 5.25 5 4.5"/><path style="opacity:1" fill="#97ab09" d="M367.5 317.5q10.735 4.735 14 16-1.79.705-2 3 3.21 7.241 9 13 2.284 6.767 7 12a65 65 0 0 1 2 9.5q-1.905 10.625-6 .5a10.3 10.3 0 0 1 4-1.5q-2.776-1.256-5.5-2.5l-1 2q-.866-4.035-3.5-8a25 25 0 0 0-7 .5q-.967 3.452-1 6.5-1.126-1.61-3-.5.21 1.896 0 3.5a9.86 9.86 0 0 1-1-6q1.843-3.361 5-6a6.85 6.85 0 0 1-1-4q1.082-4.976.5-10-1.676-2.174-2-5-1.517-1.255-3.5-1-.363-4.771 0-9-2.241-2.107-3-5-.626-1.083-2-1.5.729-3.04 0-6.5"/><path style="opacity:1" fill="#829a01" d="M84.5 343.5q-1.017 6.466 5 9a15.7 15.7 0 0 0 2-4q2.259-1.87 4.5 0 .39 5.746 3.5 10.5-4.453 3.267 1 3.5l1 8q3.479 1 7 .5-1.816-1.34 0-2.5 2.024 2.175 5 2 .329 1.763-1 3-1.343-1.509-2 0 1.574 2.418 4.5 3 3.335-2.67 3.5-7h-4q2.144-8.32 5-1 2.259 1.87 4.5 0 2.592 2.541-1.5 3.5l1 1q5.096 1.985 9 5.5a21.4 21.4 0 0 1-1 7q-3.04-.885-5 2.5a38 38 0 0 1 2 5.5 18.35 18.35 0 0 0-2 9q-4.526-1.719-1-4.5a4.93 4.93 0 0 0-3-.5 8.4 8.4 0 0 0-.5-4l-1 2q-1.34-1.817-2.5 0a6.8 6.8 0 0 0 3 1.5q-.852 1.944 0 4.5l-.5 3-1-1q-.434 2.316-2 4-.855-5.15-4-9-2.658 6.876-.5 14-2.464-.03-4 2-1.329-1.237-1-3h-3q.588 4.833-1.5 9a10.3 10.3 0 0 0-1.5-4l-1 1q2.165 4.131 2 8h-29q2.11-3.171 6-4.5l-2-2q8.595-1.008 1-3-1.89-2.72 0-5.5 1.794.519 0 1.5 7.056-.021 7-7.5 3.664-.22 2 2.5a6.54 6.54 0 0 0 3 3q-.685.621-1 1.5a32.4 32.4 0 0 1 1.5 6 55 55 0 0 0 6.5-6 7.3 7.3 0 0 1 2 3q1.297 1.624 3 .5-3.732-4.565-1.5-9.5 3.834 1.079 3.5 5 8.26 1.014 2-3.5 3.279-.905.5-3.5-2.574-1.579-3 2-3.067-6.285-9-10-1.08 2.418.5 4.5a10.8 10.8 0 0 0-3 2.5 29 29 0 0 0-4-1.5q5.697-4.476-2-6.5a7.3 7.3 0 0 1 3-2q2.734 1.87 6 2v-3a6.85 6.85 0 0 1 4 1q2.272-6.745-5-7a6.85 6.85 0 0 0-1-4 6.8 6.8 0 0 1 3-1.5q-2.1-2.157-5-2.5-.19 2.55-2 4.5a11.1 11.1 0 0 0 2 3.5q-1.346.253-3 1 1.77 2.547 0 5a6.8 6.8 0 0 1-1.5-3 3.65 3.65 0 0 1-1.5 1 34 34 0 0 0-3-6.5 186 186 0 0 1 2-7.5q2.682.428 5-1-2.129-3.687-3-8a20.95 20.95 0 0 1-3.5 8 24.7 24.7 0 0 0-2.5-9q3.801-4.562-3-4.5-2-1 0-2a7.3 7.3 0 0 1-3-2q2.917.192 5.5-.5 3.17-2.657-.5-4.5a2296 2296 0 0 0 4-8.5"/><path style="opacity:1" fill="#819701" d="M352.5 406.5a862 862 0 0 0-3 19h-10q.174-5.04 1-10h-7q1.663-1.185 4-1.5 1.842-3.21 5-5a8.4 8.4 0 0 0-4-.5q-.793-5.538 3-9.5-7.738-.638-8-8.5-2.044 2.069-3-.5-1.248-3.717-.5-7.5-1.112-1.712-2.5 0-5.365-6.823-12-12 .356-1.937-1-3a144 144 0 0 1-6-15q1.494.129 2.5-1a184 184 0 0 0 10.5 7.5l-2 1a12.4 12.4 0 0 1 4 2.5 3.65 3.65 0 0 1-1 1.5q7.638 3.474 4.5 10.5 2.684 2.594 6.5 2a101 101 0 0 0 4.5 13l4 2q.004 6.573 6.5 5 .537 6.014 1.5 12 1.063-1.321 2.5-2"/><path style="opacity:1" fill="#708701" d="M302.5 335.5q1.899.885 4 1.5l-2 1a55 55 0 0 0 7 2q1.079 3.834 5 3.5 1.056 9.073 10 12 3.531 2.619 3 7 3.604-.16 5.5 3a37 37 0 0 1 4.5 9.5q5.123.608 8 5-7.946 1.702 0 2.5v3h3a18.35 18.35 0 0 0 2 9l-2 7q.225 3.793 3 2a5.73 5.73 0 0 1-1 3q-1.437.679-2.5 2a172 172 0 0 1-1.5-12q-6.496 1.573-6.5-5l-4-2a101 101 0 0 1-4.5-13q-3.816.594-6.5-2 3.138-7.026-4.5-10.5a3.65 3.65 0 0 0 1-1.5 12.4 12.4 0 0 0-4-2.5l2-1a184 184 0 0 1-10.5-7.5q-1.006 1.129-2.5 1a128.5 128.5 0 0 0-6-17"/><path style="opacity:1" fill="#93a608" d="M399.5 360.5q2.364 3.126 5.5 6-.102 2.763.5 5h-6q-.225 5.542 3 10-1.063 1.356-3 1-.423-7.619-3-.5-1 1-2 0 2.64-1.448 1-3.5a20.5 20.5 0 0 0-3 5.5q-3.758 1.942-8 .5.693-6.154 0-12-2.835-.305-4.5 2-.354-3.466-2.5-6 .033-3.048 1-6.5a25 25 0 0 1 7-.5q2.634 3.965 3.5 8l1-2q2.724 1.244 5.5 2.5a10.3 10.3 0 0 0-4 1.5q4.095 10.125 6-.5a65 65 0 0 0-2-9.5q2.398.401 4-1"/><path style="opacity:1" fill="#8da501" d="M297.5 357.5a35 35 0 0 1 8 6 18.5 18.5 0 0 0 .5 6q1.517-1.255 3.5-1a50 50 0 0 1-2 12q3.715-.674 6 2-1.154 3.451-3 .5l1-1q-4.053-.91-8 .5a52.5 52.5 0 0 1-6.5-4q-2.59.22-.5 2a4.46 4.46 0 0 1-1.5 2 10.8 10.8 0 0 0-2.5-3q3.33-3.076 2-7a32 32 0 0 1-5 4.5q1.129 1.006 1 2.5a11 11 0 0 0-2.5-1q-3.036.45-1.5 2.5-1.5 1-3 0 1.817-1.34 0-2.5a3.65 3.65 0 0 0-1.5 1q-.75-1.75-2.5-2.5 5.46-1.701-1-3 6.942-.577 3-5.5 3.276-3.246 4-8l6 10q1.663 1.635 3 0a20.8 20.8 0 0 1-3-7.5 76 76 0 0 0 8 2.5q.371-3.537-2-6z"/><path style="opacity:1" fill="#718a00" d="M214.5 371.5a33.3 33.3 0 0 0 8 5 20.2 20.2 0 0 1 4 8.5q1.974.792 2-1.5l4 5q.913 5.34-4.5 6-4.234-4.098-10-6-2.166 2.483-5.5 2 1.316-6.789 0-13.5l.5-2.5q1.37 1.73 2.5-.5-1.129-1.006-1-2.5"/><path style="opacity:.053" fill="#96a528" d="M405.5 371.5q.759 2.893 3 5-1.606 2.738 2 3.5-3.567 1.461 1 2-1.32 2.034-1 4.5 7.49.746 3 6.5 1.701 1.524-1 2a548 548 0 0 0 2 10.5q-1.503 4.998 1 9-1.817 1.83 0 3.5l-3 3q1.842 2.235 0 4.5h-7q-.371-3.537 2-6h3q1.698-4.768 2-10-3.326-5.85-3.5-13-3.576-2.717-5-7a18.5 18.5 0 0 1-.5-6q.258-1.488-1-2-3.225-4.458-3-10z"/><path style="opacity:1" fill="#91a801" d="M315.5 370.5q6.635 5.177 12 12 1.388-1.712 2.5 0-.748 3.783.5 7.5.956 2.569 3 .5.262 7.862 8 8.5-3.793 3.962-3 9.5a8.4 8.4 0 0 1 4 .5q-3.158 1.79-5 5-2.337.315-4 1.5-1.426 2.427-4 4 .841-2.349-2-4l-2 4-3-6q1.398-2.22 0-4-.863-2.763.5-6 .724 1.422.5 3h3q-.422-3.695-3-6.5 2-1 4 0 .618 2.806 0 6 .839 3.281 4 2-3.588-3.41 0-6.5a20.5 20.5 0 0 1-3.5-7q-2.589 1.614-5.5.5 2.115-.552 3-2.5-2.087-4.934-6.5-5l-.5-2a13 13 0 0 1 3-4q-3.89-2.356-3-7-3.49.38-6-2a7.5 7.5 0 0 1 3-2"/><path style="opacity:1" fill="#8a9f01" d="M373.5 365.5a9.86 9.86 0 0 0 1 6q.21-1.604 0-3.5 1.874-1.11 3 .5 2.146 2.534 2.5 6 1.665-2.305 4.5-2 .693 5.846 0 12 4.242 1.442 8-.5a20.5 20.5 0 0 1 3-5.5q1.64 2.052-1 3.5 1 1 2 0 2.577-7.119 3 .5 1.937.356 3-1 1.258.512 1 2l-10 1q1.68 7.238-2 14a10.3 10.3 0 0 0 4 1.5 4.46 4.46 0 0 0-2 1.5 10.3 10.3 0 0 0 4 1.5q-2.159 1.486-3 4 2.342.477 3.5 2.5a7.3 7.3 0 0 1 2-3l1 7a12.8 12.8 0 0 1 3 6q1.718-1.776 3.5 0-2.371 2.463-2 6h-6v-3q-5.44-.182-7-5.5-2.911-1.114-5.5.5-.463-2.837 2.5-4-1.122-2.874 0-6 3.675-3.893-1-7.5l2-1q-.491-3.482-2.5-7.5-8.115-1.63-13.5-8l.5-1q4.57 4.215 4-2-3.251-2.947-7-1a36 36 0 0 0-4.5-3.5q9.476.86 2-4 1.817-1.34 0-2.5-7.22 4.104-12-3 1.966-1.137 1.5-3-2.355-.291-4.5 0a186 186 0 0 0 2-3q2.552-.699 4.5 0a7.3 7.3 0 0 0 2-3q4.14 4.689 10.5 4z"/><path style="opacity:1" fill="#8ea701" d="M289.5 380.5a153 153 0 0 0 7.5 14q4.077-.77 8.5 0a30 30 0 0 0-.5 5q2.082-1.58 4.5-.5-3.317 2.742 1.5 4.5 1.147-.105 1.5-1 3.28.823 2 3.5 2.176-.394 4 .5a16.1 16.1 0 0 1-2.5 5q-7.509-2.804-2.5 3.5-1.32 2.034-1 4.5-2.398.401-4-1 .197-1.314-1-2-3.015-1.503-6-3a82 82 0 0 1 7-2q1.532-.752.5-2l-3-1-1 2a35 35 0 0 0-6-8q-.432 4.458 2.5 8-2.596 4.109-3.5-1l-1 3q-2.505-3.298-6-1a2.43 2.43 0 0 1-.5-2q2.93-5.302-2-8.5 2-1.5 0-3 2.418-1.08 4.5.5a55 55 0 0 0 4.5-2q-4.192.358-9-1-2.417 1.23-2 4-2.9-.343-5-2.5a23.5 23.5 0 0 1 7-3.5 4.46 4.46 0 0 0-2-1.5q2.969-1.026 5 1.5 3.785-4.785-3-4-2.328-4.944 1-9"/><path style="opacity:1" fill="#86a001" d="M200.5 393.5q4.163-1.218 6.5-5 1.903.174 4 1 .987 2.412-1.5 3.5 7.102 3.796 4 11.5a4.46 4.46 0 0 1-2-1.5q1.817-1.34 0-2.5a4.46 4.46 0 0 0-1.5 2q-2.017-3.177-1.5-6-3.08 2.159-3 6h-3q.03 2.464-2 4 1.183 2.982 3.5 1l1 2q2.632-1.755 3-5 1.566 1.684 2 4 .981-1.794 1.5 0-2.318 1.428-5 1-2.1 10.404 3 1.5a11.3 11.3 0 0 0 3.5-1.5q.231 4.098-1.5 8 .544 1.025 1 2-3.705 2.076-5 6h-43q1.418-3.098 2-5.5a7.3 7.3 0 0 0-3-2l1-1a29 29 0 0 1-4-1.5 4.46 4.46 0 0 0-1.5 2 12.9 12.9 0 0 1-.5-5q1.983.255 3.5-1 1.114 3.217 2.5.5-1.394-1.616-2-3.5 18.857 2.732 2-6 .566-3.486 3-2a16.3 16.3 0 0 0 2-4q3.843-.913 8 .5-3.465.745-7 .5-.03 2.464 2 4a4.46 4.46 0 0 1-2 1.5q4.455 1.144 7.5 4.5a10.8 10.8 0 0 1 2.5-3 4.46 4.46 0 0 0-2-1.5q2.963-4.113 7-7.5-1.293-3.97.5-8 .724 1.422.5 3h3a8.4 8.4 0 0 1 .5-4q1.068 4.284 4.5 8.5 3.402.135 4.5-3.5.417 1.374 1.5 2"/><path style="opacity:1" fill="#8b9e09" d="M54.5 365.5q-.25 17.008.5 34 .2 4.538 3 8 2.603-6.398.5-13 5.452-.233 1-3.5 5.59-1.564-1-1.5v-4q2.217.393 4-1-2.032-1.763.5-3 1.665 2.305 4.5 2 .257 4.946-2 1-2.118 1.96-5 2.5 1.605 1.278 3 2.5.93 3.716 3 0l1 1a33.3 33.3 0 0 0-.5 6q2.838-1.676 2.5-5h3q-.41 2.927 2 4.5-7.132 3.765-1 9-1.424.926-2.5-.5-2.426 1.554-.5 3.5a18 18 0 0 0-2 3.5q.775 6.9 0 14h-12q.634-2.925 1-6a8.9 8.9 0 0 1-3-4q2-1.5 0-3-3.179-1.532-5-4.5l1-1q-3.968-3.126-2-7.5.873-2.146 2.5-4 .621.686 1.5 1a91 91 0 0 0 1-19q-.475-6.303 1-12"/><path style="opacity:1" fill="#879e01" d="M85.5 375.5a186 186 0 0 0-2 7.5 34 34 0 0 1 3 6.5 3.65 3.65 0 0 0 1.5-1 6.8 6.8 0 0 0 1.5 3q1.77-2.453 0-5 1.654-.747 3-1a11.1 11.1 0 0 1-2-3.5q1.81-1.95 2-4.5 2.9.343 5 2.5a6.8 6.8 0 0 0-3 1.5 6.85 6.85 0 0 1 1 4q7.272.255 5 7a6.85 6.85 0 0 0-4-1v3q-3.265-.13-6-2a7.3 7.3 0 0 0-3 2q7.697 2.024 2 6.5 2.058.603 4 1.5a10.8 10.8 0 0 1 3-2.5q-1.58-2.082-.5-4.5 5.933 3.715 9 10 .426-3.579 3-2 2.779 2.595-.5 3.5 6.26 4.514-2 3.5.334-3.921-3.5-5-2.232 4.935 1.5 9.5-1.703 1.124-3-.5a7.3 7.3 0 0 0-2-3 55 55 0 0 1-6.5 6 32.4 32.4 0 0 0-1.5-6 3.65 3.65 0 0 1 1-1.5 6.54 6.54 0 0 1-3-3q1.664-2.72-2-2.5.056 7.479-7 7.5 1.794-.981 0-1.5-1.89 2.78 0 5.5 7.595 1.992-1 3l2 2q-3.89 1.329-6 4.5h-6q.775-7.1 0-14a18 18 0 0 1 2-3.5q-1.926-1.946.5-3.5 1.076 1.426 2.5.5-6.132-5.235 1-9-2.41-1.573-2-4.5h-3q.338 3.324-2.5 5-.03-2.907.5-6l-1-1q-2.07 3.716-3 0 3.293-2.245 7-2.5a18 18 0 0 1-2-3.5q.519-1.794 1.5 0l1-3 1 1q1.011-1.617 3-2 4.816 4.88 4.5-2 2.551-.699 4.5 0 .417-1.374 1.5-2"/><path style="opacity:1" fill="#95ac03" d="M305.5 363.5h6a12.9 12.9 0 0 0 .5 5q2.191 2.074 2.5-1 1.356 1.063 1 3a7.5 7.5 0 0 0-3 2q2.51 2.38 6 2-.89 4.644 3 7a13 13 0 0 0-3 4l.5 2q4.413.066 6.5 5-.885 1.948-3 2.5 2.911 1.114 5.5-.5a20.5 20.5 0 0 0 3.5 7q-3.588 3.09 0 6.5-3.161 1.281-4-2 .618-3.194 0-6-2-1-4 0 2.578 2.805 3 6.5h-3a4.93 4.93 0 0 0-.5-3q-1.363 3.237-.5 6-1.494-.129-2.5 1-.225-4.381 1.5-8.5a6.8 6.8 0 0 1-3-1.5 21 21 0 0 0 2-7 8.4 8.4 0 0 1-4-.5q2.675-1.6 1.5-4.5-1.517 1.255-3.5 1 .5-4.5-4-4-2.96 6.162 2.5 10 1.294 3.373-1.5 5.5.834.574 1 1.5-.353.895-1.5 1-4.817-1.758-1.5-4.5-2.418-1.08-4.5.5.028-2.41.5-5-4.423-.77-8.5 0a153 153 0 0 1-7.5-14q1 0 1-1 .129-1.494-1-2.5a32 32 0 0 0 5-4.5q1.33 3.924-2 7a10.8 10.8 0 0 1 2.5 3 4.46 4.46 0 0 0 1.5-2q-2.09-1.78.5-2a52.5 52.5 0 0 0 6.5 4q3.947-1.41 8-.5l-1 1q1.846 2.951 3-.5-2.285-2.674-6-2a50 50 0 0 0 2-12q-1.983-.255-3.5 1a18.5 18.5 0 0 1-.5-6"/><path style="opacity:1" fill="#839801" d="M336.5 340.5a219 219 0 0 1 2.5 6 6.8 6.8 0 0 1 1.5-3l2 5q4.276.768 4 5 3.757.492 5.5 4l1-1a8.4 8.4 0 0 1 .5 4h3q-.961 1.473-2 3 2.145-.291 4.5 0 .466 1.863-1.5 3 4.78 7.104 12 3 1.817 1.16 0 2.5 7.476 4.86-2 4a36 36 0 0 1 4.5 3.5q3.749-1.947 7 1 .57 6.215-4 2l-.5 1q5.385 6.37 13.5 8 2.009 4.018 2.5 7.5l-2 1q4.675 3.607 1 7.5-.926.166-1.5 1a6.8 6.8 0 0 0-1.5-3q-1.427 1.239-3 .5a11.3 11.3 0 0 0 1.5-3.5q.981 1.794 1.5 0l-1.5-3a19 19 0 0 1-3 5 55 55 0 0 0-2-7q-4.536-3.143-3.5-8-3.313-.844-4.5-4-1.865 1.254-3.5-1a168 168 0 0 0-2-7.5q-3.34-.181-5.5-2.5l-1 1q-.348-1.856-1.5-3-2.436.21-3.5-2a4.46 4.46 0 0 1-1.5 2q-1.657-3.167-5-4-2.48-.24-4-2 1.602-1.401 4-1-.03-2.464 2-4-1.841-2.732-5-3a6.85 6.85 0 0 0-1-4 45.3 45.3 0 0 1-8-2.5q1.392-.582 2-2a22.1 22.1 0 0 1-2-9.5"/><path style="opacity:.79" fill="#687c07" d="M316.5 334.5q3 0 3 3-4.966-3.83-4.5 3a13.9 13.9 0 0 1 4 3 72 72 0 0 1 4.5 9 44.7 44.7 0 0 1 9 7 148 148 0 0 0 15 18 301 301 0 0 1 5 8h3q-.585-5.009-1.5-10-1.885-3.369-5.5-4v-3q3.343.833 5 4a4.46 4.46 0 0 0 1.5-2q1.064 2.21 3.5 2 .129 1.494-1 2.5 2.42 3.342 2 7.5a4.93 4.93 0 0 1 3 .5 25.9 25.9 0 0 0 6.5 5.5q1.294 3.373-1.5 5.5a35.5 35.5 0 0 0 7 10.5q-2.46 3.124-6 1a71 71 0 0 1-6-7.5q.795-3.205-2.5-5.5a65.5 65.5 0 0 0-6.5 11q-2.775 1.793-3-2l2-7a18.35 18.35 0 0 1-2-9h-3v-3q-7.946-.798 0-2.5-2.877-4.392-8-5a37 37 0 0 0-4.5-9.5q-1.896-3.16-5.5-3 .531-4.381-3-7-8.944-2.927-10-12-3.921.334-5-3.5a55 55 0 0 1-7-2l2-1a28 28 0 0 1-4-1.5v-3a15.4 15.4 0 0 1 2-4.5q-1.873-2.307.5-3.5a168 168 0 0 1 11.5 10"/><path style="opacity:1" fill="#819901" d="M209.5 366.5q2.549.04 4 2-.116 1.875-2 2.5a4.93 4.93 0 0 0 3 .5q-.129 1.494 1 2.5-1.13 2.23-2.5.5l-.5 2.5q1.316 6.711 0 13.5 3.334.483 5.5-2 5.766 1.902 10 6 5.413-.66 4.5-6 6.007 11.504 12 23a26.4 26.4 0 0 1-7-4 55 55 0 0 0-2 4.5q-2.166 1.171-4-.5 6.072-4.983-.5-9l-1.5 1a9.3 9.3 0 0 1 2 2.5q.355 6.585-6.5 3.5l-1.5-1 2-2q-5.54-5.649-7 3a22.6 22.6 0 0 1 2 8q-2.435-.21-3.5 2-3.06-5.52-4.5 0a44 44 0 0 0-1-2q1.731-3.902 1.5-8a11.3 11.3 0 0 1-3.5 1.5q-5.1 8.904-3-1.5 2.682.428 5-1-.519-1.794-1.5 0-.434-2.316-2-4-.368 3.245-3 5l-1-2q-2.317 1.982-3.5-1 2.03-1.536 2-4h3q-.08-3.841 3-6-.517 2.823 1.5 6a4.46 4.46 0 0 1 1.5-2q1.817 1.16 0 2.5a4.46 4.46 0 0 0 2 1.5q3.102-7.704-4-11.5 2.487-1.088 1.5-3.5-2.097-.826-4-1-2.337 3.782-6.5 5 .123-5.649 4.5-9l1 2a11.3 11.3 0 0 1 1.5-3.5q-1.61-1.875 0-4a9.3 9.3 0 0 0-2-2.5q7.761-2.655-1-3.5 5.016-.855 1-3.5a174 174 0 0 0 4-3"/><path style="opacity:.005" fill="#6b7f02" d="M356.5 400.5q2.755.209 1 2-1.197-.686-1-2"/><path style="opacity:1" fill="#849e01" d="M244.5 411.5q1.5.5 2 2 .467 1.467 2 2 2.827-6.895 7.5 0 1.171 2.166-.5 4-1.817-1.16 0-2.5-4.053-.08-8 1.5-5.86-2.701-10 2-2.674-2.284-2-6a18.5 18.5 0 0 0-6 .5l2 1a12.14 12.14 0 0 1-4.5 3.5q-1.207-3.601-2.5.5 2.731 3.205 7 2.5a4.93 4.93 0 0 1 .5-3l1.5 6h-26q1.295-3.924 5-6 1.44-5.52 4.5 0 1.065-2.21 3.5-2a22.6 22.6 0 0 0-2-8q1.46-8.649 7-3l-2 2 1.5 1q6.855 3.085 6.5-3.5a9.3 9.3 0 0 0-2-2.5l1.5-1q6.572 4.017.5 9 1.834 1.671 4 .5a55 55 0 0 1 2-4.5 26.4 26.4 0 0 0 7 4"/><path style="opacity:1" fill="#788c02" d="M358.5 372.5q1.152 1.144 1.5 3l1-1q2.16 2.319 5.5 2.5a168 168 0 0 1 2 7.5q1.635 2.254 3.5 1 1.187 3.156 4.5 4-1.036 4.857 3.5 8a55 55 0 0 1 2 7 19 19 0 0 0 3-5l1.5 3q-.519 1.794-1.5 0a11.3 11.3 0 0 1-1.5 3.5q1.573.739 3-.5a6.8 6.8 0 0 1 1.5 3q.574-.834 1.5-1-1.122 3.126 0 6-2.963 1.163-2.5 4 2.589-1.614 5.5-.5 1.56 5.318 7 5.5v3h-16q1.857-3.437-1.5-6a150 150 0 0 1-8.5-6q-5.702-2.32-6 4-2.458.639-4 0 2.506-6.01 5-12 3.54 2.124 6-1a35.5 35.5 0 0 1-7-10.5q2.794-2.127 1.5-5.5a25.9 25.9 0 0 1-6.5-5.5 4.93 4.93 0 0 0-3-.5q.42-4.158-2-7.5 1.129-1.006 1-2.5"/><path style="opacity:.998" fill="#94a706" d="M403.5 383.5a18.5 18.5 0 0 0 .5 6q1.424 4.283 5 7 .174 7.15 3.5 13-.302 5.232-2 10h-3q-1.782-1.776-3.5 0a12.8 12.8 0 0 0-3-6l-1-7a7.3 7.3 0 0 0-2 3q-1.158-2.023-3.5-2.5.841-2.514 3-4a10.3 10.3 0 0 1-4-1.5 4.46 4.46 0 0 1 2-1.5 10.3 10.3 0 0 1-4-1.5q3.68-6.762 2-14z"/><path style="opacity:1" fill="#869e02" d="M131.5 385.5q-.22 10.313-3 20l1 5q5.493-7.995 11-16a6.55 6.55 0 0 1 3-3q1.767 2.249 5 3l-2 7a14.7 14.7 0 0 0 5 1.5q1.255 1.517 1 3.5 2.519.322 3-2-.329-1.763 1-3a110 110 0 0 1 7 3.5q-.178 1.718-1 3.5.606 1.884 2 3.5-1.386 2.717-2.5-.5-1.517 1.255-3.5 1a12.9 12.9 0 0 0 .5 5 4.46 4.46 0 0 1 1.5-2 29 29 0 0 0 4 1.5l-1 1a7.3 7.3 0 0 1 3 2q-.582 2.402-2 5.5h-26q-.76-2.038-3-3 2.133-5.133 7-7a41.5 41.5 0 0 1-1-8 6.85 6.85 0 0 1-4-1q.59 2.886-2 5.5a31.6 31.6 0 0 1-8-.5v-3q-1.762.329-3-1 2.258-2.012 2-5a18.35 18.35 0 0 1 2-9 38 38 0 0 0-2-5.5q1.96-3.385 5-2.5"/><path style="opacity:1" fill="#839c01" d="M126.5 402.5q.258 2.988-2 5 1.238 1.329 3 1v3a31.6 31.6 0 0 0 8 .5q2.59-2.614 2-5.5a6.85 6.85 0 0 0 4 1 41.5 41.5 0 0 0 1 8q-4.867 1.867-7 7 2.24.962 3 3h-35q.165-3.869-2-8l1-1a10.3 10.3 0 0 1 1.5 4q2.088-4.167 1.5-9h3q-.329 1.763 1 3 1.536-2.03 4-2-2.158-7.124.5-14 3.145 3.85 4 9 1.566-1.684 2-4l1 1 .5-3q-.852-2.556 0-4.5a6.8 6.8 0 0 1-3-1.5q1.16-1.817 2.5 0l1-2a8.4 8.4 0 0 1 .5 4 4.93 4.93 0 0 1 3 .5q-3.526 2.781 1 4.5"/><path style="opacity:1" fill="#8ea500" d="M322.5 409.5q1.398 1.78 0 4l3 6 2-4q2.841 1.651 2 4 2.574-1.573 4-4h7a77 77 0 0 0-1 10h-33q-1.83-2.673-2-6 2.188.1 4-1 1.602 1.401 4 1-.32-2.466 1-4.5-5.009-6.304 2.5-3.5a16.1 16.1 0 0 0 2.5-5q-1.824-.894-4-.5 1.28-2.677-2-3.5-.166-.926-1-1.5 2.794-2.127 1.5-5.5-5.46-3.838-2.5-10 4.5-.5 4 4 1.983.255 3.5-1 1.175 2.9-1.5 4.5a8.4 8.4 0 0 0 4 .5 21 21 0 0 1-2 7 6.8 6.8 0 0 0 3 1.5q-1.725 4.119-1.5 8.5 1.006-1.129 2.5-1"/><path style="opacity:1" fill="#8ba301" d="M290.5 409.5a2.43 2.43 0 0 0 .5 2q3.495-2.298 6 1l1-3q.904 5.109 3.5 1-2.932-3.542-2.5-8a35 35 0 0 1 6 8l1-2 3 1q1.032 1.248-.5 2a82 82 0 0 0-7 2q2.985 1.497 6 3 1.197.686 1 2-1.812 1.1-4 1 .17 3.327 2 6h-28q.571-2.146 3-2.5a22.5 22.5 0 0 1-7-2l2-2a42.8 42.8 0 0 1-9-1.5q2.184.267 4-2.5-4.206-1.457-6 2.5 1.247 2.96.5 6-1.5-2-3 0 .126-4.827-3.5-9 1.237-1.329 3-1a27.2 27.2 0 0 1 2-5.5 11.3 11.3 0 0 0-1-3.5q7.214-9.972 3 2 3.872 5.94 3-2 3.271-.324 5 2.5-2.172 6.69 5 5a8.4 8.4 0 0 0-4-.5v-3q11.215-.558 1-4.5 1.75-.75 2.5-2.5 2.482 4.066 6 5-.072 4.602 3.5 7a71 71 0 0 1 2-4"/><path style="opacity:1" fill="#8ba402" d="M246.5 378.5q1.401 1.602 1 4-2.328 10.006 6 15a16.9 16.9 0 0 0 1 8q2.764.383 6 .5 1.32 2.034 1 4.5h-4a18.5 18.5 0 0 0 .5 6 9.5 9.5 0 0 0 2.5 1.5q-2.556 3.356-3 7.5h-24l-1.5-6a4.93 4.93 0 0 0-.5 3q-4.269.705-7-2.5 1.293-4.101 2.5-.5a12.14 12.14 0 0 0 4.5-3.5l-2-1a18.5 18.5 0 0 1 6-.5q-.674 3.716 2 6 4.14-4.701 10-2 3.947-1.58 8-1.5-1.817 1.34 0 2.5 1.671-1.834.5-4-4.673-6.895-7.5 0-1.533-.533-2-2 1.916-1.79 2-5a227 227 0 0 1-2-30"/><path style="opacity:1" fill="#506100" d="M383.5 425.5h-26q1.563-5.145 6-8 1.542.639 4 0 .298-6.32 6-4a150 150 0 0 0 8.5 6q3.357 2.563 1.5 6"/></svg>`,
